Get a list of assertions

Retrieve all assertions from the CIViC DB

getAllAssertions(page = 1, count = 25)

Arguments

page
the page number to retrieve

count
the number of assertions to retrieve

Value

An S3 Object of type civic_api containing the content, url, and response

Examples

getAllAssertions(count = 10)

#> $content
#> $content$`_meta`
#> $content$`_meta`$current_page
#> [1] 1
#> 
#> $content$`_meta`$per_page
#> [1] "10"
#> 
#> $content$`_meta`$total_pages
#> [1] 3
#> 
#> $content$`_meta`$total_count
#> [1] 27
#> 
#> $content$`_meta`$links
#> $content$`_meta`$links$`next`
#> [1] "https://civicdb.org/api/assertions?count=10&page=2"
#> 
#> $content$`_meta`$links$previous
#> NULL
#> 
#> 
#> 
#> $content$records
#> $content$records[[1]]
#> $content$records[[1]]$id
#> [1] 1
#> 
#> $content$records[[1]]$type
#> [1] "assertion"
#> 
#> $content$records[[1]]$name
#> [1] "AID1"
#> 
#> $content$records[[1]]$summary
#> [1] "HER2 amplification predicts sensitivity to Trastuzumab"
#> 
#> $content$records[[1]]$description
#> [1] "HER2 amplification defines a clinically relevant subtype of breast cancer. HER2 amplification predicts sensitivity to various targeted therapies including the monoclonal antibody Trastuzumab. The use of Trastuzumab, often in combination with chemotherapy and/or endocrine therapy (depending on hormone receptor status), is now standard of care for HER2-positive breast cancer patients."
#> 
#> $content$records[[1]]$gene
#> $content$records[[1]]$gene$name
#> [1] "ERBB2"
#> 
#> $content$records[[1]]$gene$id
#> [1] 20
#> 
#> 
#> $content$records[[1]]$variant
#> $content$records[[1]]$variant$name
#> [1] "AMPLIFICATION"
#> 
#> $content$records[[1]]$variant$id
#> [1] 18
#> 
#> 
#> $content$records[[1]]$disease
#> $content$records[[1]]$disease$id
#> [1] 368
#> 
#> $content$records[[1]]$disease$name
#> [1] "Her2-receptor Positive Breast Cancer"
#> 
#> $content$records[[1]]$disease$display_name
#> [1] "Her2-receptor Positive Breast Cancer"
#> 
#> $content$records[[1]]$disease$doid
#> [1] "0060079"
#> 
#> $content$records[[1]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:0060079"
#> 
#> 
#> $content$records[[1]]$drugs
#> $content$records[[1]]$drugs[[1]]
#> $content$records[[1]]$drugs[[1]]$id
#> [1] 84
#> 
#> $content$records[[1]]$drugs[[1]]$name
#> [1] "Trastuzumab"
#> 
#> $content$records[[1]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[1]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[1]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[1]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[1]]$evidence_item_count
#> [1] 3
#> 
#> $content$records[[1]]$fda_regulatory_approval
#> [1] TRUE
#> 
#> $content$records[[1]]$status
#> [1] "rejected"
#> 
#> $content$records[[1]]$open_change_count
#> [1] 0
#> 
#> $content$records[[1]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[2]]
#> $content$records[[2]]$id
#> [1] 2
#> 
#> $content$records[[2]]$type
#> [1] "assertion"
#> 
#> $content$records[[2]]$name
#> [1] "AID2"
#> 
#> $content$records[[2]]$summary
#> [1] "HER2 amplification predicts sensitivity to Trastuzumab"
#> 
#> $content$records[[2]]$description
#> [1] "HER2 amplification defines a clinically relevant subtype of breast cancer. HER2 amplification predicts sensitivity to various targeted therapies including the monoclonal antibody Trastuzumab. The use of Trastuzumab, often in combination with chemotherapy and/or endocrine therapy (depending on hormone receptor status), is now standard of care for HER2-positive breast cancer patients."
#> 
#> $content$records[[2]]$gene
#> $content$records[[2]]$gene$name
#> [1] "ERBB2"
#> 
#> $content$records[[2]]$gene$id
#> [1] 20
#> 
#> 
#> $content$records[[2]]$variant
#> $content$records[[2]]$variant$name
#> [1] "AMPLIFICATION"
#> 
#> $content$records[[2]]$variant$id
#> [1] 18
#> 
#> 
#> $content$records[[2]]$disease
#> $content$records[[2]]$disease$id
#> [1] 368
#> 
#> $content$records[[2]]$disease$name
#> [1] "Her2-receptor Positive Breast Cancer"
#> 
#> $content$records[[2]]$disease$display_name
#> [1] "Her2-receptor Positive Breast Cancer"
#> 
#> $content$records[[2]]$disease$doid
#> [1] "0060079"
#> 
#> $content$records[[2]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:0060079"
#> 
#> 
#> $content$records[[2]]$drugs
#> $content$records[[2]]$drugs[[1]]
#> $content$records[[2]]$drugs[[1]]$id
#> [1] 84
#> 
#> $content$records[[2]]$drugs[[1]]$name
#> [1] "Trastuzumab"
#> 
#> $content$records[[2]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[2]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[2]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[2]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[2]]$evidence_item_count
#> [1] 3
#> 
#> $content$records[[2]]$fda_regulatory_approval
#> [1] TRUE
#> 
#> $content$records[[2]]$status
#> [1] "accepted"
#> 
#> $content$records[[2]]$open_change_count
#> [1] 0
#> 
#> $content$records[[2]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[3]]
#> $content$records[[3]]$id
#> [1] 3
#> 
#> $content$records[[3]]$type
#> [1] "assertion"
#> 
#> $content$records[[3]]$name
#> [1] "AID3"
#> 
#> $content$records[[3]]$summary
#> [1] "Lung adenocarcinoma positive for ALK-FUSIONS have been found to be sensitive to crizotinib treatment"
#> 
#> $content$records[[3]]$description
#> [1] "ALK-FUSIONS induce an active form of the receptor tyrosine kinase ALK.  ALK-FUSIONS are found in 3-5% of non-small cell lung cancer and act as a targetable driver mutation. the ALK inhibitor crizotinib was effective in lung cancer patients with ALK fusions and was granted accelerated FDA approval in 2011. The Vysis break-apart FISH assay is an FDA approved test for the presence of ALK fusion. NCCN guideline version 9.2017 lists crizotinib as approved first-line NSCLC treatment in the presence of ALK fusions but now lists alectinib as preferred first-line therapy."
#> 
#> $content$records[[3]]$gene
#> $content$records[[3]]$gene$name
#> [1] "ALK"
#> 
#> $content$records[[3]]$gene$id
#> [1] 1
#> 
#> 
#> $content$records[[3]]$variant
#> $content$records[[3]]$variant$name
#> [1] "ALK FUSIONS"
#> 
#> $content$records[[3]]$variant$id
#> [1] 499
#> 
#> 
#> $content$records[[3]]$disease
#> $content$records[[3]]$disease$id
#> [1] 8
#> 
#> $content$records[[3]]$disease$name
#> [1] "Lung Non-small Cell Carcinoma"
#> 
#> $content$records[[3]]$disease$display_name
#> [1] "Lung Non-small Cell Carcinoma"
#> 
#> $content$records[[3]]$disease$doid
#> [1] "3908"
#> 
#> $content$records[[3]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:3908"
#> 
#> 
#> $content$records[[3]]$drugs
#> $content$records[[3]]$drugs[[1]]
#> $content$records[[3]]$drugs[[1]]$id
#> [1] 12
#> 
#> $content$records[[3]]$drugs[[1]]$name
#> [1] "Crizotinib"
#> 
#> $content$records[[3]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[3]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[3]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[3]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[3]]$evidence_item_count
#> [1] 18
#> 
#> $content$records[[3]]$fda_regulatory_approval
#> [1] TRUE
#> 
#> $content$records[[3]]$status
#> [1] "accepted"
#> 
#> $content$records[[3]]$open_change_count
#> [1] 0
#> 
#> $content$records[[3]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[4]]
#> $content$records[[4]]$id
#> [1] 4
#> 
#> $content$records[[4]]$type
#> [1] "assertion"
#> 
#> $content$records[[4]]$name
#> [1] "AID4"
#> 
#> $content$records[[4]]$summary
#> [1] "R167Q (c.500G>A) is a pathogenic variant for Von Hippel-Lindau"
#> 
#> $content$records[[4]]$description
#> [1] "R167Q is the most common mutation associated with Von Hippel-Lindau syndrome. This variant is very rare in the general population at 4.061e-6 in the gnomAD exomes (v2.0.2) proving ACMG code PM2. The variant occurs within the functional domain, disrupting VHL binding to elongin C (ACMG code PM1). Additional codes are provided by the following EIDs. 4913 (PS4), 5062 (PP4), 5264 (PS4), 5354 (PP4), 5487 and 4913 (PP1)."
#> 
#> $content$records[[4]]$gene
#> $content$records[[4]]$gene$name
#> [1] "VHL"
#> 
#> $content$records[[4]]$gene$id
#> [1] 58
#> 
#> 
#> $content$records[[4]]$variant
#> $content$records[[4]]$variant$name
#> [1] "R167Q (c.500G>A)"
#> 
#> $content$records[[4]]$variant$id
#> [1] 1739
#> 
#> 
#> $content$records[[4]]$disease
#> $content$records[[4]]$disease$id
#> [1] 2198
#> 
#> $content$records[[4]]$disease$name
#> [1] "Von Hippel-Lindau Disease"
#> 
#> $content$records[[4]]$disease$display_name
#> [1] "Von Hippel-Lindau Disease"
#> 
#> $content$records[[4]]$disease$doid
#> [1] "14175"
#> 
#> $content$records[[4]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:14175"
#> 
#> 
#> $content$records[[4]]$drugs
#> list()
#> 
#> $content$records[[4]]$evidence_type
#> [1] "Predisposing"
#> 
#> $content$records[[4]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[4]]$clinical_significance
#> [1] "Pathogenic"
#> 
#> $content$records[[4]]$evidence_item_count
#> [1] 7
#> 
#> $content$records[[4]]$fda_regulatory_approval
#> [1] FALSE
#> 
#> $content$records[[4]]$status
#> [1] "accepted"
#> 
#> $content$records[[4]]$open_change_count
#> [1] 0
#> 
#> $content$records[[4]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[5]]
#> $content$records[[5]]$id
#> [1] 5
#> 
#> $content$records[[5]]$type
#> [1] "assertion"
#> 
#> $content$records[[5]]$name
#> [1] "AID5"
#> 
#> $content$records[[5]]$summary
#> [1] "Non-small cell lung cancer with EGFR L858R mutation is sensitive to erlotininb or gefitinib."
#> 
#> $content$records[[5]]$description
#> [1] "L858R is among the most common sensitizing EGFR mutations in NSCLC, and is assessed via DNA mutational analysis including Sanger sequencing and next generation sequencing methods. Tyrosine kinase inhibitors erlotinib and gefitinib are associated with improved progression free survival over chemotherapy in EGFR L858R patients. NCCN guidelines recommend (category 1) erlotinib and gefitinib for NSCLC with sensitizing EGFR mutations, along with afatinib and osimertinib."
#> 
#> $content$records[[5]]$gene
#> $content$records[[5]]$gene$name
#> [1] "EGFR"
#> 
#> $content$records[[5]]$gene$id
#> [1] 19
#> 
#> 
#> $content$records[[5]]$variant
#> $content$records[[5]]$variant$name
#> [1] "L858R"
#> 
#> $content$records[[5]]$variant$id
#> [1] 33
#> 
#> 
#> $content$records[[5]]$disease
#> $content$records[[5]]$disease$id
#> [1] 8
#> 
#> $content$records[[5]]$disease$name
#> [1] "Lung Non-small Cell Carcinoma"
#> 
#> $content$records[[5]]$disease$display_name
#> [1] "Lung Non-small Cell Carcinoma"
#> 
#> $content$records[[5]]$disease$doid
#> [1] "3908"
#> 
#> $content$records[[5]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:3908"
#> 
#> 
#> $content$records[[5]]$drugs
#> $content$records[[5]]$drugs[[1]]
#> $content$records[[5]]$drugs[[1]]$id
#> [1] 14
#> 
#> $content$records[[5]]$drugs[[1]]$name
#> [1] "Gefitinib"
#> 
#> $content$records[[5]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> $content$records[[5]]$drugs[[2]]
#> $content$records[[5]]$drugs[[2]]$id
#> [1] 15
#> 
#> $content$records[[5]]$drugs[[2]]$name
#> [1] "Erlotinib"
#> 
#> $content$records[[5]]$drugs[[2]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[5]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[5]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[5]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[5]]$evidence_item_count
#> [1] 14
#> 
#> $content$records[[5]]$fda_regulatory_approval
#> [1] TRUE
#> 
#> $content$records[[5]]$status
#> [1] "accepted"
#> 
#> $content$records[[5]]$open_change_count
#> [1] 0
#> 
#> $content$records[[5]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[6]]
#> $content$records[[6]]$id
#> [1] 6
#> 
#> $content$records[[6]]$type
#> [1] "assertion"
#> 
#> $content$records[[6]]$name
#> [1] "AID6"
#> 
#> $content$records[[6]]$summary
#> [1] "EGFR L858R positive NSCLC is sensitive to afatinib."
#> 
#> $content$records[[6]]$description
#> [1] "L858R is among the most common sensitizing EGFR mutations in NSCLC, and is assessed via DNA mutational analysis, including Sanger sequencing and next generation sequencing methods. Tyrosine kinase inhibitor afatinib is FDA approved, and is recommended (category 1) by NCCN guidelines along with erlotinib, gefitinib and osimertinib as first line systemic therapy in NSCLC with sensitizing EGFR mutation."
#> 
#> $content$records[[6]]$gene
#> $content$records[[6]]$gene$name
#> [1] "EGFR"
#> 
#> $content$records[[6]]$gene$id
#> [1] 19
#> 
#> 
#> $content$records[[6]]$variant
#> $content$records[[6]]$variant$name
#> [1] "L858R"
#> 
#> $content$records[[6]]$variant$id
#> [1] 33
#> 
#> 
#> $content$records[[6]]$disease
#> $content$records[[6]]$disease$id
#> [1] 8
#> 
#> $content$records[[6]]$disease$name
#> [1] "Lung Non-small Cell Carcinoma"
#> 
#> $content$records[[6]]$disease$display_name
#> [1] "Lung Non-small Cell Carcinoma"
#> 
#> $content$records[[6]]$disease$doid
#> [1] "3908"
#> 
#> $content$records[[6]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:3908"
#> 
#> 
#> $content$records[[6]]$drugs
#> $content$records[[6]]$drugs[[1]]
#> $content$records[[6]]$drugs[[1]]$id
#> [1] 146
#> 
#> $content$records[[6]]$drugs[[1]]$name
#> [1] "Afatinib"
#> 
#> $content$records[[6]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[6]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[6]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[6]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[6]]$evidence_item_count
#> [1] 6
#> 
#> $content$records[[6]]$fda_regulatory_approval
#> [1] TRUE
#> 
#> $content$records[[6]]$status
#> [1] "accepted"
#> 
#> $content$records[[6]]$open_change_count
#> [1] 0
#> 
#> $content$records[[6]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[7]]
#> $content$records[[7]]$id
#> [1] 7
#> 
#> $content$records[[7]]$type
#> [1] "assertion"
#> 
#> $content$records[[7]]$name
#> [1] "AID7"
#> 
#> $content$records[[7]]$summary
#> [1] "BRAF V600E mutant melanoma is sensitive to dabrafenib and trametinib combination therapy"
#> 
#> $content$records[[7]]$description
#> [1] "Combination treatment of BRAF inhibitor dabrafenib and MEK inhibitor trametinib is recommended for adjuvant treatment of stage III or recurrent melanoma with BRAF V600E mutation detected by the approved THxID kit, as well as first line treatment for metastatic melanoma. The treatments are FDA approved and NCCN guidelines recommend these treatments as category 1 based on studies including the Phase III COMBI-V, COMBI-D and COMBI-AD Trials. Combination therapy is now recommended above BRAF inhibitor monotherapy. Dabrafenib and trametinib are recommend as NCCN Category 2A for second line therapy in metastatic melanoma due to lack of clear Phase III trial data for this use case. Cutaneous squamous-cell carcinoma and keratoacanthoma occur at lower rates with combination therapy than with BRAF inhibitor alone."
#> 
#> $content$records[[7]]$gene
#> $content$records[[7]]$gene$name
#> [1] "BRAF"
#> 
#> $content$records[[7]]$gene$id
#> [1] 5
#> 
#> 
#> $content$records[[7]]$variant
#> $content$records[[7]]$variant$name
#> [1] "V600E"
#> 
#> $content$records[[7]]$variant$id
#> [1] 12
#> 
#> 
#> $content$records[[7]]$disease
#> $content$records[[7]]$disease$id
#> [1] 7
#> 
#> $content$records[[7]]$disease$name
#> [1] "Melanoma"
#> 
#> $content$records[[7]]$disease$display_name
#> [1] "Melanoma"
#> 
#> $content$records[[7]]$disease$doid
#> [1] "1909"
#> 
#> $content$records[[7]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:1909"
#> 
#> 
#> $content$records[[7]]$drugs
#> $content$records[[7]]$drugs[[1]]
#> $content$records[[7]]$drugs[[1]]$id
#> [1] 19
#> 
#> $content$records[[7]]$drugs[[1]]$name
#> [1] "Trametinib"
#> 
#> $content$records[[7]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> $content$records[[7]]$drugs[[2]]
#> $content$records[[7]]$drugs[[2]]$id
#> [1] 22
#> 
#> $content$records[[7]]$drugs[[2]]$name
#> [1] "Dabrafenib"
#> 
#> $content$records[[7]]$drugs[[2]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[7]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[7]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[7]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[7]]$evidence_item_count
#> [1] 4
#> 
#> $content$records[[7]]$fda_regulatory_approval
#> [1] TRUE
#> 
#> $content$records[[7]]$status
#> [1] "accepted"
#> 
#> $content$records[[7]]$open_change_count
#> [1] 0
#> 
#> $content$records[[7]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[8]]
#> $content$records[[8]]$id
#> [1] 8
#> 
#> $content$records[[8]]$type
#> [1] "assertion"
#> 
#> $content$records[[8]]$name
#> [1] "AID8"
#> 
#> $content$records[[8]]$summary
#> [1] "NTRK1 Fusions predict sensitivity to Larotrectinib"
#> 
#> $content$records[[8]]$description
#> [1] "NTRK1 Fusions predict sensitivity to Larotrectinib across cancer types and patient ages. Objective response rates have been reported as high as 75% in adults and over 90% in children. NTRK1 forms fusions with multiple partner genes. Larotrectinib is FDA approved for the treatment of adult and pediatric patients with solid tumors characterized by an NTRK gene fusion and without an acquired resistance mutation. Eligible patients have metastatic disease, have progressed on the treatment they’re receiving, are without other treatment options, and are not candidates for surgical resection of their tumor."
#> 
#> $content$records[[8]]$gene
#> $content$records[[8]]$gene$name
#> [1] "NTRK1"
#> 
#> $content$records[[8]]$gene$id
#> [1] 3983
#> 
#> 
#> $content$records[[8]]$variant
#> $content$records[[8]]$variant$name
#> [1] "NTRK1 FUSIONS"
#> 
#> $content$records[[8]]$variant$id
#> [1] 419
#> 
#> 
#> $content$records[[8]]$disease
#> $content$records[[8]]$disease$id
#> [1] 216
#> 
#> $content$records[[8]]$disease$name
#> [1] "Cancer"
#> 
#> $content$records[[8]]$disease$display_name
#> [1] "Cancer"
#> 
#> $content$records[[8]]$disease$doid
#> [1] "162"
#> 
#> $content$records[[8]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:162"
#> 
#> 
#> $content$records[[8]]$drugs
#> $content$records[[8]]$drugs[[1]]
#> $content$records[[8]]$drugs[[1]]$id
#> [1] 573
#> 
#> $content$records[[8]]$drugs[[1]]$name
#> [1] "Larotrectinib"
#> 
#> $content$records[[8]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[8]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[8]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[8]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[8]]$evidence_item_count
#> [1] 4
#> 
#> $content$records[[8]]$fda_regulatory_approval
#> [1] TRUE
#> 
#> $content$records[[8]]$status
#> [1] "accepted"
#> 
#> $content$records[[8]]$open_change_count
#> [1] 0
#> 
#> $content$records[[8]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[9]]
#> $content$records[[9]]$id
#> [1] 9
#> 
#> $content$records[[9]]$type
#> [1] "assertion"
#> 
#> $content$records[[9]]$name
#> [1] "AID9"
#> 
#> $content$records[[9]]$summary
#> [1] "Supports diagnosis of diffuse intrinsic pontine glioma."
#> 
#> $content$records[[9]]$description
#> [1] "ACVR1 G328V mutations occur within the kinase domain, leading to activation of downstream signaling. Exclusively seen in high-grade pediatric gliomas, supporting diagnosis of diffuse intrinsic pontine glioma."
#> 
#> $content$records[[9]]$gene
#> $content$records[[9]]$gene$name
#> [1] "ACVR1"
#> 
#> $content$records[[9]]$gene$id
#> [1] 154
#> 
#> 
#> $content$records[[9]]$variant
#> $content$records[[9]]$variant$name
#> [1] "G328V"
#> 
#> $content$records[[9]]$variant$id
#> [1] 1686
#> 
#> 
#> $content$records[[9]]$disease
#> $content$records[[9]]$disease$id
#> [1] 2950
#> 
#> $content$records[[9]]$disease$name
#> [1] "Diffuse Intrinsic Pontine Glioma"
#> 
#> $content$records[[9]]$disease$display_name
#> [1] "Diffuse Intrinsic Pontine Glioma"
#> 
#> $content$records[[9]]$disease$doid
#> NULL
#> 
#> $content$records[[9]]$disease$url
#> [1] "http://www.disease-ontology.org/"
#> 
#> 
#> $content$records[[9]]$drugs
#> list()
#> 
#> $content$records[[9]]$evidence_type
#> [1] "Diagnostic"
#> 
#> $content$records[[9]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[9]]$clinical_significance
#> [1] "Positive"
#> 
#> $content$records[[9]]$evidence_item_count
#> [1] 2
#> 
#> $content$records[[9]]$fda_regulatory_approval
#> [1] FALSE
#> 
#> $content$records[[9]]$status
#> [1] "accepted"
#> 
#> $content$records[[9]]$open_change_count
#> [1] 0
#> 
#> $content$records[[9]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[10]]
#> $content$records[[10]]$id
#> [1] 10
#> 
#> $content$records[[10]]$type
#> [1] "assertion"
#> 
#> $content$records[[10]]$name
#> [1] "AID10"
#> 
#> $content$records[[10]]$summary
#> [1] "BRAF V600E mutant melanoma is sensitive to vemurafenib and cobimetinib combination therapy"
#> 
#> $content$records[[10]]$description
#> [1] "Vemurafenib and cobimetinib combination is an FDA approved and NCCN Category 1 first line treatment for BRAF V600E mutant metastatic melanoma based on clinical data including the Phase III coBRIM trial. NCCN guidelines recommend combination BRAF/MEK inhibitor therapy over BRAF inhibitor monotherapy in this treatment context. Vemurafenib and cobimetinib combination is recommend as Category 2A treatment in second-line or later contexts, and it is recommended to use treatment options different from those used with the patient during first-line therapy. The cobas 4800 BRAF V600 Mutation Test is approved as an FDA companion test for Cotellic (cobimetinib) in combination with Zelboraf (vemurafenib)."
#> 
#> $content$records[[10]]$gene
#> $content$records[[10]]$gene$name
#> [1] "BRAF"
#> 
#> $content$records[[10]]$gene$id
#> [1] 5
#> 
#> 
#> $content$records[[10]]$variant
#> $content$records[[10]]$variant$name
#> [1] "V600E"
#> 
#> $content$records[[10]]$variant$id
#> [1] 12
#> 
#> 
#> $content$records[[10]]$disease
#> $content$records[[10]]$disease$id
#> [1] 7
#> 
#> $content$records[[10]]$disease$name
#> [1] "Melanoma"
#> 
#> $content$records[[10]]$disease$display_name
#> [1] "Melanoma"
#> 
#> $content$records[[10]]$disease$doid
#> [1] "1909"
#> 
#> $content$records[[10]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:1909"
#> 
#> 
#> $content$records[[10]]$drugs
#> $content$records[[10]]$drugs[[1]]
#> $content$records[[10]]$drugs[[1]]$id
#> [1] 4
#> 
#> $content$records[[10]]$drugs[[1]]$name
#> [1] "Vemurafenib"
#> 
#> $content$records[[10]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> $content$records[[10]]$drugs[[2]]
#> $content$records[[10]]$drugs[[2]]$id
#> [1] 342
#> 
#> $content$records[[10]]$drugs[[2]]$name
#> [1] "Cobimetinib"
#> 
#> $content$records[[10]]$drugs[[2]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[10]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[10]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[10]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[10]]$evidence_item_count
#> [1] 3
#> 
#> $content$records[[10]]$fda_regulatory_approval
#> [1] TRUE
#> 
#> $content$records[[10]]$status
#> [1] "accepted"
#> 
#> $content$records[[10]]$open_change_count
#> [1] 0
#> 
#> $content$records[[10]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> 
#> 
#> $url
#> [1] "https://civicdb.org/api/assertions"
#> 
#> $response
#> Response [https://civicdb.org/api/assertions?page=1&count=10]
#>   Date: 2019-09-03 16:44
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 11.9 kB
#> 
#> 
#> attr(,"class")
#> [1] "civic_api"

getAllAssertions(page = 2, count = 10)

#> $content
#> $content$`_meta`
#> $content$`_meta`$current_page
#> [1] 2
#> 
#> $content$`_meta`$per_page
#> [1] "10"
#> 
#> $content$`_meta`$total_pages
#> [1] 3
#> 
#> $content$`_meta`$total_count
#> [1] 27
#> 
#> $content$`_meta`$links
#> $content$`_meta`$links$`next`
#> [1] "https://civicdb.org/api/assertions?count=10&page=3"
#> 
#> $content$`_meta`$links$previous
#> [1] "https://civicdb.org/api/assertions?count=10&page=1"
#> 
#> 
#> 
#> $content$records
#> $content$records[[1]]
#> $content$records[[1]]$id
#> [1] 11
#> 
#> $content$records[[1]]$type
#> [1] "assertion"
#> 
#> $content$records[[1]]$name
#> [1] "AID11"
#> 
#> $content$records[[1]]$summary
#> [1] "BRAF V600K mutant melanoma is sensitive to vemurafenib and cobimetinib combination therapy"
#> 
#> $content$records[[1]]$description
#> [1] "Vemurafenib and cobimentinb is an FDA approved and NCCN Category 1 first line treatment for BRAF V600K mutant metastatic melanoma based on clinical data including the Phase III coBRIM trial. \nNCCN guidelines recommend combination BRAF/MEK inhibitor therapy over BRAF inhibitor monotherapy in this treatment context. BRAF-targeted therapy are recommend as Category 2A treatment options in second-line or later treatment contexts, and it is recommended to use treatment options different from those used with the patient during first-line therapy. The cobas 4800 BRAF V600 Mutation Test is approved as an FDA companion test for Cotellic (cobimetinib) in combination with Zelboraf (vemurafenib)."
#> 
#> $content$records[[1]]$gene
#> $content$records[[1]]$gene$name
#> [1] "BRAF"
#> 
#> $content$records[[1]]$gene$id
#> [1] 5
#> 
#> 
#> $content$records[[1]]$variant
#> $content$records[[1]]$variant$name
#> [1] "V600K"
#> 
#> $content$records[[1]]$variant$id
#> [1] 563
#> 
#> 
#> $content$records[[1]]$disease
#> $content$records[[1]]$disease$id
#> [1] 7
#> 
#> $content$records[[1]]$disease$name
#> [1] "Melanoma"
#> 
#> $content$records[[1]]$disease$display_name
#> [1] "Melanoma"
#> 
#> $content$records[[1]]$disease$doid
#> [1] "1909"
#> 
#> $content$records[[1]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:1909"
#> 
#> 
#> $content$records[[1]]$drugs
#> $content$records[[1]]$drugs[[1]]
#> $content$records[[1]]$drugs[[1]]$id
#> [1] 4
#> 
#> $content$records[[1]]$drugs[[1]]$name
#> [1] "Vemurafenib"
#> 
#> $content$records[[1]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> $content$records[[1]]$drugs[[2]]
#> $content$records[[1]]$drugs[[2]]$id
#> [1] 342
#> 
#> $content$records[[1]]$drugs[[2]]$name
#> [1] "Cobimetinib"
#> 
#> $content$records[[1]]$drugs[[2]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[1]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[1]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[1]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[1]]$evidence_item_count
#> [1] 3
#> 
#> $content$records[[1]]$fda_regulatory_approval
#> [1] TRUE
#> 
#> $content$records[[1]]$status
#> [1] "accepted"
#> 
#> $content$records[[1]]$open_change_count
#> [1] 0
#> 
#> $content$records[[1]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[2]]
#> $content$records[[2]]$id
#> [1] 12
#> 
#> $content$records[[2]]$type
#> [1] "assertion"
#> 
#> $content$records[[2]]$name
#> [1] "AID12"
#> 
#> $content$records[[2]]$summary
#> [1] "BRAF V600E mutant melanoma is sensitive to dabrafenib and trametinib combination therapy"
#> 
#> $content$records[[2]]$description
#> [1] "Combination treatment of BRAF inhibitor dabrafenib and MEK inhibitor trametinib is recommended for adjuvant treatment of stage III or recurrent melanoma with BRAF V600E mutation detected by the approved THxID kit, as well as first line treatment for metastatic melanoma. The treatments are FDA approved and NCCN guidelines recommend these treatments as category 1 based on studies including the Phase III COMBI-V, COMBI-D and COMBI-AD Trials. Combination therapy is now recommended above BRAF inhibitor monotherapy. Dabrafenib and trametinib are recommend as NCCN Category 2A for second line therapy in metastatic melanoma due to lack of clear Phase III trial data for this use case. Cutaneous squamous-cell carcinoma and keratoacanthoma occur at lower rates with combination therapy."
#> 
#> $content$records[[2]]$gene
#> $content$records[[2]]$gene$name
#> [1] "BRAF"
#> 
#> $content$records[[2]]$gene$id
#> [1] 5
#> 
#> 
#> $content$records[[2]]$variant
#> $content$records[[2]]$variant$name
#> [1] "V600E"
#> 
#> $content$records[[2]]$variant$id
#> [1] 12
#> 
#> 
#> $content$records[[2]]$disease
#> $content$records[[2]]$disease$id
#> [1] 7
#> 
#> $content$records[[2]]$disease$name
#> [1] "Melanoma"
#> 
#> $content$records[[2]]$disease$display_name
#> [1] "Melanoma"
#> 
#> $content$records[[2]]$disease$doid
#> [1] "1909"
#> 
#> $content$records[[2]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:1909"
#> 
#> 
#> $content$records[[2]]$drugs
#> $content$records[[2]]$drugs[[1]]
#> $content$records[[2]]$drugs[[1]]$id
#> [1] 19
#> 
#> $content$records[[2]]$drugs[[1]]$name
#> [1] "Trametinib"
#> 
#> $content$records[[2]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> $content$records[[2]]$drugs[[2]]
#> $content$records[[2]]$drugs[[2]]$id
#> [1] 22
#> 
#> $content$records[[2]]$drugs[[2]]$name
#> [1] "Dabrafenib"
#> 
#> $content$records[[2]]$drugs[[2]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[2]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[2]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[2]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[2]]$evidence_item_count
#> [1] 8
#> 
#> $content$records[[2]]$fda_regulatory_approval
#> [1] TRUE
#> 
#> $content$records[[2]]$status
#> [1] "rejected"
#> 
#> $content$records[[2]]$open_change_count
#> [1] 0
#> 
#> $content$records[[2]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[3]]
#> $content$records[[3]]$id
#> [1] 13
#> 
#> $content$records[[3]]$type
#> [1] "assertion"
#> 
#> $content$records[[3]]$name
#> [1] "AID13"
#> 
#> $content$records[[3]]$summary
#> [1] "BRAF V600K mutant melanoma is sensitive to dabrafenib and trametinib combination therapy"
#> 
#> $content$records[[3]]$description
#> [1] "Combination treatment of BRAF inhibitor dabrafenib and MEK inhibitor trametinib is recommended for adjuvant treatment of stage III or recurrent melanoma with the less common BRAF V600K mutation detected by the approved THxID kit, as well as first line treatment for metastatic melanoma. The treatments are FDA approved and NCCN guidelines recommend these treatments as category 1 based on studies including the Phase III COMBI-V, COMBI-D and COMBI-AD Trials. Combination therapy is now recommended above BRAF inhibitor monotherapy. Dabrafenib and trametinib are recommend as NCCN Category 2A for second line therapy in metastatic melanoma due to lack of clear Phase III trial data for this use case. Cutaneous squamous-cell carcinoma and keratoacanthoma occur at lower rates with combination therapy."
#> 
#> $content$records[[3]]$gene
#> $content$records[[3]]$gene$name
#> [1] "BRAF"
#> 
#> $content$records[[3]]$gene$id
#> [1] 5
#> 
#> 
#> $content$records[[3]]$variant
#> $content$records[[3]]$variant$name
#> [1] "V600K"
#> 
#> $content$records[[3]]$variant$id
#> [1] 563
#> 
#> 
#> $content$records[[3]]$disease
#> $content$records[[3]]$disease$id
#> [1] 7
#> 
#> $content$records[[3]]$disease$name
#> [1] "Melanoma"
#> 
#> $content$records[[3]]$disease$display_name
#> [1] "Melanoma"
#> 
#> $content$records[[3]]$disease$doid
#> [1] "1909"
#> 
#> $content$records[[3]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:1909"
#> 
#> 
#> $content$records[[3]]$drugs
#> $content$records[[3]]$drugs[[1]]
#> $content$records[[3]]$drugs[[1]]$id
#> [1] 19
#> 
#> $content$records[[3]]$drugs[[1]]$name
#> [1] "Trametinib"
#> 
#> $content$records[[3]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> $content$records[[3]]$drugs[[2]]
#> $content$records[[3]]$drugs[[2]]$id
#> [1] 22
#> 
#> $content$records[[3]]$drugs[[2]]$name
#> [1] "Dabrafenib"
#> 
#> $content$records[[3]]$drugs[[2]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[3]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[3]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[3]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[3]]$evidence_item_count
#> [1] 3
#> 
#> $content$records[[3]]$fda_regulatory_approval
#> [1] TRUE
#> 
#> $content$records[[3]]$status
#> [1] "accepted"
#> 
#> $content$records[[3]]$open_change_count
#> [1] 0
#> 
#> $content$records[[3]]$pending_evidence_count
#> [1] 1
#> 
#> 
#> $content$records[[4]]
#> $content$records[[4]]$id
#> [1] 14
#> 
#> $content$records[[4]]$type
#> [1] "assertion"
#> 
#> $content$records[[4]]$name
#> [1] "AID14"
#> 
#> $content$records[[4]]$summary
#> [1] "VHL E70K (c.208G>A) is Likely Pathogenic"
#> 
#> $content$records[[4]]$description
#> [1] "E70K missense variant occurs at a very low allele frequency in the general population (0.00001489 allele frequency in ExAC, 4.365e-6 allele frequency in gnomAD) and was previously identified in several unrelated individuals with VHL disease symptoms (see evidence statements).  PM1 (6860); PM2 (per above); PP1 (5805);  PP4 (5805;6742)."
#> 
#> $content$records[[4]]$gene
#> $content$records[[4]]$gene$name
#> [1] "VHL"
#> 
#> $content$records[[4]]$gene$id
#> [1] 58
#> 
#> 
#> $content$records[[4]]$variant
#> $content$records[[4]]$variant$name
#> [1] "E70K (c.208G>A)"
#> 
#> $content$records[[4]]$variant$id
#> [1] 1956
#> 
#> 
#> $content$records[[4]]$disease
#> $content$records[[4]]$disease$id
#> [1] 2198
#> 
#> $content$records[[4]]$disease$name
#> [1] "Von Hippel-Lindau Disease"
#> 
#> $content$records[[4]]$disease$display_name
#> [1] "Von Hippel-Lindau Disease"
#> 
#> $content$records[[4]]$disease$doid
#> [1] "14175"
#> 
#> $content$records[[4]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:14175"
#> 
#> 
#> $content$records[[4]]$drugs
#> list()
#> 
#> $content$records[[4]]$evidence_type
#> [1] "Predisposing"
#> 
#> $content$records[[4]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[4]]$clinical_significance
#> [1] "Likely Pathogenic"
#> 
#> $content$records[[4]]$evidence_item_count
#> [1] 6
#> 
#> $content$records[[4]]$fda_regulatory_approval
#> [1] FALSE
#> 
#> $content$records[[4]]$status
#> [1] "accepted"
#> 
#> $content$records[[4]]$open_change_count
#> [1] 0
#> 
#> $content$records[[4]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[5]]
#> $content$records[[5]]$id
#> [1] 15
#> 
#> $content$records[[5]]$type
#> [1] "assertion"
#> 
#> $content$records[[5]]$name
#> [1] "AID15"
#> 
#> $content$records[[5]]$summary
#> [1] "Patients with PML-RARA fusions have improved outcomes following ATRA treatment."
#> 
#> $content$records[[5]]$description
#> [1] "The most recent WHO classification of myeloid neoplasms and acute leukemia changed the definition of APL from the cytogenetic criteria of t(15;17) to the molecular definition of “APL with PML-RARA” to be inclusive of complex or cryptic rearrangements that lead to a functional transcription factor. The incorporation of all-trans retinoic acid (ATRA) and the use of risk stratification (based on WBC counts) in the management of APL has largely improved outcomes for patients with this subtype (NCCN guidelines v3.2018)"
#> 
#> $content$records[[5]]$gene
#> $content$records[[5]]$gene$name
#> [1] "PML"
#> 
#> $content$records[[5]]$gene$id
#> [1] 39
#> 
#> 
#> $content$records[[5]]$variant
#> $content$records[[5]]$variant$name
#> [1] "PML-RARA"
#> 
#> $content$records[[5]]$variant$id
#> [1] 108
#> 
#> 
#> $content$records[[5]]$disease
#> $content$records[[5]]$disease$id
#> [1] 379
#> 
#> $content$records[[5]]$disease$name
#> [1] "Acute Promyelocytic Leukemia"
#> 
#> $content$records[[5]]$disease$display_name
#> [1] "Acute Promyelocytic Leukemia"
#> 
#> $content$records[[5]]$disease$doid
#> [1] "0060318"
#> 
#> $content$records[[5]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:0060318"
#> 
#> 
#> $content$records[[5]]$drugs
#> $content$records[[5]]$drugs[[1]]
#> $content$records[[5]]$drugs[[1]]$id
#> [1] 265
#> 
#> $content$records[[5]]$drugs[[1]]$name
#> [1] "All-trans Retinoic Acid"
#> 
#> $content$records[[5]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[5]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[5]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[5]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[5]]$evidence_item_count
#> [1] 2
#> 
#> $content$records[[5]]$fda_regulatory_approval
#> [1] FALSE
#> 
#> $content$records[[5]]$status
#> [1] "submitted"
#> 
#> $content$records[[5]]$open_change_count
#> [1] 0
#> 
#> $content$records[[5]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[6]]
#> $content$records[[6]]$id
#> [1] 16
#> 
#> $content$records[[6]]$type
#> [1] "assertion"
#> 
#> $content$records[[6]]$name
#> [1] "AID16"
#> 
#> $content$records[[6]]$summary
#> [1] "Adult Patients With De Novo Acute Myeloid Leukemia and t(9; 11)(p22; q23) Have a Superior Outcome to Patients With Other Translocations Involving Band 11q23: A Cancer and Leukemia Group B Study (PMID:9373264)"
#> 
#> $content$records[[6]]$description
#> [1] "AML with t(9;11)(p21.3;q23.3) has an intermediate survival, superior to that of AML with other 11q23.3 translocations (PMID:9373264) (WHO 2016)"
#> 
#> $content$records[[6]]$gene
#> $content$records[[6]]$gene$name
#> [1] "KMT2A"
#> 
#> $content$records[[6]]$gene$id
#> [1] 3537
#> 
#> 
#> $content$records[[6]]$variant
#> $content$records[[6]]$variant$name
#> [1] "MLL-MLLT3"
#> 
#> $content$records[[6]]$variant$id
#> [1] 432
#> 
#> 
#> $content$records[[6]]$disease
#> $content$records[[6]]$disease$id
#> [1] 3
#> 
#> $content$records[[6]]$disease$name
#> [1] "Acute Myeloid Leukemia"
#> 
#> $content$records[[6]]$disease$display_name
#> [1] "Acute Myeloid Leukemia"
#> 
#> $content$records[[6]]$disease$doid
#> [1] "9119"
#> 
#> $content$records[[6]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:9119"
#> 
#> 
#> $content$records[[6]]$drugs
#> list()
#> 
#> $content$records[[6]]$evidence_type
#> [1] "Prognostic"
#> 
#> $content$records[[6]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[6]]$clinical_significance
#> [1] "Better Outcome"
#> 
#> $content$records[[6]]$evidence_item_count
#> [1] 0
#> 
#> $content$records[[6]]$fda_regulatory_approval
#> [1] FALSE
#> 
#> $content$records[[6]]$status
#> [1] "submitted"
#> 
#> $content$records[[6]]$open_change_count
#> [1] 0
#> 
#> $content$records[[6]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[7]]
#> $content$records[[7]]$id
#> [1] 17
#> 
#> $content$records[[7]]$type
#> [1] "assertion"
#> 
#> $content$records[[7]]$name
#> [1] "AID17"
#> 
#> $content$records[[7]]$summary
#> [1] "The inframe variant, F76del, is pathogenic for Von Hippel-Lindau Disease."
#> 
#> $content$records[[7]]$description
#> [1] "EID5682 shows a large family with the variant cosegregating with affected individuals (PP1). However, confirmed de novo mutations are also described EID5340 (PS2). Both are supported by several other reports with familial and sporadic VHL and this variant. This inframe deletion is not in a repetitive region (PM4) and absent from gnomAD v2.1 (PM2)."
#> 
#> $content$records[[7]]$gene
#> $content$records[[7]]$gene$name
#> [1] "VHL"
#> 
#> $content$records[[7]]$gene$id
#> [1] 58
#> 
#> 
#> $content$records[[7]]$variant
#> $content$records[[7]]$variant$name
#> [1] "F76del (c.227_229delTCT)"
#> 
#> $content$records[[7]]$variant$id
#> [1] 1926
#> 
#> 
#> $content$records[[7]]$disease
#> $content$records[[7]]$disease$id
#> [1] 2198
#> 
#> $content$records[[7]]$disease$name
#> [1] "Von Hippel-Lindau Disease"
#> 
#> $content$records[[7]]$disease$display_name
#> [1] "Von Hippel-Lindau Disease"
#> 
#> $content$records[[7]]$disease$doid
#> [1] "14175"
#> 
#> $content$records[[7]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:14175"
#> 
#> 
#> $content$records[[7]]$drugs
#> list()
#> 
#> $content$records[[7]]$evidence_type
#> [1] "Predisposing"
#> 
#> $content$records[[7]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[7]]$clinical_significance
#> [1] "Pathogenic"
#> 
#> $content$records[[7]]$evidence_item_count
#> [1] 13
#> 
#> $content$records[[7]]$fda_regulatory_approval
#> [1] FALSE
#> 
#> $content$records[[7]]$status
#> [1] "accepted"
#> 
#> $content$records[[7]]$open_change_count
#> [1] 0
#> 
#> $content$records[[7]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[8]]
#> $content$records[[8]]$id
#> [1] 18
#> 
#> $content$records[[8]]$type
#> [1] "assertion"
#> 
#> $content$records[[8]]$name
#> [1] "AID18"
#> 
#> $content$records[[8]]$summary
#> [1] "VHL nonsense variant Q195* (c.583C>T) is pathogenic for Von Hippel Landau Disease"
#> 
#> $content$records[[8]]$description
#> [1] "VHL variant Q195* (c.583C>T) introduces an early stop codon resulting in a truncated protein. Loss of function VHL variants have been implicated with pathogenicity for Von Hippel Landau disease (PVS1), and this variant has been observed in patients with symptoms and family history highly characteristic of Von Hippel Landau disease (PP4). The variant does not appear in the gnomAD population database (PM2)."
#> 
#> $content$records[[8]]$gene
#> $content$records[[8]]$gene$name
#> [1] "VHL"
#> 
#> $content$records[[8]]$gene$id
#> [1] 58
#> 
#> 
#> $content$records[[8]]$variant
#> $content$records[[8]]$variant$name
#> [1] "Q195* (c.583C>T)"
#> 
#> $content$records[[8]]$variant$id
#> [1] 1810
#> 
#> 
#> $content$records[[8]]$disease
#> $content$records[[8]]$disease$id
#> [1] 2198
#> 
#> $content$records[[8]]$disease$name
#> [1] "Von Hippel-Lindau Disease"
#> 
#> $content$records[[8]]$disease$display_name
#> [1] "Von Hippel-Lindau Disease"
#> 
#> $content$records[[8]]$disease$doid
#> [1] "14175"
#> 
#> $content$records[[8]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:14175"
#> 
#> 
#> $content$records[[8]]$drugs
#> list()
#> 
#> $content$records[[8]]$evidence_type
#> [1] "Predisposing"
#> 
#> $content$records[[8]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[8]]$clinical_significance
#> [1] "Pathogenic"
#> 
#> $content$records[[8]]$evidence_item_count
#> [1] 9
#> 
#> $content$records[[8]]$fda_regulatory_approval
#> [1] FALSE
#> 
#> $content$records[[8]]$status
#> [1] "accepted"
#> 
#> $content$records[[8]]$open_change_count
#> [1] 0
#> 
#> $content$records[[8]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> $content$records[[9]]
#> $content$records[[9]]$id
#> [1] 19
#> 
#> $content$records[[9]]$type
#> [1] "assertion"
#> 
#> $content$records[[9]]$name
#> [1] "AID19"
#> 
#> $content$records[[9]]$summary
#> [1] "EBF1-PDGFRB is associated with Ph-like ALL and predicts sensitivity to TKIs"
#> 
#> $content$records[[9]]$description
#> [1] "The EBF1-PDGFRB fusion is associated with Philadelphia-like molecular subtype of B-ALL. B-ALLs characterized by this fusion show poor response to conventional chemotherapy, with an increased proportion of cases being completely resistant, showing residual disease at the end of induction or relapsing. However, a very good response of EBF1-PDGFRB-positive B-ALL  to ABL-class tyrosine kinase inhibitors (imatinib and dasatinib) in combination with standard chemotherapy has been shown in a small number of cases. Clinical trials for ph-like ALL are ongoing."
#> 
#> $content$records[[9]]$gene
#> $content$records[[9]]$gene$name
#> [1] "PDGFRB"
#> 
#> $content$records[[9]]$gene$id
#> [1] 4176
#> 
#> 
#> $content$records[[9]]$variant
#> $content$records[[9]]$variant$name
#> [1] "EBF1-PDGFRB"
#> 
#> $content$records[[9]]$variant$id
#> [1] 535
#> 
#> 
#> $content$records[[9]]$disease
#> $content$records[[9]]$disease$id
#> [1] 2955
#> 
#> $content$records[[9]]$disease$name
#> [1] "Pediatric B-cell Acute Lymphoblastic Leukemia"
#> 
#> $content$records[[9]]$disease$display_name
#> [1] "Pediatric B-cell Acute Lymphoblastic Leukemia"
#> 
#> $content$records[[9]]$disease$doid
#> NULL
#> 
#> $content$records[[9]]$disease$url
#> [1] "http://www.disease-ontology.org/"
#> 
#> 
#> $content$records[[9]]$drugs
#> $content$records[[9]]$drugs[[1]]
#> $content$records[[9]]$drugs[[1]]$id
#> [1] 5
#> 
#> $content$records[[9]]$drugs[[1]]$name
#> [1] "Imatinib"
#> 
#> $content$records[[9]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> $content$records[[9]]$drugs[[2]]
#> $content$records[[9]]$drugs[[2]]$id
#> [1] 20
#> 
#> $content$records[[9]]$drugs[[2]]$name
#> [1] "Dasatinib"
#> 
#> $content$records[[9]]$drugs[[2]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[9]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[9]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[9]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[9]]$evidence_item_count
#> [1] 4
#> 
#> $content$records[[9]]$fda_regulatory_approval
#> [1] FALSE
#> 
#> $content$records[[9]]$status
#> [1] "submitted"
#> 
#> $content$records[[9]]$open_change_count
#> [1] 0
#> 
#> $content$records[[9]]$pending_evidence_count
#> [1] 1
#> 
#> 
#> $content$records[[10]]
#> $content$records[[10]]$id
#> [1] 20
#> 
#> $content$records[[10]]$type
#> [1] "assertion"
#> 
#> $content$records[[10]]$name
#> [1] "AID20"
#> 
#> $content$records[[10]]$summary
#> [1] "BRAF V600E indicates poor prognosis in advanced colorectal cancer"
#> 
#> $content$records[[10]]$description
#> [1] "BRAF V600E was associated with worse prognosis in Phase II and III colorectal cancer, with a stronger effect in MSI-Low or MSI-Stable tumors. In metastatic CRC, V600E was associated with worse prognosis, and meta-analysis showed BRAF mutation in CRC associated with multiple negative prognostic markers. \nNCCN Guidelines state that that\nmutations in BRAF are a strong prognostic marker, and recommend BRAF genotyping of either primary or metastatic tumor tissue at diagnosis of stage IV disease."
#> 
#> $content$records[[10]]$gene
#> $content$records[[10]]$gene$name
#> [1] "BRAF"
#> 
#> $content$records[[10]]$gene$id
#> [1] 5
#> 
#> 
#> $content$records[[10]]$variant
#> $content$records[[10]]$variant$name
#> [1] "V600E"
#> 
#> $content$records[[10]]$variant$id
#> [1] 12
#> 
#> 
#> $content$records[[10]]$disease
#> $content$records[[10]]$disease$id
#> [1] 11
#> 
#> $content$records[[10]]$disease$name
#> [1] "Colorectal Cancer"
#> 
#> $content$records[[10]]$disease$display_name
#> [1] "Colorectal Cancer"
#> 
#> $content$records[[10]]$disease$doid
#> [1] "9256"
#> 
#> $content$records[[10]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:9256"
#> 
#> 
#> $content$records[[10]]$drugs
#> list()
#> 
#> $content$records[[10]]$evidence_type
#> [1] "Prognostic"
#> 
#> $content$records[[10]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[10]]$clinical_significance
#> [1] "Poor Outcome"
#> 
#> $content$records[[10]]$evidence_item_count
#> [1] 6
#> 
#> $content$records[[10]]$fda_regulatory_approval
#> [1] FALSE
#> 
#> $content$records[[10]]$status
#> [1] "accepted"
#> 
#> $content$records[[10]]$open_change_count
#> [1] 0
#> 
#> $content$records[[10]]$pending_evidence_count
#> [1] 0
#> 
#> 
#> 
#> 
#> $url
#> [1] "https://civicdb.org/api/assertions"
#> 
#> $response
#> Response [https://civicdb.org/api/assertions?page=2&count=10]
#>   Date: 2019-09-03 16:44
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 12.1 kB
#> 
#> 
#> attr(,"class")
#> [1] "civic_api"