Get a list of evidence items

Retrieve all evidence items from the CIViC DB

getAllEvidenceItems(page = 1, count = 25)

Arguments

page
the page number to retrieve

count
the number of evidence items to retrieve

Value

An S3 Object of type civic_api containing the content, url, and response

Examples

getAllEvidenceItems(count = 10)

#> $content
#> $content$`_meta`
#> $content$`_meta`$current_page
#> [1] 1
#> 
#> $content$`_meta`$per_page
#> [1] "10"
#> 
#> $content$`_meta`$total_pages
#> [1] 642
#> 
#> $content$`_meta`$total_count
#> [1] 6411
#> 
#> $content$`_meta`$links
#> $content$`_meta`$links$`next`
#> [1] "https://civicdb.org/api/evidence_items?count=10&page=2"
#> 
#> $content$`_meta`$links$previous
#> NULL
#> 
#> 
#> 
#> $content$records
#> $content$records[[1]]
#> $content$records[[1]]$id
#> [1] 1
#> 
#> $content$records[[1]]$name
#> [1] "EID1"
#> 
#> $content$records[[1]]$description
#> [1] "JAK2 V617F is not associated with lymphoid leukemia (B-lineage ALL, T-ALL or CLL)."
#> 
#> $content$records[[1]]$disease
#> $content$records[[1]]$disease$id
#> [1] 1
#> 
#> $content$records[[1]]$disease$name
#> [1] "Lymphoid Leukemia"
#> 
#> $content$records[[1]]$disease$display_name
#> [1] "Lymphoid Leukemia"
#> 
#> $content$records[[1]]$disease$doid
#> [1] "10747"
#> 
#> $content$records[[1]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:10747"
#> 
#> 
#> $content$records[[1]]$drugs
#> list()
#> 
#> $content$records[[1]]$rating
#> [1] 4
#> 
#> $content$records[[1]]$evidence_level
#> [1] "B"
#> 
#> $content$records[[1]]$evidence_type
#> [1] "Diagnostic"
#> 
#> $content$records[[1]]$clinical_significance
#> [1] "Negative"
#> 
#> $content$records[[1]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[1]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[1]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[1]]$status
#> [1] "accepted"
#> 
#> $content$records[[1]]$open_change_count
#> [1] 0
#> 
#> $content$records[[1]]$type
#> [1] "evidence"
#> 
#> $content$records[[1]]$source
#> $content$records[[1]]$source$id
#> [1] 51
#> 
#> $content$records[[1]]$source$name
#> [1] "The JAK2V617F activating mutation occurs in chronic myelomonocytic leukemia and acute myeloid leukemia, but not in acute lymphoblastic leukemia or chronic lymphocytic leukemia."
#> 
#> $content$records[[1]]$source$citation
#> [1] "Levine et al., 2005, Blood"
#> 
#> $content$records[[1]]$source$citation_id
#> [1] "16081687"
#> 
#> $content$records[[1]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[1]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[1]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/16081687"
#> 
#> $content$records[[1]]$source$open_access
#> [1] TRUE
#> 
#> $content$records[[1]]$source$pmc_id
#> [1] "PMC1895066"
#> 
#> $content$records[[1]]$source$publication_date
#> $content$records[[1]]$source$publication_date$year
#> [1] 2005
#> 
#> $content$records[[1]]$source$publication_date$month
#> [1] 11
#> 
#> $content$records[[1]]$source$publication_date$day
#> [1] 15
#> 
#> 
#> $content$records[[1]]$source$journal
#> [1] "Blood"
#> 
#> $content$records[[1]]$source$full_journal_title
#> [1] "Blood"
#> 
#> $content$records[[1]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[1]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[1]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[1]]$variant_id
#> [1] 64
#> 
#> $content$records[[1]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[2]]
#> $content$records[[2]]$id
#> [1] 2
#> 
#> $content$records[[2]]$name
#> [1] "EID2"
#> 
#> $content$records[[2]]$description
#> [1] "GIST tumors harboring PDGFRA D842V mutation are more likely to be benign than malignant."
#> 
#> $content$records[[2]]$disease
#> $content$records[[2]]$disease$id
#> [1] 2
#> 
#> $content$records[[2]]$disease$name
#> [1] "Gastrointestinal Stromal Tumor"
#> 
#> $content$records[[2]]$disease$display_name
#> [1] "Gastrointestinal Stromal Tumor"
#> 
#> $content$records[[2]]$disease$doid
#> [1] "9253"
#> 
#> $content$records[[2]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:9253"
#> 
#> 
#> $content$records[[2]]$drugs
#> list()
#> 
#> $content$records[[2]]$rating
#> [1] 3
#> 
#> $content$records[[2]]$evidence_level
#> [1] "B"
#> 
#> $content$records[[2]]$evidence_type
#> [1] "Diagnostic"
#> 
#> $content$records[[2]]$clinical_significance
#> [1] "Negative"
#> 
#> $content$records[[2]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[2]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[2]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[2]]$status
#> [1] "accepted"
#> 
#> $content$records[[2]]$open_change_count
#> [1] 0
#> 
#> $content$records[[2]]$type
#> [1] "evidence"
#> 
#> $content$records[[2]]$source
#> $content$records[[2]]$source$id
#> [1] 52
#> 
#> $content$records[[2]]$source$name
#> [1] "A great majority of GISTs with PDGFRA mutations represent gastric tumors of low or no malignant potential."
#> 
#> $content$records[[2]]$source$citation
#> [1] "Lasota et al., 2004, Lab. Invest."
#> 
#> $content$records[[2]]$source$citation_id
#> [1] "15146165"
#> 
#> $content$records[[2]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[2]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[2]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/15146165"
#> 
#> $content$records[[2]]$source$open_access
#> NULL
#> 
#> $content$records[[2]]$source$pmc_id
#> NULL
#> 
#> $content$records[[2]]$source$publication_date
#> $content$records[[2]]$source$publication_date$year
#> [1] 2004
#> 
#> $content$records[[2]]$source$publication_date$month
#> [1] 7
#> 
#> 
#> $content$records[[2]]$source$journal
#> [1] "Lab. Invest."
#> 
#> $content$records[[2]]$source$full_journal_title
#> [1] "Laboratory investigation; a journal of technical methods and pathology"
#> 
#> $content$records[[2]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[2]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[2]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[2]]$variant_id
#> [1] 99
#> 
#> $content$records[[2]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[3]]
#> $content$records[[3]]$id
#> [1] 3
#> 
#> $content$records[[3]]$name
#> [1] "EID3"
#> 
#> $content$records[[3]]$description
#> [1] "DNMT3A R882 mutations occur most often in de novo AML patients with intermediate risk cytogenetics (39/194 intermediate risk patients vs 0/89 low and high risk)."
#> 
#> $content$records[[3]]$disease
#> $content$records[[3]]$disease$id
#> [1] 3
#> 
#> $content$records[[3]]$disease$name
#> [1] "Acute Myeloid Leukemia"
#> 
#> $content$records[[3]]$disease$display_name
#> [1] "Acute Myeloid Leukemia"
#> 
#> $content$records[[3]]$disease$doid
#> [1] "9119"
#> 
#> $content$records[[3]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:9119"
#> 
#> 
#> $content$records[[3]]$drugs
#> list()
#> 
#> $content$records[[3]]$rating
#> [1] 2
#> 
#> $content$records[[3]]$evidence_level
#> [1] "B"
#> 
#> $content$records[[3]]$evidence_type
#> [1] "Diagnostic"
#> 
#> $content$records[[3]]$clinical_significance
#> [1] "Positive"
#> 
#> $content$records[[3]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[3]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[3]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[3]]$status
#> [1] "accepted"
#> 
#> $content$records[[3]]$open_change_count
#> [1] 0
#> 
#> $content$records[[3]]$type
#> [1] "evidence"
#> 
#> $content$records[[3]]$source
#> $content$records[[3]]$source$id
#> [1] 53
#> 
#> $content$records[[3]]$source$name
#> [1] "Do AML patients with DNMT3A exon 23 mutations benefit from idarubicin as compared to daunorubicin? A single center experience."
#> 
#> $content$records[[3]]$source$citation
#> [1] "LaRochelle et al., 2011, Oncotarget"
#> 
#> $content$records[[3]]$source$citation_id
#> [1] "22081665"
#> 
#> $content$records[[3]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[3]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[3]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/22081665"
#> 
#> $content$records[[3]]$source$open_access
#> [1] TRUE
#> 
#> $content$records[[3]]$source$pmc_id
#> [1] "PMC3260002"
#> 
#> $content$records[[3]]$source$publication_date
#> $content$records[[3]]$source$publication_date$year
#> [1] 2011
#> 
#> $content$records[[3]]$source$publication_date$month
#> [1] 11
#> 
#> 
#> $content$records[[3]]$source$journal
#> [1] "Oncotarget"
#> 
#> $content$records[[3]]$source$full_journal_title
#> [1] "Oncotarget"
#> 
#> $content$records[[3]]$source$status
#> [1] "partially curated"
#> 
#> $content$records[[3]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[3]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[3]]$variant_id
#> [1] 32
#> 
#> $content$records[[3]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[4]]
#> $content$records[[4]]$id
#> [1] 4
#> 
#> $content$records[[4]]$name
#> [1] "EID4"
#> 
#> $content$records[[4]]$description
#> [1] "Young AML patients (<60 years old) with DNMT3A mutations (60% of which were R882) were older in age, had higher white blood cell counts and had higher platelet counts than patients wildtype for DNMT3A."
#> 
#> $content$records[[4]]$disease
#> $content$records[[4]]$disease$id
#> [1] 3
#> 
#> $content$records[[4]]$disease$name
#> [1] "Acute Myeloid Leukemia"
#> 
#> $content$records[[4]]$disease$display_name
#> [1] "Acute Myeloid Leukemia"
#> 
#> $content$records[[4]]$disease$doid
#> [1] "9119"
#> 
#> $content$records[[4]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:9119"
#> 
#> 
#> $content$records[[4]]$drugs
#> list()
#> 
#> $content$records[[4]]$rating
#> [1] 3
#> 
#> $content$records[[4]]$evidence_level
#> [1] "B"
#> 
#> $content$records[[4]]$evidence_type
#> [1] "Diagnostic"
#> 
#> $content$records[[4]]$clinical_significance
#> [1] "Positive"
#> 
#> $content$records[[4]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[4]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[4]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[4]]$status
#> [1] "accepted"
#> 
#> $content$records[[4]]$open_change_count
#> [1] 0
#> 
#> $content$records[[4]]$type
#> [1] "evidence"
#> 
#> $content$records[[4]]$source
#> $content$records[[4]]$source$id
#> [1] 54
#> 
#> $content$records[[4]]$source$name
#> [1] "Mutant DNMT3A: a marker of poor prognosis in acute myeloid leukemia."
#> 
#> $content$records[[4]]$source$citation
#> [1] "Ribeiro et al., 2012, Blood"
#> 
#> $content$records[[4]]$source$citation_id
#> [1] "22490330"
#> 
#> $content$records[[4]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[4]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[4]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/22490330"
#> 
#> $content$records[[4]]$source$open_access
#> NULL
#> 
#> $content$records[[4]]$source$pmc_id
#> NULL
#> 
#> $content$records[[4]]$source$publication_date
#> $content$records[[4]]$source$publication_date$year
#> [1] 2012
#> 
#> $content$records[[4]]$source$publication_date$month
#> [1] 6
#> 
#> $content$records[[4]]$source$publication_date$day
#> [1] 14
#> 
#> 
#> $content$records[[4]]$source$journal
#> [1] "Blood"
#> 
#> $content$records[[4]]$source$full_journal_title
#> [1] "Blood"
#> 
#> $content$records[[4]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[4]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[4]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[4]]$variant_id
#> [1] 32
#> 
#> $content$records[[4]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[5]]
#> $content$records[[5]]$id
#> [1] 5
#> 
#> $content$records[[5]]$name
#> [1] "EID5"
#> 
#> $content$records[[5]]$description
#> [1] "JAK2 V617F is associated with myeloid malignanices (AML, MDS, CMML/atypical CML)."
#> 
#> $content$records[[5]]$disease
#> $content$records[[5]]$disease$id
#> [1] 4
#> 
#> $content$records[[5]]$disease$name
#> [1] "Chronic Myeloid Leukemia"
#> 
#> $content$records[[5]]$disease$display_name
#> [1] "Chronic Myeloid Leukemia"
#> 
#> $content$records[[5]]$disease$doid
#> [1] "8552"
#> 
#> $content$records[[5]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:8552"
#> 
#> 
#> $content$records[[5]]$drugs
#> list()
#> 
#> $content$records[[5]]$rating
#> [1] 4
#> 
#> $content$records[[5]]$evidence_level
#> [1] "B"
#> 
#> $content$records[[5]]$evidence_type
#> [1] "Diagnostic"
#> 
#> $content$records[[5]]$clinical_significance
#> [1] "Positive"
#> 
#> $content$records[[5]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[5]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[5]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[5]]$status
#> [1] "accepted"
#> 
#> $content$records[[5]]$open_change_count
#> [1] 0
#> 
#> $content$records[[5]]$type
#> [1] "evidence"
#> 
#> $content$records[[5]]$source
#> $content$records[[5]]$source$id
#> [1] 51
#> 
#> $content$records[[5]]$source$name
#> [1] "The JAK2V617F activating mutation occurs in chronic myelomonocytic leukemia and acute myeloid leukemia, but not in acute lymphoblastic leukemia or chronic lymphocytic leukemia."
#> 
#> $content$records[[5]]$source$citation
#> [1] "Levine et al., 2005, Blood"
#> 
#> $content$records[[5]]$source$citation_id
#> [1] "16081687"
#> 
#> $content$records[[5]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[5]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[5]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/16081687"
#> 
#> $content$records[[5]]$source$open_access
#> [1] TRUE
#> 
#> $content$records[[5]]$source$pmc_id
#> [1] "PMC1895066"
#> 
#> $content$records[[5]]$source$publication_date
#> $content$records[[5]]$source$publication_date$year
#> [1] 2005
#> 
#> $content$records[[5]]$source$publication_date$month
#> [1] 11
#> 
#> $content$records[[5]]$source$publication_date$day
#> [1] 15
#> 
#> 
#> $content$records[[5]]$source$journal
#> [1] "Blood"
#> 
#> $content$records[[5]]$source$full_journal_title
#> [1] "Blood"
#> 
#> $content$records[[5]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[5]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[5]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[5]]$variant_id
#> [1] 64
#> 
#> $content$records[[5]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[6]]
#> $content$records[[6]]$id
#> [1] 7
#> 
#> $content$records[[6]]$name
#> [1] "EID7"
#> 
#> $content$records[[6]]$description
#> [1] "The SNP rs10974944 in the Jak2 locus is associated with increased predisposition for JAK2 V617F mutation and its associated cancer."
#> 
#> $content$records[[6]]$disease
#> $content$records[[6]]$disease$id
#> [1] 5
#> 
#> $content$records[[6]]$disease$name
#> [1] "Bone Marrow Cancer"
#> 
#> $content$records[[6]]$disease$display_name
#> [1] "Bone Marrow Cancer"
#> 
#> $content$records[[6]]$disease$doid
#> [1] "4960"
#> 
#> $content$records[[6]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:4960"
#> 
#> 
#> $content$records[[6]]$drugs
#> list()
#> 
#> $content$records[[6]]$rating
#> [1] 4
#> 
#> $content$records[[6]]$evidence_level
#> [1] "B"
#> 
#> $content$records[[6]]$evidence_type
#> [1] "Diagnostic"
#> 
#> $content$records[[6]]$clinical_significance
#> [1] "Positive"
#> 
#> $content$records[[6]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[6]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[6]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[6]]$status
#> [1] "accepted"
#> 
#> $content$records[[6]]$open_change_count
#> [1] 0
#> 
#> $content$records[[6]]$type
#> [1] "evidence"
#> 
#> $content$records[[6]]$source
#> $content$records[[6]]$source$id
#> [1] 55
#> 
#> $content$records[[6]]$source$name
#> [1] "A germline JAK2 SNP is associated with predisposition to the development of JAK2(V617F)-positive myeloproliferative neoplasms."
#> 
#> $content$records[[6]]$source$citation
#> [1] "Kilpivaara et al., 2009, Nat. Genet."
#> 
#> $content$records[[6]]$source$citation_id
#> [1] "19287384"
#> 
#> $content$records[[6]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[6]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[6]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/19287384"
#> 
#> $content$records[[6]]$source$open_access
#> [1] TRUE
#> 
#> $content$records[[6]]$source$pmc_id
#> [1] "PMC3676425"
#> 
#> $content$records[[6]]$source$publication_date
#> $content$records[[6]]$source$publication_date$year
#> [1] 2009
#> 
#> $content$records[[6]]$source$publication_date$month
#> [1] 4
#> 
#> 
#> $content$records[[6]]$source$journal
#> [1] "Nat. Genet."
#> 
#> $content$records[[6]]$source$full_journal_title
#> [1] "Nature genetics"
#> 
#> $content$records[[6]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[6]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[6]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[6]]$variant_id
#> [1] 64
#> 
#> $content$records[[6]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[7]]
#> $content$records[[7]]$id
#> [1] 8
#> 
#> $content$records[[7]]$name
#> [1] "EID8"
#> 
#> $content$records[[7]]$description
#> [1] "Children with early age leukaemia who had second hand smoke exposure are more likely to harbor KRAS mutation."
#> 
#> $content$records[[7]]$disease
#> $content$records[[7]]$disease$id
#> [1] 6
#> 
#> $content$records[[7]]$disease$name
#> [1] "Acute Leukemia"
#> 
#> $content$records[[7]]$disease$display_name
#> [1] "Acute Leukemia"
#> 
#> $content$records[[7]]$disease$doid
#> [1] "12603"
#> 
#> $content$records[[7]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:12603"
#> 
#> 
#> $content$records[[7]]$drugs
#> list()
#> 
#> $content$records[[7]]$rating
#> [1] 3
#> 
#> $content$records[[7]]$evidence_level
#> [1] "B"
#> 
#> $content$records[[7]]$evidence_type
#> [1] "Diagnostic"
#> 
#> $content$records[[7]]$clinical_significance
#> [1] "Positive"
#> 
#> $content$records[[7]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[7]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[7]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[7]]$status
#> [1] "accepted"
#> 
#> $content$records[[7]]$open_change_count
#> [1] 0
#> 
#> $content$records[[7]]$type
#> [1] "evidence"
#> 
#> $content$records[[7]]$source
#> $content$records[[7]]$source$id
#> [1] 56
#> 
#> $content$records[[7]]$source$name
#> [1] "RAS mutations in early age leukaemia modulated by NQO1 rs1800566 (C609T) are associated with second-hand smoking exposures."
#> 
#> $content$records[[7]]$source$citation
#> [1] "Andrade et al., 2014, BMC Cancer"
#> 
#> $content$records[[7]]$source$citation_id
#> [1] "24571676"
#> 
#> $content$records[[7]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[7]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[7]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/24571676"
#> 
#> $content$records[[7]]$source$open_access
#> [1] TRUE
#> 
#> $content$records[[7]]$source$pmc_id
#> [1] "PMC3946262"
#> 
#> $content$records[[7]]$source$publication_date
#> $content$records[[7]]$source$publication_date$year
#> [1] 2014
#> 
#> 
#> $content$records[[7]]$source$journal
#> [1] "BMC Cancer"
#> 
#> $content$records[[7]]$source$full_journal_title
#> [1] "BMC cancer"
#> 
#> $content$records[[7]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[7]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[7]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[7]]$variant_id
#> [1] 76
#> 
#> $content$records[[7]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[8]]
#> $content$records[[8]]$id
#> [1] 10
#> 
#> $content$records[[8]]$name
#> [1] "EID10"
#> 
#> $content$records[[8]]$description
#> [1] "Melanoma associated with NRAS Q61 mutation was more often associated with those at the extremity than those at the trunk."
#> 
#> $content$records[[8]]$disease
#> $content$records[[8]]$disease$id
#> [1] 7
#> 
#> $content$records[[8]]$disease$name
#> [1] "Melanoma"
#> 
#> $content$records[[8]]$disease$display_name
#> [1] "Melanoma"
#> 
#> $content$records[[8]]$disease$doid
#> [1] "1909"
#> 
#> $content$records[[8]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:1909"
#> 
#> 
#> $content$records[[8]]$drugs
#> list()
#> 
#> $content$records[[8]]$rating
#> [1] 3
#> 
#> $content$records[[8]]$evidence_level
#> [1] "B"
#> 
#> $content$records[[8]]$evidence_type
#> [1] "Diagnostic"
#> 
#> $content$records[[8]]$clinical_significance
#> [1] "Positive"
#> 
#> $content$records[[8]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[8]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[8]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[8]]$status
#> [1] "accepted"
#> 
#> $content$records[[8]]$open_change_count
#> [1] 0
#> 
#> $content$records[[8]]$type
#> [1] "evidence"
#> 
#> $content$records[[8]]$source
#> $content$records[[8]]$source$id
#> [1] 57
#> 
#> $content$records[[8]]$source$name
#> [1] "NRAS and BRAF mutations in melanoma-associated nevi and uninvolved nevi."
#> 
#> $content$records[[8]]$source$citation
#> [1] "Tschandl et al., 2013, PLoS ONE"
#> 
#> $content$records[[8]]$source$citation_id
#> [1] "23861977"
#> 
#> $content$records[[8]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[8]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[8]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/23861977"
#> 
#> $content$records[[8]]$source$open_access
#> [1] TRUE
#> 
#> $content$records[[8]]$source$pmc_id
#> [1] "PMC3704624"
#> 
#> $content$records[[8]]$source$publication_date
#> $content$records[[8]]$source$publication_date$year
#> [1] 2013
#> 
#> 
#> $content$records[[8]]$source$journal
#> [1] "PLoS ONE"
#> 
#> $content$records[[8]]$source$full_journal_title
#> [1] "PloS one"
#> 
#> $content$records[[8]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[8]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[8]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[8]]$variant_id
#> [1] 94
#> 
#> $content$records[[8]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[9]]
#> $content$records[[9]]$id
#> [1] 11
#> 
#> $content$records[[9]]$name
#> [1] "EID11"
#> 
#> $content$records[[9]]$description
#> [1] "Daunorubicin treatment resulted in similar overall survival and disease free survival in de novo AML patients with DNMT3A R882 mutation compared to those who do not harbor this mutation."
#> 
#> $content$records[[9]]$disease
#> $content$records[[9]]$disease$id
#> [1] 3
#> 
#> $content$records[[9]]$disease$name
#> [1] "Acute Myeloid Leukemia"
#> 
#> $content$records[[9]]$disease$display_name
#> [1] "Acute Myeloid Leukemia"
#> 
#> $content$records[[9]]$disease$doid
#> [1] "9119"
#> 
#> $content$records[[9]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:9119"
#> 
#> 
#> $content$records[[9]]$drugs
#> $content$records[[9]]$drugs[[1]]
#> $content$records[[9]]$drugs[[1]]$id
#> [1] 2
#> 
#> $content$records[[9]]$drugs[[1]]$name
#> [1] "Daunorubicin"
#> 
#> $content$records[[9]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[9]]$rating
#> [1] 4
#> 
#> $content$records[[9]]$evidence_level
#> [1] "B"
#> 
#> $content$records[[9]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[9]]$clinical_significance
#> [1] "Resistance"
#> 
#> $content$records[[9]]$evidence_direction
#> [1] "Does Not Support"
#> 
#> $content$records[[9]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[9]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[9]]$status
#> [1] "accepted"
#> 
#> $content$records[[9]]$open_change_count
#> [1] 0
#> 
#> $content$records[[9]]$type
#> [1] "evidence"
#> 
#> $content$records[[9]]$source
#> $content$records[[9]]$source$id
#> [1] 53
#> 
#> $content$records[[9]]$source$name
#> [1] "Do AML patients with DNMT3A exon 23 mutations benefit from idarubicin as compared to daunorubicin? A single center experience."
#> 
#> $content$records[[9]]$source$citation
#> [1] "LaRochelle et al., 2011, Oncotarget"
#> 
#> $content$records[[9]]$source$citation_id
#> [1] "22081665"
#> 
#> $content$records[[9]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[9]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[9]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/22081665"
#> 
#> $content$records[[9]]$source$open_access
#> [1] TRUE
#> 
#> $content$records[[9]]$source$pmc_id
#> [1] "PMC3260002"
#> 
#> $content$records[[9]]$source$publication_date
#> $content$records[[9]]$source$publication_date$year
#> [1] 2011
#> 
#> $content$records[[9]]$source$publication_date$month
#> [1] 11
#> 
#> 
#> $content$records[[9]]$source$journal
#> [1] "Oncotarget"
#> 
#> $content$records[[9]]$source$full_journal_title
#> [1] "Oncotarget"
#> 
#> $content$records[[9]]$source$status
#> [1] "partially curated"
#> 
#> $content$records[[9]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[9]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[9]]$variant_id
#> [1] 32
#> 
#> $content$records[[9]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[10]]
#> $content$records[[10]]$id
#> [1] 12
#> 
#> $content$records[[10]]$name
#> [1] "EID12"
#> 
#> $content$records[[10]]$description
#> [1] "A375 cells expressing MAP2K1 P124S mutation conferred an approximately 5-fold increased resistance to selumetinib (AZD6244) induced growth inhibition when compared to untransduced cells, cells expressing wild-type MAP2K1 or cells expressing a constitutively active form. This mutation also attenuated pERK inhibition by selumetinib in these cells."
#> 
#> $content$records[[10]]$disease
#> $content$records[[10]]$disease$id
#> [1] 7
#> 
#> $content$records[[10]]$disease$name
#> [1] "Melanoma"
#> 
#> $content$records[[10]]$disease$display_name
#> [1] "Melanoma"
#> 
#> $content$records[[10]]$disease$doid
#> [1] "1909"
#> 
#> $content$records[[10]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:1909"
#> 
#> 
#> $content$records[[10]]$drugs
#> $content$records[[10]]$drugs[[1]]
#> $content$records[[10]]$drugs[[1]]$id
#> [1] 122
#> 
#> $content$records[[10]]$drugs[[1]]$name
#> [1] "Selumetinib"
#> 
#> $content$records[[10]]$drugs[[1]]$pubchem_id
#> [1] ""
#> 
#> 
#> 
#> $content$records[[10]]$rating
#> [1] 3
#> 
#> $content$records[[10]]$evidence_level
#> [1] "D"
#> 
#> $content$records[[10]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[10]]$clinical_significance
#> [1] "Resistance"
#> 
#> $content$records[[10]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[10]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[10]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[10]]$status
#> [1] "accepted"
#> 
#> $content$records[[10]]$open_change_count
#> [1] 0
#> 
#> $content$records[[10]]$type
#> [1] "evidence"
#> 
#> $content$records[[10]]$source
#> $content$records[[10]]$source$id
#> [1] 58
#> 
#> $content$records[[10]]$source$name
#> [1] "MEK1 mutations confer resistance to MEK and B-RAF inhibition."
#> 
#> $content$records[[10]]$source$citation
#> [1] "Emery et al., 2009, Proc. Natl. Acad. Sci. U.S.A."
#> 
#> $content$records[[10]]$source$citation_id
#> [1] "19915144"
#> 
#> $content$records[[10]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[10]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[10]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/19915144"
#> 
#> $content$records[[10]]$source$open_access
#> [1] TRUE
#> 
#> $content$records[[10]]$source$pmc_id
#> [1] "PMC2777185"
#> 
#> $content$records[[10]]$source$publication_date
#> $content$records[[10]]$source$publication_date$year
#> [1] 2009
#> 
#> $content$records[[10]]$source$publication_date$month
#> [1] 12
#> 
#> $content$records[[10]]$source$publication_date$day
#> [1] 1
#> 
#> 
#> $content$records[[10]]$source$journal
#> [1] "Proc. Natl. Acad. Sci. U.S.A."
#> 
#> $content$records[[10]]$source$full_journal_title
#> [1] "Proceedings of the National Academy of Sciences of the United States of America"
#> 
#> $content$records[[10]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[10]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[10]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[10]]$variant_id
#> [1] 82
#> 
#> $content$records[[10]]$phenotypes
#> list()
#> 
#> 
#> 
#> 
#> $url
#> [1] "https://civicdb.org/api/evidence_items"
#> 
#> $response
#> Response [https://civicdb.org/api/evidence_items?page=1&count=10]
#>   Date: 2019-09-03 16:44
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 12.1 kB
#> 
#> 
#> attr(,"class")
#> [1] "civic_api"

getAllEvidenceItems(page = 2, count = 10)

#> $content
#> $content$`_meta`
#> $content$`_meta`$current_page
#> [1] 2
#> 
#> $content$`_meta`$per_page
#> [1] "10"
#> 
#> $content$`_meta`$total_pages
#> [1] 642
#> 
#> $content$`_meta`$total_count
#> [1] 6411
#> 
#> $content$`_meta`$links
#> $content$`_meta`$links$`next`
#> [1] "https://civicdb.org/api/evidence_items?count=10&page=3"
#> 
#> $content$`_meta`$links$previous
#> [1] "https://civicdb.org/api/evidence_items?count=10&page=1"
#> 
#> 
#> 
#> $content$records
#> $content$records[[1]]
#> $content$records[[1]]$id
#> [1] 13
#> 
#> $content$records[[1]]$name
#> [1] "EID13"
#> 
#> $content$records[[1]]$description
#> [1] "A375 cells expressing MAP2K1 Q56P mutation conferred an approximately 100-fold increased resistance to selumetinib (AZD6244) induced growth inhibition when compared to untransduced cells, cells expressing wild-type MAP2K1, or cells expressing a constitutively active form. This mutation also attenuated pERK inhibition by selumetinib in these cells."
#> 
#> $content$records[[1]]$disease
#> $content$records[[1]]$disease$id
#> [1] 7
#> 
#> $content$records[[1]]$disease$name
#> [1] "Melanoma"
#> 
#> $content$records[[1]]$disease$display_name
#> [1] "Melanoma"
#> 
#> $content$records[[1]]$disease$doid
#> [1] "1909"
#> 
#> $content$records[[1]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:1909"
#> 
#> 
#> $content$records[[1]]$drugs
#> $content$records[[1]]$drugs[[1]]
#> $content$records[[1]]$drugs[[1]]$id
#> [1] 122
#> 
#> $content$records[[1]]$drugs[[1]]$name
#> [1] "Selumetinib"
#> 
#> $content$records[[1]]$drugs[[1]]$pubchem_id
#> [1] ""
#> 
#> 
#> 
#> $content$records[[1]]$rating
#> [1] 3
#> 
#> $content$records[[1]]$evidence_level
#> [1] "D"
#> 
#> $content$records[[1]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[1]]$clinical_significance
#> [1] "Resistance"
#> 
#> $content$records[[1]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[1]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[1]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[1]]$status
#> [1] "accepted"
#> 
#> $content$records[[1]]$open_change_count
#> [1] 0
#> 
#> $content$records[[1]]$type
#> [1] "evidence"
#> 
#> $content$records[[1]]$source
#> $content$records[[1]]$source$id
#> [1] 58
#> 
#> $content$records[[1]]$source$name
#> [1] "MEK1 mutations confer resistance to MEK and B-RAF inhibition."
#> 
#> $content$records[[1]]$source$citation
#> [1] "Emery et al., 2009, Proc. Natl. Acad. Sci. U.S.A."
#> 
#> $content$records[[1]]$source$citation_id
#> [1] "19915144"
#> 
#> $content$records[[1]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[1]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[1]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/19915144"
#> 
#> $content$records[[1]]$source$open_access
#> [1] TRUE
#> 
#> $content$records[[1]]$source$pmc_id
#> [1] "PMC2777185"
#> 
#> $content$records[[1]]$source$publication_date
#> $content$records[[1]]$source$publication_date$year
#> [1] 2009
#> 
#> $content$records[[1]]$source$publication_date$month
#> [1] 12
#> 
#> $content$records[[1]]$source$publication_date$day
#> [1] 1
#> 
#> 
#> $content$records[[1]]$source$journal
#> [1] "Proc. Natl. Acad. Sci. U.S.A."
#> 
#> $content$records[[1]]$source$full_journal_title
#> [1] "Proceedings of the National Academy of Sciences of the United States of America"
#> 
#> $content$records[[1]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[1]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[1]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[1]]$variant_id
#> [1] 83
#> 
#> $content$records[[1]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[2]]
#> $content$records[[2]]$id
#> [1] 14
#> 
#> $content$records[[2]]$name
#> [1] "EID14"
#> 
#> $content$records[[2]]$description
#> [1] "Vemurafenib resistance is associated with gain of Q61 mutation in NRAS."
#> 
#> $content$records[[2]]$disease
#> $content$records[[2]]$disease$id
#> [1] 7
#> 
#> $content$records[[2]]$disease$name
#> [1] "Melanoma"
#> 
#> $content$records[[2]]$disease$display_name
#> [1] "Melanoma"
#> 
#> $content$records[[2]]$disease$doid
#> [1] "1909"
#> 
#> $content$records[[2]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:1909"
#> 
#> 
#> $content$records[[2]]$drugs
#> $content$records[[2]]$drugs[[1]]
#> $content$records[[2]]$drugs[[1]]$id
#> [1] 4
#> 
#> $content$records[[2]]$drugs[[1]]$name
#> [1] "Vemurafenib"
#> 
#> $content$records[[2]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[2]]$rating
#> [1] 4
#> 
#> $content$records[[2]]$evidence_level
#> [1] "B"
#> 
#> $content$records[[2]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[2]]$clinical_significance
#> [1] "Resistance"
#> 
#> $content$records[[2]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[2]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[2]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[2]]$status
#> [1] "accepted"
#> 
#> $content$records[[2]]$open_change_count
#> [1] 0
#> 
#> $content$records[[2]]$type
#> [1] "evidence"
#> 
#> $content$records[[2]]$source
#> $content$records[[2]]$source$id
#> [1] 60
#> 
#> $content$records[[2]]$source$name
#> [1] "Pharmacodynamic effects and mechanisms of resistance to vemurafenib in patients with metastatic melanoma."
#> 
#> $content$records[[2]]$source$citation
#> [1] "Trunzer et al., 2013, J. Clin. Oncol."
#> 
#> $content$records[[2]]$source$citation_id
#> [1] "23569304"
#> 
#> $content$records[[2]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[2]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[2]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/23569304"
#> 
#> $content$records[[2]]$source$open_access
#> NULL
#> 
#> $content$records[[2]]$source$pmc_id
#> NULL
#> 
#> $content$records[[2]]$source$publication_date
#> $content$records[[2]]$source$publication_date$year
#> [1] 2013
#> 
#> $content$records[[2]]$source$publication_date$month
#> [1] 5
#> 
#> $content$records[[2]]$source$publication_date$day
#> [1] 10
#> 
#> 
#> $content$records[[2]]$source$journal
#> [1] "J. Clin. Oncol."
#> 
#> $content$records[[2]]$source$full_journal_title
#> [1] "Journal of clinical oncology : official journal of the American Society of Clinical Oncology"
#> 
#> $content$records[[2]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[2]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[2]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[2]]$variant_id
#> [1] 94
#> 
#> $content$records[[2]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[3]]
#> $content$records[[3]]$id
#> [1] 15
#> 
#> $content$records[[3]]$name
#> [1] "EID15"
#> 
#> $content$records[[3]]$description
#> [1] "GIST cancer with D842V mutation is resistant to imatinib."
#> 
#> $content$records[[3]]$disease
#> $content$records[[3]]$disease$id
#> [1] 2
#> 
#> $content$records[[3]]$disease$name
#> [1] "Gastrointestinal Stromal Tumor"
#> 
#> $content$records[[3]]$disease$display_name
#> [1] "Gastrointestinal Stromal Tumor"
#> 
#> $content$records[[3]]$disease$doid
#> [1] "9253"
#> 
#> $content$records[[3]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:9253"
#> 
#> 
#> $content$records[[3]]$drugs
#> $content$records[[3]]$drugs[[1]]
#> $content$records[[3]]$drugs[[1]]$id
#> [1] 5
#> 
#> $content$records[[3]]$drugs[[1]]$name
#> [1] "Imatinib"
#> 
#> $content$records[[3]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[3]]$rating
#> [1] 4
#> 
#> $content$records[[3]]$evidence_level
#> [1] "B"
#> 
#> $content$records[[3]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[3]]$clinical_significance
#> [1] "Resistance"
#> 
#> $content$records[[3]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[3]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[3]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[3]]$status
#> [1] "accepted"
#> 
#> $content$records[[3]]$open_change_count
#> [1] 0
#> 
#> $content$records[[3]]$type
#> [1] "evidence"
#> 
#> $content$records[[3]]$source
#> $content$records[[3]]$source$id
#> [1] 61
#> 
#> $content$records[[3]]$source$name
#> [1] "PDGFRA mutations in gastrointestinal stromal tumors: frequency, spectrum and in vitro sensitivity to imatinib."
#> 
#> $content$records[[3]]$source$citation
#> [1] "Corless et al., 2005, J. Clin. Oncol."
#> 
#> $content$records[[3]]$source$citation_id
#> [1] "15928335"
#> 
#> $content$records[[3]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[3]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[3]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/15928335"
#> 
#> $content$records[[3]]$source$open_access
#> NULL
#> 
#> $content$records[[3]]$source$pmc_id
#> NULL
#> 
#> $content$records[[3]]$source$publication_date
#> $content$records[[3]]$source$publication_date$year
#> [1] 2005
#> 
#> $content$records[[3]]$source$publication_date$month
#> [1] 8
#> 
#> $content$records[[3]]$source$publication_date$day
#> [1] 10
#> 
#> 
#> $content$records[[3]]$source$journal
#> [1] "J. Clin. Oncol."
#> 
#> $content$records[[3]]$source$full_journal_title
#> [1] "Journal of clinical oncology : official journal of the American Society of Clinical Oncology"
#> 
#> $content$records[[3]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[3]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[3]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[3]]$variant_id
#> [1] 99
#> 
#> $content$records[[3]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[4]]
#> $content$records[[4]]$id
#> [1] 16
#> 
#> $content$records[[4]]$name
#> [1] "EID16"
#> 
#> $content$records[[4]]$description
#> [1] "While cancer with PDGFRA V561D mutation is known to be sensitive to Imatinib, double mutation of V561D and D842V mutants are resistant to imatinib."
#> 
#> $content$records[[4]]$disease
#> $content$records[[4]]$disease$id
#> [1] 2
#> 
#> $content$records[[4]]$disease$name
#> [1] "Gastrointestinal Stromal Tumor"
#> 
#> $content$records[[4]]$disease$display_name
#> [1] "Gastrointestinal Stromal Tumor"
#> 
#> $content$records[[4]]$disease$doid
#> [1] "9253"
#> 
#> $content$records[[4]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:9253"
#> 
#> 
#> $content$records[[4]]$drugs
#> $content$records[[4]]$drugs[[1]]
#> $content$records[[4]]$drugs[[1]]$id
#> [1] 5
#> 
#> $content$records[[4]]$drugs[[1]]$name
#> [1] "Imatinib"
#> 
#> $content$records[[4]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[4]]$rating
#> [1] 4
#> 
#> $content$records[[4]]$evidence_level
#> [1] "B"
#> 
#> $content$records[[4]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[4]]$clinical_significance
#> [1] "Resistance"
#> 
#> $content$records[[4]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[4]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[4]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[4]]$status
#> [1] "accepted"
#> 
#> $content$records[[4]]$open_change_count
#> [1] 0
#> 
#> $content$records[[4]]$type
#> [1] "evidence"
#> 
#> $content$records[[4]]$source
#> $content$records[[4]]$source$id
#> [1] 62
#> 
#> $content$records[[4]]$source$name
#> [1] "Molecular correlates of imatinib resistance in gastrointestinal stromal tumors."
#> 
#> $content$records[[4]]$source$citation
#> [1] "Heinrich et al., 2006, J. Clin. Oncol."
#> 
#> $content$records[[4]]$source$citation_id
#> [1] "16954519"
#> 
#> $content$records[[4]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[4]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[4]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/16954519"
#> 
#> $content$records[[4]]$source$open_access
#> NULL
#> 
#> $content$records[[4]]$source$pmc_id
#> NULL
#> 
#> $content$records[[4]]$source$publication_date
#> $content$records[[4]]$source$publication_date$year
#> [1] 2006
#> 
#> $content$records[[4]]$source$publication_date$month
#> [1] 10
#> 
#> $content$records[[4]]$source$publication_date$day
#> [1] 10
#> 
#> 
#> $content$records[[4]]$source$journal
#> [1] "J. Clin. Oncol."
#> 
#> $content$records[[4]]$source$full_journal_title
#> [1] "Journal of clinical oncology : official journal of the American Society of Clinical Oncology"
#> 
#> $content$records[[4]]$source$status
#> [1] "partially curated"
#> 
#> $content$records[[4]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[4]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[4]]$variant_id
#> [1] 99
#> 
#> $content$records[[4]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[5]]
#> $content$records[[5]]$id
#> [1] 17
#> 
#> $content$records[[5]]$name
#> [1] "EID17"
#> 
#> $content$records[[5]]$description
#> [1] "In one patient with S214C mutation, the use of sorafenib has led to more than 5 years of survival and near remission."
#> 
#> $content$records[[5]]$disease
#> $content$records[[5]]$disease$id
#> [1] 8
#> 
#> $content$records[[5]]$disease$name
#> [1] "Lung Non-small Cell Carcinoma"
#> 
#> $content$records[[5]]$disease$display_name
#> [1] "Lung Non-small Cell Carcinoma"
#> 
#> $content$records[[5]]$disease$doid
#> [1] "3908"
#> 
#> $content$records[[5]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:3908"
#> 
#> 
#> $content$records[[5]]$drugs
#> $content$records[[5]]$drugs[[1]]
#> $content$records[[5]]$drugs[[1]]$id
#> [1] 6
#> 
#> $content$records[[5]]$drugs[[1]]$name
#> [1] "Sorafenib"
#> 
#> $content$records[[5]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[5]]$rating
#> [1] 2
#> 
#> $content$records[[5]]$evidence_level
#> [1] "C"
#> 
#> $content$records[[5]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[5]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[5]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[5]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[5]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[5]]$status
#> [1] "accepted"
#> 
#> $content$records[[5]]$open_change_count
#> [1] 0
#> 
#> $content$records[[5]]$type
#> [1] "evidence"
#> 
#> $content$records[[5]]$source
#> $content$records[[5]]$source$id
#> [1] 6
#> 
#> $content$records[[5]]$source$name
#> [1] "Oncogenic and sorafenib-sensitive ARAF mutations in lung adenocarcinoma."
#> 
#> $content$records[[5]]$source$citation
#> [1] "Imielinski et al., 2014, J. Clin. Invest."
#> 
#> $content$records[[5]]$source$citation_id
#> [1] "24569458"
#> 
#> $content$records[[5]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[5]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[5]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/24569458"
#> 
#> $content$records[[5]]$source$open_access
#> [1] TRUE
#> 
#> $content$records[[5]]$source$pmc_id
#> [1] "PMC3973082"
#> 
#> $content$records[[5]]$source$publication_date
#> $content$records[[5]]$source$publication_date$year
#> [1] 2014
#> 
#> $content$records[[5]]$source$publication_date$month
#> [1] 4
#> 
#> 
#> $content$records[[5]]$source$journal
#> [1] "J. Clin. Invest."
#> 
#> $content$records[[5]]$source$full_journal_title
#> [1] "The Journal of clinical investigation"
#> 
#> $content$records[[5]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[5]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[5]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[5]]$variant_id
#> [1] 10
#> 
#> $content$records[[5]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[6]]
#> $content$records[[6]]$id
#> [1] 18
#> 
#> $content$records[[6]]$name
#> [1] "EID18"
#> 
#> $content$records[[6]]$description
#> [1] "Idarubicin increases the overall survival and disease free survival in de novo AML patients with DNMT3A R882 mutation compared to those who do not harbor this mutation."
#> 
#> $content$records[[6]]$disease
#> $content$records[[6]]$disease$id
#> [1] 3
#> 
#> $content$records[[6]]$disease$name
#> [1] "Acute Myeloid Leukemia"
#> 
#> $content$records[[6]]$disease$display_name
#> [1] "Acute Myeloid Leukemia"
#> 
#> $content$records[[6]]$disease$doid
#> [1] "9119"
#> 
#> $content$records[[6]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:9119"
#> 
#> 
#> $content$records[[6]]$drugs
#> $content$records[[6]]$drugs[[1]]
#> $content$records[[6]]$drugs[[1]]$id
#> [1] 7
#> 
#> $content$records[[6]]$drugs[[1]]$name
#> [1] "Idarubicin"
#> 
#> $content$records[[6]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[6]]$rating
#> [1] 4
#> 
#> $content$records[[6]]$evidence_level
#> [1] "B"
#> 
#> $content$records[[6]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[6]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[6]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[6]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[6]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[6]]$status
#> [1] "accepted"
#> 
#> $content$records[[6]]$open_change_count
#> [1] 0
#> 
#> $content$records[[6]]$type
#> [1] "evidence"
#> 
#> $content$records[[6]]$source
#> $content$records[[6]]$source$id
#> [1] 53
#> 
#> $content$records[[6]]$source$name
#> [1] "Do AML patients with DNMT3A exon 23 mutations benefit from idarubicin as compared to daunorubicin? A single center experience."
#> 
#> $content$records[[6]]$source$citation
#> [1] "LaRochelle et al., 2011, Oncotarget"
#> 
#> $content$records[[6]]$source$citation_id
#> [1] "22081665"
#> 
#> $content$records[[6]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[6]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[6]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/22081665"
#> 
#> $content$records[[6]]$source$open_access
#> [1] TRUE
#> 
#> $content$records[[6]]$source$pmc_id
#> [1] "PMC3260002"
#> 
#> $content$records[[6]]$source$publication_date
#> $content$records[[6]]$source$publication_date$year
#> [1] 2011
#> 
#> $content$records[[6]]$source$publication_date$month
#> [1] 11
#> 
#> 
#> $content$records[[6]]$source$journal
#> [1] "Oncotarget"
#> 
#> $content$records[[6]]$source$full_journal_title
#> [1] "Oncotarget"
#> 
#> $content$records[[6]]$source$status
#> [1] "partially curated"
#> 
#> $content$records[[6]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[6]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[6]]$variant_id
#> [1] 32
#> 
#> $content$records[[6]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[7]]
#> $content$records[[7]]$id
#> [1] 19
#> 
#> $content$records[[7]]$name
#> [1] "EID19"
#> 
#> $content$records[[7]]$description
#> [1] "In patients with JAK2 V617F, the use of pegylated IFN-alpha-2a leads to reduction in the percentage of cells harboring JAK2 V617F."
#> 
#> $content$records[[7]]$disease
#> $content$records[[7]]$disease$id
#> [1] 9
#> 
#> $content$records[[7]]$disease$name
#> [1] "Polycythemia Vera"
#> 
#> $content$records[[7]]$disease$display_name
#> [1] "Polycythemia Vera"
#> 
#> $content$records[[7]]$disease$doid
#> [1] "8997"
#> 
#> $content$records[[7]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:8997"
#> 
#> 
#> $content$records[[7]]$drugs
#> $content$records[[7]]$drugs[[1]]
#> $content$records[[7]]$drugs[[1]]$id
#> [1] 8
#> 
#> $content$records[[7]]$drugs[[1]]$name
#> [1] "Pegylated IFN-alpha-2a"
#> 
#> $content$records[[7]]$drugs[[1]]$pubchem_id
#> [1] ""
#> 
#> 
#> 
#> $content$records[[7]]$rating
#> [1] 4
#> 
#> $content$records[[7]]$evidence_level
#> [1] "B"
#> 
#> $content$records[[7]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[7]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[7]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[7]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[7]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[7]]$status
#> [1] "accepted"
#> 
#> $content$records[[7]]$open_change_count
#> [1] 1
#> 
#> $content$records[[7]]$type
#> [1] "evidence"
#> 
#> $content$records[[7]]$source
#> $content$records[[7]]$source$id
#> [1] 64
#> 
#> $content$records[[7]]$source$name
#> [1] "High molecular response rate of polycythemia vera patients treated with pegylated interferon alpha-2a."
#> 
#> $content$records[[7]]$source$citation
#> [1] "Kiladjian et al., 2006, Blood"
#> 
#> $content$records[[7]]$source$citation_id
#> [1] "16709929"
#> 
#> $content$records[[7]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[7]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[7]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/16709929"
#> 
#> $content$records[[7]]$source$open_access
#> NULL
#> 
#> $content$records[[7]]$source$pmc_id
#> NULL
#> 
#> $content$records[[7]]$source$publication_date
#> $content$records[[7]]$source$publication_date$year
#> [1] 2006
#> 
#> $content$records[[7]]$source$publication_date$month
#> [1] 9
#> 
#> $content$records[[7]]$source$publication_date$day
#> [1] 15
#> 
#> 
#> $content$records[[7]]$source$journal
#> [1] "Blood"
#> 
#> $content$records[[7]]$source$full_journal_title
#> [1] "Blood"
#> 
#> $content$records[[7]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[7]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[7]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[7]]$variant_id
#> [1] 64
#> 
#> $content$records[[7]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[8]]
#> $content$records[[8]]$id
#> [1] 20
#> 
#> $content$records[[8]]$name
#> [1] "EID20"
#> 
#> $content$records[[8]]$description
#> [1] "TG101348 effectively inhibits STAT5 signaling in JAK2 V617F cells."
#> 
#> $content$records[[8]]$disease
#> $content$records[[8]]$disease$id
#> [1] 9
#> 
#> $content$records[[8]]$disease$name
#> [1] "Polycythemia Vera"
#> 
#> $content$records[[8]]$disease$display_name
#> [1] "Polycythemia Vera"
#> 
#> $content$records[[8]]$disease$doid
#> [1] "8997"
#> 
#> $content$records[[8]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:8997"
#> 
#> 
#> $content$records[[8]]$drugs
#> $content$records[[8]]$drugs[[1]]
#> $content$records[[8]]$drugs[[1]]$id
#> [1] 9
#> 
#> $content$records[[8]]$drugs[[1]]$name
#> [1] "TG101348"
#> 
#> $content$records[[8]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[8]]$rating
#> [1] 3
#> 
#> $content$records[[8]]$evidence_level
#> [1] "D"
#> 
#> $content$records[[8]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[8]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[8]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[8]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[8]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[8]]$status
#> [1] "accepted"
#> 
#> $content$records[[8]]$open_change_count
#> [1] 0
#> 
#> $content$records[[8]]$type
#> [1] "evidence"
#> 
#> $content$records[[8]]$source
#> $content$records[[8]]$source$id
#> [1] 65
#> 
#> $content$records[[8]]$source$name
#> [1] "Efficacy of TG101348, a selective JAK2 inhibitor, in treatment of a murine model of JAK2V617F-induced polycythemia vera."
#> 
#> $content$records[[8]]$source$citation
#> [1] "Wernig et al., 2008, Cancer Cell"
#> 
#> $content$records[[8]]$source$citation_id
#> [1] "18394554"
#> 
#> $content$records[[8]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[8]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[8]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/18394554"
#> 
#> $content$records[[8]]$source$open_access
#> NULL
#> 
#> $content$records[[8]]$source$pmc_id
#> NULL
#> 
#> $content$records[[8]]$source$publication_date
#> $content$records[[8]]$source$publication_date$year
#> [1] 2008
#> 
#> $content$records[[8]]$source$publication_date$month
#> [1] 4
#> 
#> 
#> $content$records[[8]]$source$journal
#> [1] "Cancer Cell"
#> 
#> $content$records[[8]]$source$full_journal_title
#> [1] "Cancer cell"
#> 
#> $content$records[[8]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[8]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[8]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[8]]$variant_id
#> [1] 64
#> 
#> $content$records[[8]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[9]]
#> $content$records[[9]]$id
#> [1] 21
#> 
#> $content$records[[9]]$name
#> [1] "EID21"
#> 
#> $content$records[[9]]$description
#> [1] "Likely due to increased reliance of mutant NRAS on HSP90 for stabilization, inhibition of HSP90 by 17-AAG was shown to be effective in a patient with metastatic malignant melanoma with an NRAS G13D mutation."
#> 
#> $content$records[[9]]$disease
#> $content$records[[9]]$disease$id
#> [1] 7
#> 
#> $content$records[[9]]$disease$name
#> [1] "Melanoma"
#> 
#> $content$records[[9]]$disease$display_name
#> [1] "Melanoma"
#> 
#> $content$records[[9]]$disease$doid
#> [1] "1909"
#> 
#> $content$records[[9]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:1909"
#> 
#> 
#> $content$records[[9]]$drugs
#> $content$records[[9]]$drugs[[1]]
#> $content$records[[9]]$drugs[[1]]$id
#> [1] 10
#> 
#> $content$records[[9]]$drugs[[1]]$name
#> [1] "17-AAG"
#> 
#> $content$records[[9]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$records[[9]]$rating
#> [1] 2
#> 
#> $content$records[[9]]$evidence_level
#> [1] "C"
#> 
#> $content$records[[9]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[9]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[9]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[9]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[9]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[9]]$status
#> [1] "accepted"
#> 
#> $content$records[[9]]$open_change_count
#> [1] 0
#> 
#> $content$records[[9]]$type
#> [1] "evidence"
#> 
#> $content$records[[9]]$source
#> $content$records[[9]]$source$id
#> [1] 66
#> 
#> $content$records[[9]]$source$name
#> [1] "BRAF and NRAS mutations in melanoma: potential relationships to clinical response to HSP90 inhibitors."
#> 
#> $content$records[[9]]$source$citation
#> [1] "Banerji et al., 2008, Mol. Cancer Ther."
#> 
#> $content$records[[9]]$source$citation_id
#> [1] "18375819"
#> 
#> $content$records[[9]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[9]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[9]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/18375819"
#> 
#> $content$records[[9]]$source$open_access
#> NULL
#> 
#> $content$records[[9]]$source$pmc_id
#> NULL
#> 
#> $content$records[[9]]$source$publication_date
#> $content$records[[9]]$source$publication_date$year
#> [1] 2008
#> 
#> $content$records[[9]]$source$publication_date$month
#> [1] 4
#> 
#> 
#> $content$records[[9]]$source$journal
#> [1] "Mol. Cancer Ther."
#> 
#> $content$records[[9]]$source$full_journal_title
#> [1] "Molecular cancer therapeutics"
#> 
#> $content$records[[9]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[9]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[9]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[9]]$variant_id
#> [1] 93
#> 
#> $content$records[[9]]$phenotypes
#> list()
#> 
#> 
#> $content$records[[10]]
#> $content$records[[10]]$id
#> [1] 22
#> 
#> $content$records[[10]]$name
#> [1] "EID22"
#> 
#> $content$records[[10]]$description
#> [1] "In a melanoma patient with NRAS Q61L mutation, treatment with temozolomide resulted in disease free survival of 14 months."
#> 
#> $content$records[[10]]$disease
#> $content$records[[10]]$disease$id
#> [1] 7
#> 
#> $content$records[[10]]$disease$name
#> [1] "Melanoma"
#> 
#> $content$records[[10]]$disease$display_name
#> [1] "Melanoma"
#> 
#> $content$records[[10]]$disease$doid
#> [1] "1909"
#> 
#> $content$records[[10]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:1909"
#> 
#> 
#> $content$records[[10]]$drugs
#> $content$records[[10]]$drugs[[1]]
#> $content$records[[10]]$drugs[[1]]$id
#> [1] 11
#> 
#> $content$records[[10]]$drugs[[1]]$name
#> [1] "Temozolomide"
#> 
#> $content$records[[10]]$drugs[[1]]$pubchem_id
#> [1] ""
#> 
#> 
#> 
#> $content$records[[10]]$rating
#> [1] 2
#> 
#> $content$records[[10]]$evidence_level
#> [1] "C"
#> 
#> $content$records[[10]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$records[[10]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$records[[10]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$records[[10]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$records[[10]]$drug_interaction_type
#> NULL
#> 
#> $content$records[[10]]$status
#> [1] "accepted"
#> 
#> $content$records[[10]]$open_change_count
#> [1] 0
#> 
#> $content$records[[10]]$type
#> [1] "evidence"
#> 
#> $content$records[[10]]$source
#> $content$records[[10]]$source$id
#> [1] 67
#> 
#> $content$records[[10]]$source$name
#> [1] "NRAS-mutant melanoma: response to chemotherapy."
#> 
#> $content$records[[10]]$source$citation
#> [1] "Soon et al., 2011, Arch Dermatol"
#> 
#> $content$records[[10]]$source$citation_id
#> [1] "21576590"
#> 
#> $content$records[[10]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$records[[10]]$source$asco_abstract_id
#> NULL
#> 
#> $content$records[[10]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/21576590"
#> 
#> $content$records[[10]]$source$open_access
#> NULL
#> 
#> $content$records[[10]]$source$pmc_id
#> NULL
#> 
#> $content$records[[10]]$source$publication_date
#> $content$records[[10]]$source$publication_date$year
#> [1] 2011
#> 
#> $content$records[[10]]$source$publication_date$month
#> [1] 5
#> 
#> 
#> $content$records[[10]]$source$journal
#> [1] "Arch Dermatol"
#> 
#> $content$records[[10]]$source$full_journal_title
#> [1] "Archives of dermatology"
#> 
#> $content$records[[10]]$source$status
#> [1] "fully curated"
#> 
#> $content$records[[10]]$source$is_review
#> [1] FALSE
#> 
#> $content$records[[10]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$records[[10]]$variant_id
#> [1] 95
#> 
#> $content$records[[10]]$phenotypes
#> list()
#> 
#> 
#> 
#> 
#> $url
#> [1] "https://civicdb.org/api/evidence_items"
#> 
#> $response
#> Response [https://civicdb.org/api/evidence_items?page=2&count=10]
#>   Date: 2019-09-03 16:44
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 12.6 kB
#> 
#> 
#> attr(,"class")
#> [1] "civic_api"