Get a list of variants

Retrieve all variants from the CIViC DB

getAllVariants(page = 1, count = 25)

Arguments

page
the page number to retrieve

count
the number of variants to retrieve

Value

An S3 Object of type civic_api containing the content, url, and response

Examples

getAllVariants(count = 10)

#> $content
#> $content$`_meta`
#> $content$`_meta`$current_page
#> [1] 1
#> 
#> $content$`_meta`$per_page
#> [1] "10"
#> 
#> $content$`_meta`$total_pages
#> [1] 229
#> 
#> $content$`_meta`$total_count
#> [1] 2286
#> 
#> $content$`_meta`$links
#> $content$`_meta`$links$`next`
#> [1] "https://civicdb.org/api/variants?count=10&page=2"
#> 
#> $content$`_meta`$links$previous
#> NULL
#> 
#> 
#> 
#> $content$records
#> $content$records[[1]]
#> $content$records[[1]]$id
#> [1] 1
#> 
#> $content$records[[1]]$entrez_name
#> [1] "ABL1"
#> 
#> $content$records[[1]]$entrez_id
#> [1] 25
#> 
#> $content$records[[1]]$name
#> [1] "BCR-ABL"
#> 
#> $content$records[[1]]$description
#> [1] "The BCR-ABL fusion protein, commonly referred to as the Philadelphia chromosome, is one of the most studied fusion genes in cancer. It has widely been considered the initiating event in chronic myelogenous leukemia (CML). However, despite its ability to initiate disease in mice, its status as an initiating mutation is in dispute. In what is commonly used as the poster-child for targeted therapeutics, the development and use of imatinib in the clinic has led to profound improvements in the prognosis of the disease. However, imatinib resistance is still seen in patients with mutations in the ABL kinase domain of the fusion, most notably the T315I variant. In patients resistant to imatinib, either acquired or otherwise, second generation BCR-ABL TKI's (dasatinib and nilotinib) have seen some success in delivering a tumor response. Third generation ABL1 inhibitor ponatinib is the only FDA approved drug with activity against T315I . However due to risk of life-threatening blood clots and severe narrowing of blood vessels ponatinib is ONLY approved for T315I-positive CML or T315I-positive Ph+ ALL or in cases of CML, Ph+ ALL with resistance or intolerance to other approved ABL1 inhibitors."
#> 
#> $content$records[[1]]$gene_id
#> [1] 4
#> 
#> $content$records[[1]]$type
#> [1] "variant"
#> 
#> $content$records[[1]]$variant_types
#> $content$records[[1]]$variant_types[[1]]
#> $content$records[[1]]$variant_types[[1]]$id
#> [1] 120
#> 
#> $content$records[[1]]$variant_types[[1]]$name
#> [1] "transcript_fusion"
#> 
#> $content$records[[1]]$variant_types[[1]]$display_name
#> [1] "Transcript Fusion"
#> 
#> $content$records[[1]]$variant_types[[1]]$so_id
#> [1] "SO:0001886"
#> 
#> $content$records[[1]]$variant_types[[1]]$description
#> [1] "A feature fusion where the deletion brings together transcript regions."
#> 
#> $content$records[[1]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001886"
#> 
#> 
#> 
#> $content$records[[1]]$civic_actionability_score
#> [1] 238
#> 
#> $content$records[[1]]$coordinates
#> $content$records[[1]]$coordinates$chromosome
#> [1] "22"
#> 
#> $content$records[[1]]$coordinates$start
#> [1] 23522397
#> 
#> $content$records[[1]]$coordinates$stop
#> [1] 23632600
#> 
#> $content$records[[1]]$coordinates$reference_bases
#> NULL
#> 
#> $content$records[[1]]$coordinates$variant_bases
#> NULL
#> 
#> $content$records[[1]]$coordinates$representative_transcript
#> [1] "ENST00000305877.8"
#> 
#> $content$records[[1]]$coordinates$chromosome2
#> [1] "9"
#> 
#> $content$records[[1]]$coordinates$start2
#> [1] 133729451
#> 
#> $content$records[[1]]$coordinates$stop2
#> [1] 133763063
#> 
#> $content$records[[1]]$coordinates$representative_transcript2
#> [1] "ENST00000318560.5"
#> 
#> $content$records[[1]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[1]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[2]]
#> $content$records[[2]]$id
#> [1] 2
#> 
#> $content$records[[2]]$entrez_name
#> [1] "ABL1"
#> 
#> $content$records[[2]]$entrez_id
#> [1] 25
#> 
#> $content$records[[2]]$name
#> [1] "BCR-ABL T315I"
#> 
#> $content$records[[2]]$description
#> [1] "While the efficacy of imatinib has revolutionized chronic myelogenous leukemia (CML) treatment, it is still not a cure-all. Both initial resistance and acquired resistance as a result of selection have been seen in a small subset of CML patients. The ABL kinase domain mutation T315I (aka T334I) has been shown to be one such mutation that confers resistance to imatinib. Second generation TKI's (dasatinib and ponatinib) specific to BCR-ABL have shown efficacy in treating resistant cases."
#> 
#> $content$records[[2]]$gene_id
#> [1] 4
#> 
#> $content$records[[2]]$type
#> [1] "variant"
#> 
#> $content$records[[2]]$variant_types
#> $content$records[[2]]$variant_types[[1]]
#> $content$records[[2]]$variant_types[[1]]$id
#> [1] 47
#> 
#> $content$records[[2]]$variant_types[[1]]$name
#> [1] "missense_variant"
#> 
#> $content$records[[2]]$variant_types[[1]]$display_name
#> [1] "Missense Variant"
#> 
#> $content$records[[2]]$variant_types[[1]]$so_id
#> [1] "SO:0001583"
#> 
#> $content$records[[2]]$variant_types[[1]]$description
#> [1] "A sequence variant, that changes one or more bases, resulting in a different amino acid sequence but where the length is preserved."
#> 
#> $content$records[[2]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001583"
#> 
#> 
#> $content$records[[2]]$variant_types[[2]]
#> $content$records[[2]]$variant_types[[2]]$id
#> [1] 120
#> 
#> $content$records[[2]]$variant_types[[2]]$name
#> [1] "transcript_fusion"
#> 
#> $content$records[[2]]$variant_types[[2]]$display_name
#> [1] "Transcript Fusion"
#> 
#> $content$records[[2]]$variant_types[[2]]$so_id
#> [1] "SO:0001886"
#> 
#> $content$records[[2]]$variant_types[[2]]$description
#> [1] "A feature fusion where the deletion brings together transcript regions."
#> 
#> $content$records[[2]]$variant_types[[2]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001886"
#> 
#> 
#> 
#> $content$records[[2]]$civic_actionability_score
#> [1] 105
#> 
#> $content$records[[2]]$coordinates
#> $content$records[[2]]$coordinates$chromosome
#> [1] "9"
#> 
#> $content$records[[2]]$coordinates$start
#> [1] 133748283
#> 
#> $content$records[[2]]$coordinates$stop
#> [1] 133748283
#> 
#> $content$records[[2]]$coordinates$reference_bases
#> [1] "C"
#> 
#> $content$records[[2]]$coordinates$variant_bases
#> [1] "T"
#> 
#> $content$records[[2]]$coordinates$representative_transcript
#> [1] "ENST00000318560.5"
#> 
#> $content$records[[2]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[2]]$coordinates$start2
#> NULL
#> 
#> $content$records[[2]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[2]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[2]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[2]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[3]]
#> $content$records[[3]]$id
#> [1] 3
#> 
#> $content$records[[3]]$entrez_name
#> [1] "ABL1"
#> 
#> $content$records[[3]]$entrez_id
#> [1] 25
#> 
#> $content$records[[3]]$name
#> [1] "BCR-ABL E255K"
#> 
#> $content$records[[3]]$description
#> [1] "While the efficacy of imatinib has revolutionized chronic myelogenous leukemia (CML) treatment, it is still not a cure-all. Both initial resistance and acquired resistance as a result of selection have been seen in a small subset of CML patients. The ABL kinase domain mutation E255K has been shown to be one such mutation that confers resistance to imatinib. Second generation TKI's (dasatinib and nilotinib) specific to BCR-ABL have shown efficacy in treating resistant cases."
#> 
#> $content$records[[3]]$gene_id
#> [1] 4
#> 
#> $content$records[[3]]$type
#> [1] "variant"
#> 
#> $content$records[[3]]$variant_types
#> $content$records[[3]]$variant_types[[1]]
#> $content$records[[3]]$variant_types[[1]]$id
#> [1] 47
#> 
#> $content$records[[3]]$variant_types[[1]]$name
#> [1] "missense_variant"
#> 
#> $content$records[[3]]$variant_types[[1]]$display_name
#> [1] "Missense Variant"
#> 
#> $content$records[[3]]$variant_types[[1]]$so_id
#> [1] "SO:0001583"
#> 
#> $content$records[[3]]$variant_types[[1]]$description
#> [1] "A sequence variant, that changes one or more bases, resulting in a different amino acid sequence but where the length is preserved."
#> 
#> $content$records[[3]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001583"
#> 
#> 
#> $content$records[[3]]$variant_types[[2]]
#> $content$records[[3]]$variant_types[[2]]$id
#> [1] 120
#> 
#> $content$records[[3]]$variant_types[[2]]$name
#> [1] "transcript_fusion"
#> 
#> $content$records[[3]]$variant_types[[2]]$display_name
#> [1] "Transcript Fusion"
#> 
#> $content$records[[3]]$variant_types[[2]]$so_id
#> [1] "SO:0001886"
#> 
#> $content$records[[3]]$variant_types[[2]]$description
#> [1] "A feature fusion where the deletion brings together transcript regions."
#> 
#> $content$records[[3]]$variant_types[[2]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001886"
#> 
#> 
#> 
#> $content$records[[3]]$civic_actionability_score
#> [1] 56
#> 
#> $content$records[[3]]$coordinates
#> $content$records[[3]]$coordinates$chromosome
#> [1] "9"
#> 
#> $content$records[[3]]$coordinates$start
#> [1] 133738363
#> 
#> $content$records[[3]]$coordinates$stop
#> [1] 133738363
#> 
#> $content$records[[3]]$coordinates$reference_bases
#> [1] "G"
#> 
#> $content$records[[3]]$coordinates$variant_bases
#> [1] "A"
#> 
#> $content$records[[3]]$coordinates$representative_transcript
#> [1] "ENST00000318560.5"
#> 
#> $content$records[[3]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[3]]$coordinates$start2
#> NULL
#> 
#> $content$records[[3]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[3]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[3]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[3]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[4]]
#> $content$records[[4]]$id
#> [1] 4
#> 
#> $content$records[[4]]$entrez_name
#> [1] "AKT1"
#> 
#> $content$records[[4]]$entrez_id
#> [1] 207
#> 
#> $content$records[[4]]$name
#> [1] "E17K"
#> 
#> $content$records[[4]]$description
#> [1] "AKT1 E17K is a recurrent mutation that has been observed in breast, colorectal, lung, and ovarian cancer. It has been convincingly shown to be an activating mutation resulting in PI3K/AKT/mTOR pathway activity. It has been suggested that this mutation decreases the cell's sensitivity to AKT1 allosteric kinase inhibitors. This, and other AKT1 mutations, are the subject of much research and development for therapeutics."
#> 
#> $content$records[[4]]$gene_id
#> [1] 2
#> 
#> $content$records[[4]]$type
#> [1] "variant"
#> 
#> $content$records[[4]]$variant_types
#> $content$records[[4]]$variant_types[[1]]
#> $content$records[[4]]$variant_types[[1]]$id
#> [1] 47
#> 
#> $content$records[[4]]$variant_types[[1]]$name
#> [1] "missense_variant"
#> 
#> $content$records[[4]]$variant_types[[1]]$display_name
#> [1] "Missense Variant"
#> 
#> $content$records[[4]]$variant_types[[1]]$so_id
#> [1] "SO:0001583"
#> 
#> $content$records[[4]]$variant_types[[1]]$description
#> [1] "A sequence variant, that changes one or more bases, resulting in a different amino acid sequence but where the length is preserved."
#> 
#> $content$records[[4]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001583"
#> 
#> 
#> 
#> $content$records[[4]]$civic_actionability_score
#> [1] 32.5
#> 
#> $content$records[[4]]$coordinates
#> $content$records[[4]]$coordinates$chromosome
#> [1] "14"
#> 
#> $content$records[[4]]$coordinates$start
#> [1] 105246551
#> 
#> $content$records[[4]]$coordinates$stop
#> [1] 105246551
#> 
#> $content$records[[4]]$coordinates$reference_bases
#> [1] "C"
#> 
#> $content$records[[4]]$coordinates$variant_bases
#> [1] "T"
#> 
#> $content$records[[4]]$coordinates$representative_transcript
#> [1] "ENST00000407796.2"
#> 
#> $content$records[[4]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[4]]$coordinates$start2
#> NULL
#> 
#> $content$records[[4]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[4]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[4]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[4]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[5]]
#> $content$records[[5]]$id
#> [1] 5
#> 
#> $content$records[[5]]$entrez_name
#> [1] "ALK"
#> 
#> $content$records[[5]]$entrez_id
#> [1] 238
#> 
#> $content$records[[5]]$name
#> [1] "EML4-ALK"
#> 
#> $content$records[[5]]$description
#> [1] "The EML4-ALK fusion variant 1 consisting of ALK kinase domain (exons 20-29) fused to EML4 exons 1-13 is the most common EML4-ALK variant, and was discovered in non-small cell lung cancer. Multiple EML4 breakpoint shave been described with differential sensitivity to inhibitors with variant 1 showing greater sensitivity than 3a in cell lines. EML4-ALK is crizotinib sensitive; however, several mutations that confer resistance mutations have been described in case studies. In the only clinical trial for crizotinib that included determination of EML4-ALK variant type in a subset of its participants, a very high response rate was observed, although the numbers were insufficient to validate correlation of variant type to outcome. Preclinical studies with this variant have indicated sensitivity to Hsp90 inhibitors."
#> 
#> $content$records[[5]]$gene_id
#> [1] 1
#> 
#> $content$records[[5]]$type
#> [1] "variant"
#> 
#> $content$records[[5]]$variant_types
#> $content$records[[5]]$variant_types[[1]]
#> $content$records[[5]]$variant_types[[1]]$id
#> [1] 120
#> 
#> $content$records[[5]]$variant_types[[1]]$name
#> [1] "transcript_fusion"
#> 
#> $content$records[[5]]$variant_types[[1]]$display_name
#> [1] "Transcript Fusion"
#> 
#> $content$records[[5]]$variant_types[[1]]$so_id
#> [1] "SO:0001886"
#> 
#> $content$records[[5]]$variant_types[[1]]$description
#> [1] "A feature fusion where the deletion brings together transcript regions."
#> 
#> $content$records[[5]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001886"
#> 
#> 
#> 
#> $content$records[[5]]$civic_actionability_score
#> [1] 48
#> 
#> $content$records[[5]]$coordinates
#> $content$records[[5]]$coordinates$chromosome
#> [1] "2"
#> 
#> $content$records[[5]]$coordinates$start
#> [1] 42396490
#> 
#> $content$records[[5]]$coordinates$stop
#> [1] 42522656
#> 
#> $content$records[[5]]$coordinates$reference_bases
#> NULL
#> 
#> $content$records[[5]]$coordinates$variant_bases
#> NULL
#> 
#> $content$records[[5]]$coordinates$representative_transcript
#> [1] "ENST00000318522.5"
#> 
#> $content$records[[5]]$coordinates$chromosome2
#> [1] "2"
#> 
#> $content$records[[5]]$coordinates$start2
#> [1] 29415640
#> 
#> $content$records[[5]]$coordinates$stop2
#> [1] 29446394
#> 
#> $content$records[[5]]$coordinates$representative_transcript2
#> [1] "ENST00000389048.3"
#> 
#> $content$records[[5]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[5]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[6]]
#> $content$records[[6]]$id
#> [1] 6
#> 
#> $content$records[[6]]$entrez_name
#> [1] "ALK"
#> 
#> $content$records[[6]]$entrez_id
#> [1] 238
#> 
#> $content$records[[6]]$name
#> [1] "EML4-ALK C1156Y"
#> 
#> $content$records[[6]]$description
#> [1] "In patients with non-small cell lung cancer exhibiting EML4-ALK fusion, C1156Y has been shown to confer resistance to crizotinib. Case reports indicate that secondary mutations can modulate drug sensitivity. EML4-ALK C1156Y/L1196M maintained crizotinib resistance while the lorlatinib resistant combination EML4-ALK C1156Y/L1198F re-sensitized the tumor to crizotinib treatment."
#> 
#> $content$records[[6]]$gene_id
#> [1] 1
#> 
#> $content$records[[6]]$type
#> [1] "variant"
#> 
#> $content$records[[6]]$variant_types
#> $content$records[[6]]$variant_types[[1]]
#> $content$records[[6]]$variant_types[[1]]$id
#> [1] 47
#> 
#> $content$records[[6]]$variant_types[[1]]$name
#> [1] "missense_variant"
#> 
#> $content$records[[6]]$variant_types[[1]]$display_name
#> [1] "Missense Variant"
#> 
#> $content$records[[6]]$variant_types[[1]]$so_id
#> [1] "SO:0001583"
#> 
#> $content$records[[6]]$variant_types[[1]]$description
#> [1] "A sequence variant, that changes one or more bases, resulting in a different amino acid sequence but where the length is preserved."
#> 
#> $content$records[[6]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001583"
#> 
#> 
#> $content$records[[6]]$variant_types[[2]]
#> $content$records[[6]]$variant_types[[2]]$id
#> [1] 120
#> 
#> $content$records[[6]]$variant_types[[2]]$name
#> [1] "transcript_fusion"
#> 
#> $content$records[[6]]$variant_types[[2]]$display_name
#> [1] "Transcript Fusion"
#> 
#> $content$records[[6]]$variant_types[[2]]$so_id
#> [1] "SO:0001886"
#> 
#> $content$records[[6]]$variant_types[[2]]$description
#> [1] "A feature fusion where the deletion brings together transcript regions."
#> 
#> $content$records[[6]]$variant_types[[2]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001886"
#> 
#> 
#> 
#> $content$records[[6]]$civic_actionability_score
#> [1] 19
#> 
#> $content$records[[6]]$coordinates
#> $content$records[[6]]$coordinates$chromosome
#> [1] "2"
#> 
#> $content$records[[6]]$coordinates$start
#> [1] 29445258
#> 
#> $content$records[[6]]$coordinates$stop
#> [1] 29445258
#> 
#> $content$records[[6]]$coordinates$reference_bases
#> [1] "C"
#> 
#> $content$records[[6]]$coordinates$variant_bases
#> [1] "T"
#> 
#> $content$records[[6]]$coordinates$representative_transcript
#> [1] "ENST00000389048.3"
#> 
#> $content$records[[6]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[6]]$coordinates$start2
#> NULL
#> 
#> $content$records[[6]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[6]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[6]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[6]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[7]]
#> $content$records[[7]]$id
#> [1] 7
#> 
#> $content$records[[7]]$entrez_name
#> [1] "ALK"
#> 
#> $content$records[[7]]$entrez_id
#> [1] 238
#> 
#> $content$records[[7]]$name
#> [1] "EML4-ALK L1196M"
#> 
#> $content$records[[7]]$description
#> [1] "In patients with non-small cell lung cancer exhibiting EML4-ALK fusion, L1196M has been shown to confer resistance to crizotinib. This was also true in a patient with both EML4-ALK C1156Y & L1196M."
#> 
#> $content$records[[7]]$gene_id
#> [1] 1
#> 
#> $content$records[[7]]$type
#> [1] "variant"
#> 
#> $content$records[[7]]$variant_types
#> $content$records[[7]]$variant_types[[1]]
#> $content$records[[7]]$variant_types[[1]]$id
#> [1] 47
#> 
#> $content$records[[7]]$variant_types[[1]]$name
#> [1] "missense_variant"
#> 
#> $content$records[[7]]$variant_types[[1]]$display_name
#> [1] "Missense Variant"
#> 
#> $content$records[[7]]$variant_types[[1]]$so_id
#> [1] "SO:0001583"
#> 
#> $content$records[[7]]$variant_types[[1]]$description
#> [1] "A sequence variant, that changes one or more bases, resulting in a different amino acid sequence but where the length is preserved."
#> 
#> $content$records[[7]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001583"
#> 
#> 
#> $content$records[[7]]$variant_types[[2]]
#> $content$records[[7]]$variant_types[[2]]$id
#> [1] 120
#> 
#> $content$records[[7]]$variant_types[[2]]$name
#> [1] "transcript_fusion"
#> 
#> $content$records[[7]]$variant_types[[2]]$display_name
#> [1] "Transcript Fusion"
#> 
#> $content$records[[7]]$variant_types[[2]]$so_id
#> [1] "SO:0001886"
#> 
#> $content$records[[7]]$variant_types[[2]]$description
#> [1] "A feature fusion where the deletion brings together transcript regions."
#> 
#> $content$records[[7]]$variant_types[[2]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001886"
#> 
#> 
#> 
#> $content$records[[7]]$civic_actionability_score
#> [1] 31
#> 
#> $content$records[[7]]$coordinates
#> $content$records[[7]]$coordinates$chromosome
#> [1] "2"
#> 
#> $content$records[[7]]$coordinates$start
#> [1] 29443631
#> 
#> $content$records[[7]]$coordinates$stop
#> [1] 29443631
#> 
#> $content$records[[7]]$coordinates$reference_bases
#> [1] "G"
#> 
#> $content$records[[7]]$coordinates$variant_bases
#> [1] "T"
#> 
#> $content$records[[7]]$coordinates$representative_transcript
#> [1] "ENST00000389048.3"
#> 
#> $content$records[[7]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[7]]$coordinates$start2
#> NULL
#> 
#> $content$records[[7]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[7]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[7]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[7]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[8]]
#> $content$records[[8]]$id
#> [1] 8
#> 
#> $content$records[[8]]$entrez_name
#> [1] "ALK"
#> 
#> $content$records[[8]]$entrez_id
#> [1] 238
#> 
#> $content$records[[8]]$name
#> [1] "F1174L"
#> 
#> $content$records[[8]]$description
#> [1] "ALK F1174L has been observed as recurrent in neuroblastoma, non-small cell lung cancer (NSCLC), and other cancer types. Neuroblastoma cells containing this mutation have shown resistance to low doses of criztonib. However, increased dosage can overcome this resistance in cell lines studies. TAE684 has also proven effective in both NSCLC and neuroblastoma F1174L containing cells."
#> 
#> $content$records[[8]]$gene_id
#> [1] 1
#> 
#> $content$records[[8]]$type
#> [1] "variant"
#> 
#> $content$records[[8]]$variant_types
#> $content$records[[8]]$variant_types[[1]]
#> $content$records[[8]]$variant_types[[1]]$id
#> [1] 47
#> 
#> $content$records[[8]]$variant_types[[1]]$name
#> [1] "missense_variant"
#> 
#> $content$records[[8]]$variant_types[[1]]$display_name
#> [1] "Missense Variant"
#> 
#> $content$records[[8]]$variant_types[[1]]$so_id
#> [1] "SO:0001583"
#> 
#> $content$records[[8]]$variant_types[[1]]$description
#> [1] "A sequence variant, that changes one or more bases, resulting in a different amino acid sequence but where the length is preserved."
#> 
#> $content$records[[8]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001583"
#> 
#> 
#> 
#> $content$records[[8]]$civic_actionability_score
#> [1] 33.5
#> 
#> $content$records[[8]]$coordinates
#> $content$records[[8]]$coordinates$chromosome
#> [1] "2"
#> 
#> $content$records[[8]]$coordinates$start
#> [1] 29443695
#> 
#> $content$records[[8]]$coordinates$stop
#> [1] 29443695
#> 
#> $content$records[[8]]$coordinates$reference_bases
#> [1] "G"
#> 
#> $content$records[[8]]$coordinates$variant_bases
#> [1] "T"
#> 
#> $content$records[[8]]$coordinates$representative_transcript
#> [1] "ENST00000389048.3"
#> 
#> $content$records[[8]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[8]]$coordinates$start2
#> NULL
#> 
#> $content$records[[8]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[8]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[8]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[8]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[9]]
#> $content$records[[9]]$id
#> [1] 9
#> 
#> $content$records[[9]]$entrez_name
#> [1] "ALK"
#> 
#> $content$records[[9]]$entrez_id
#> [1] 238
#> 
#> $content$records[[9]]$name
#> [1] "R1275Q"
#> 
#> $content$records[[9]]$description
#> [1] "ALK R1275Q has been observed as a recurrent mutation in  neuroblastoma, non-small cell lung cancer (NSCLC), as well as other cancer types. Neuroblastoma cells with this mutation have shown sensitivity to the ALK-inhibitor TAE684. This and the geldanamycin deriviative 17-DMAG have been shown to be effective in NSCLC cell lines."
#> 
#> $content$records[[9]]$gene_id
#> [1] 1
#> 
#> $content$records[[9]]$type
#> [1] "variant"
#> 
#> $content$records[[9]]$variant_types
#> $content$records[[9]]$variant_types[[1]]
#> $content$records[[9]]$variant_types[[1]]$id
#> [1] 47
#> 
#> $content$records[[9]]$variant_types[[1]]$name
#> [1] "missense_variant"
#> 
#> $content$records[[9]]$variant_types[[1]]$display_name
#> [1] "Missense Variant"
#> 
#> $content$records[[9]]$variant_types[[1]]$so_id
#> [1] "SO:0001583"
#> 
#> $content$records[[9]]$variant_types[[1]]$description
#> [1] "A sequence variant, that changes one or more bases, resulting in a different amino acid sequence but where the length is preserved."
#> 
#> $content$records[[9]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001583"
#> 
#> 
#> 
#> $content$records[[9]]$civic_actionability_score
#> [1] 23
#> 
#> $content$records[[9]]$coordinates
#> $content$records[[9]]$coordinates$chromosome
#> [1] "2"
#> 
#> $content$records[[9]]$coordinates$start
#> [1] 29432664
#> 
#> $content$records[[9]]$coordinates$stop
#> [1] 29432664
#> 
#> $content$records[[9]]$coordinates$reference_bases
#> [1] "C"
#> 
#> $content$records[[9]]$coordinates$variant_bases
#> [1] "T"
#> 
#> $content$records[[9]]$coordinates$representative_transcript
#> [1] "ENST00000389048.3"
#> 
#> $content$records[[9]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[9]]$coordinates$start2
#> NULL
#> 
#> $content$records[[9]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[9]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[9]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[9]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[10]]
#> $content$records[[10]]$id
#> [1] 10
#> 
#> $content$records[[10]]$entrez_name
#> [1] "ARAF"
#> 
#> $content$records[[10]]$entrez_id
#> [1] 369
#> 
#> $content$records[[10]]$name
#> [1] "S214C"
#> 
#> $content$records[[10]]$description
#> [1] "ARAF S214C has been found to be a recurrent oncogenic mutation in non-small cell lung cancer. It has been shown to confer sensitivity to sorafenib and trametenib in cell lines. In a case study of advanced stage lung adenocarcinoma harboring this mutation, sorafenib also acheived near-complete clinical remission. This case has brought more interest to the variant from a research and clinical perspective."
#> 
#> $content$records[[10]]$gene_id
#> [1] 3
#> 
#> $content$records[[10]]$type
#> [1] "variant"
#> 
#> $content$records[[10]]$variant_types
#> $content$records[[10]]$variant_types[[1]]
#> $content$records[[10]]$variant_types[[1]]$id
#> [1] 47
#> 
#> $content$records[[10]]$variant_types[[1]]$name
#> [1] "missense_variant"
#> 
#> $content$records[[10]]$variant_types[[1]]$display_name
#> [1] "Missense Variant"
#> 
#> $content$records[[10]]$variant_types[[1]]$so_id
#> [1] "SO:0001583"
#> 
#> $content$records[[10]]$variant_types[[1]]$description
#> [1] "A sequence variant, that changes one or more bases, resulting in a different amino acid sequence but where the length is preserved."
#> 
#> $content$records[[10]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001583"
#> 
#> 
#> 
#> $content$records[[10]]$civic_actionability_score
#> [1] 11
#> 
#> $content$records[[10]]$coordinates
#> $content$records[[10]]$coordinates$chromosome
#> [1] "X"
#> 
#> $content$records[[10]]$coordinates$start
#> [1] 47426121
#> 
#> $content$records[[10]]$coordinates$stop
#> [1] 47426121
#> 
#> $content$records[[10]]$coordinates$reference_bases
#> [1] "C"
#> 
#> $content$records[[10]]$coordinates$variant_bases
#> [1] "G"
#> 
#> $content$records[[10]]$coordinates$representative_transcript
#> [1] "ENST00000377045.4"
#> 
#> $content$records[[10]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[10]]$coordinates$start2
#> NULL
#> 
#> $content$records[[10]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[10]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[10]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[10]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> 
#> 
#> $url
#> [1] "https://civicdb.org/api/variants"
#> 
#> $response
#> Response [https://civicdb.org/api/variants?page=1&count=10]
#>   Date: 2019-09-03 16:44
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 13.9 kB
#> 
#> 
#> attr(,"class")
#> [1] "civic_api"

getAllVariants(page = 2, count = 10)

#> $content
#> $content$`_meta`
#> $content$`_meta`$current_page
#> [1] 2
#> 
#> $content$`_meta`$per_page
#> [1] "10"
#> 
#> $content$`_meta`$total_pages
#> [1] 229
#> 
#> $content$`_meta`$total_count
#> [1] 2286
#> 
#> $content$`_meta`$links
#> $content$`_meta`$links$`next`
#> [1] "https://civicdb.org/api/variants?count=10&page=3"
#> 
#> $content$`_meta`$links$previous
#> [1] "https://civicdb.org/api/variants?count=10&page=1"
#> 
#> 
#> 
#> $content$records
#> $content$records[[1]]
#> $content$records[[1]]$id
#> [1] 11
#> 
#> $content$records[[1]]$entrez_name
#> [1] "BRAF"
#> 
#> $content$records[[1]]$entrez_id
#> [1] 673
#> 
#> $content$records[[1]]$name
#> [1] "V600D"
#> 
#> $content$records[[1]]$description
#> [1] "Patients harboring mutations in valine 600 residue of BRAF have been shown to be sensitive to dabrafenib. For more information on the V600 locus, see the V600E entry."
#> 
#> $content$records[[1]]$gene_id
#> [1] 5
#> 
#> $content$records[[1]]$type
#> [1] "variant"
#> 
#> $content$records[[1]]$variant_types
#> $content$records[[1]]$variant_types[[1]]
#> $content$records[[1]]$variant_types[[1]]$id
#> [1] 47
#> 
#> $content$records[[1]]$variant_types[[1]]$name
#> [1] "missense_variant"
#> 
#> $content$records[[1]]$variant_types[[1]]$display_name
#> [1] "Missense Variant"
#> 
#> $content$records[[1]]$variant_types[[1]]$so_id
#> [1] "SO:0001583"
#> 
#> $content$records[[1]]$variant_types[[1]]$description
#> [1] "A sequence variant, that changes one or more bases, resulting in a different amino acid sequence but where the length is preserved."
#> 
#> $content$records[[1]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001583"
#> 
#> 
#> 
#> $content$records[[1]]$civic_actionability_score
#> [1] 25
#> 
#> $content$records[[1]]$coordinates
#> $content$records[[1]]$coordinates$chromosome
#> [1] "7"
#> 
#> $content$records[[1]]$coordinates$start
#> [1] 140453135
#> 
#> $content$records[[1]]$coordinates$stop
#> [1] 140453136
#> 
#> $content$records[[1]]$coordinates$reference_bases
#> [1] "CA"
#> 
#> $content$records[[1]]$coordinates$variant_bases
#> [1] "AT"
#> 
#> $content$records[[1]]$coordinates$representative_transcript
#> [1] "ENST00000288602.6"
#> 
#> $content$records[[1]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[1]]$coordinates$start2
#> NULL
#> 
#> $content$records[[1]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[1]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[1]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[1]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[2]]
#> $content$records[[2]]$id
#> [1] 12
#> 
#> $content$records[[2]]$entrez_name
#> [1] "BRAF"
#> 
#> $content$records[[2]]$entrez_id
#> [1] 673
#> 
#> $content$records[[2]]$name
#> [1] "V600E"
#> 
#> $content$records[[2]]$description
#> [1] "BRAF V600E has been shown to be recurrent in many cancer types. It is one of the most widely studied variants in cancer. This variant is correlated with poor prognosis in certain cancer types, including colorectal cancer and papillary thyroid cancer. The targeted therapeutic dabrafenib has been shown to be effective in clinical trials with an array of BRAF mutations and cancer types. Dabrafenib has also shown to be effective when combined with the MEK inhibitor trametinib in colorectal cancer and melanoma. However, in patients with TP53, CDKN2A and KRAS mutations, dabrafenib resistance has been reported. Ipilimumab, regorafenib, vemurafenib, and a number of combination therapies have been successful in treating V600E mutations. However, cetuximab and panitumumab have been largely shown to be ineffective without supplementary treatment."
#> 
#> $content$records[[2]]$gene_id
#> [1] 5
#> 
#> $content$records[[2]]$type
#> [1] "variant"
#> 
#> $content$records[[2]]$variant_types
#> $content$records[[2]]$variant_types[[1]]
#> $content$records[[2]]$variant_types[[1]]$id
#> [1] 47
#> 
#> $content$records[[2]]$variant_types[[1]]$name
#> [1] "missense_variant"
#> 
#> $content$records[[2]]$variant_types[[1]]$display_name
#> [1] "Missense Variant"
#> 
#> $content$records[[2]]$variant_types[[1]]$so_id
#> [1] "SO:0001583"
#> 
#> $content$records[[2]]$variant_types[[1]]$description
#> [1] "A sequence variant, that changes one or more bases, resulting in a different amino acid sequence but where the length is preserved."
#> 
#> $content$records[[2]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001583"
#> 
#> 
#> 
#> $content$records[[2]]$civic_actionability_score
#> [1] 1019
#> 
#> $content$records[[2]]$coordinates
#> $content$records[[2]]$coordinates$chromosome
#> [1] "7"
#> 
#> $content$records[[2]]$coordinates$start
#> [1] 140453136
#> 
#> $content$records[[2]]$coordinates$stop
#> [1] 140453136
#> 
#> $content$records[[2]]$coordinates$reference_bases
#> [1] "A"
#> 
#> $content$records[[2]]$coordinates$variant_bases
#> [1] "T"
#> 
#> $content$records[[2]]$coordinates$representative_transcript
#> [1] "ENST00000288602.6"
#> 
#> $content$records[[2]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[2]]$coordinates$start2
#> NULL
#> 
#> $content$records[[2]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[2]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[2]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[2]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[3]]
#> $content$records[[3]]$id
#> [1] 13
#> 
#> $content$records[[3]]$entrez_name
#> [1] "BRAF"
#> 
#> $content$records[[3]]$entrez_id
#> [1] 673
#> 
#> $content$records[[3]]$name
#> [1] "V600E+V600M"
#> 
#> $content$records[[3]]$description
#> [1] "A case study of a single patient harboring both a V600E and a V600M mutation, dabrafenib was shown to acheive clinical response."
#> 
#> $content$records[[3]]$gene_id
#> [1] 5
#> 
#> $content$records[[3]]$type
#> [1] "variant"
#> 
#> $content$records[[3]]$variant_types
#> $content$records[[3]]$variant_types[[1]]
#> $content$records[[3]]$variant_types[[1]]$id
#> [1] 47
#> 
#> $content$records[[3]]$variant_types[[1]]$name
#> [1] "missense_variant"
#> 
#> $content$records[[3]]$variant_types[[1]]$display_name
#> [1] "Missense Variant"
#> 
#> $content$records[[3]]$variant_types[[1]]$so_id
#> [1] "SO:0001583"
#> 
#> $content$records[[3]]$variant_types[[1]]$description
#> [1] "A sequence variant, that changes one or more bases, resulting in a different amino acid sequence but where the length is preserved."
#> 
#> $content$records[[3]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001583"
#> 
#> 
#> 
#> $content$records[[3]]$civic_actionability_score
#> [1] 7.5
#> 
#> $content$records[[3]]$coordinates
#> $content$records[[3]]$coordinates$chromosome
#> [1] "7"
#> 
#> $content$records[[3]]$coordinates$start
#> [1] 140453135
#> 
#> $content$records[[3]]$coordinates$stop
#> [1] 140453137
#> 
#> $content$records[[3]]$coordinates$reference_bases
#> NULL
#> 
#> $content$records[[3]]$coordinates$variant_bases
#> NULL
#> 
#> $content$records[[3]]$coordinates$representative_transcript
#> [1] "ENST00000288602.6"
#> 
#> $content$records[[3]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[3]]$coordinates$start2
#> NULL
#> 
#> $content$records[[3]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[3]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[3]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[3]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[4]]
#> $content$records[[4]]$id
#> [1] 14
#> 
#> $content$records[[4]]$entrez_name
#> [1] "BRAF"
#> 
#> $content$records[[4]]$entrez_id
#> [1] 673
#> 
#> $content$records[[4]]$name
#> [1] "V600E AMPLIFICATION"
#> 
#> $content$records[[4]]$description
#> [1] "Amplification of BRAF V600E has been shown to confer resistance to MEK inhibitors. For more information on the V600 locus, see the V600E entry."
#> 
#> $content$records[[4]]$gene_id
#> [1] 5
#> 
#> $content$records[[4]]$type
#> [1] "variant"
#> 
#> $content$records[[4]]$variant_types
#> $content$records[[4]]$variant_types[[1]]
#> $content$records[[4]]$variant_types[[1]]$id
#> [1] 47
#> 
#> $content$records[[4]]$variant_types[[1]]$name
#> [1] "missense_variant"
#> 
#> $content$records[[4]]$variant_types[[1]]$display_name
#> [1] "Missense Variant"
#> 
#> $content$records[[4]]$variant_types[[1]]$so_id
#> [1] "SO:0001583"
#> 
#> $content$records[[4]]$variant_types[[1]]$description
#> [1] "A sequence variant, that changes one or more bases, resulting in a different amino acid sequence but where the length is preserved."
#> 
#> $content$records[[4]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001583"
#> 
#> 
#> $content$records[[4]]$variant_types[[2]]
#> $content$records[[4]]$variant_types[[2]]$id
#> [1] 123
#> 
#> $content$records[[4]]$variant_types[[2]]$name
#> [1] "transcript_amplification"
#> 
#> $content$records[[4]]$variant_types[[2]]$display_name
#> [1] "Transcript Amplification"
#> 
#> $content$records[[4]]$variant_types[[2]]$so_id
#> [1] "SO:0001889"
#> 
#> $content$records[[4]]$variant_types[[2]]$description
#> [1] "A feature amplification of a region containing a transcript."
#> 
#> $content$records[[4]]$variant_types[[2]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001889"
#> 
#> 
#> 
#> $content$records[[4]]$civic_actionability_score
#> [1] 1
#> 
#> $content$records[[4]]$coordinates
#> $content$records[[4]]$coordinates$chromosome
#> [1] "7"
#> 
#> $content$records[[4]]$coordinates$start
#> [1] 140434279
#> 
#> $content$records[[4]]$coordinates$stop
#> [1] 140624564
#> 
#> $content$records[[4]]$coordinates$reference_bases
#> NULL
#> 
#> $content$records[[4]]$coordinates$variant_bases
#> NULL
#> 
#> $content$records[[4]]$coordinates$representative_transcript
#> [1] "ENST00000288602.6"
#> 
#> $content$records[[4]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[4]]$coordinates$start2
#> NULL
#> 
#> $content$records[[4]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[4]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[4]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[4]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[5]]
#> $content$records[[5]]$id
#> [1] 17
#> 
#> $content$records[[5]]$entrez_name
#> [1] "BRAF"
#> 
#> $content$records[[5]]$entrez_id
#> [1] 673
#> 
#> $content$records[[5]]$name
#> [1] "V600"
#> 
#> $content$records[[5]]$description
#> [1] "BRAF mutations of the valine 600 residue have been shown to be recurrent in many cancer types. Of the V600 mutations, V600E is the most widely researched. V600 mutations as a whole have been correlated to poorer prognosis in colorectal and papilarry thyroid cancers. V600 mutations have also been shown to confer sensitivity to the BRAF inhibitor dabrafenib. For a more detailed summary, click the individual mutations."
#> 
#> $content$records[[5]]$gene_id
#> [1] 5
#> 
#> $content$records[[5]]$type
#> [1] "variant"
#> 
#> $content$records[[5]]$variant_types
#> $content$records[[5]]$variant_types[[1]]
#> $content$records[[5]]$variant_types[[1]]$id
#> [1] 103
#> 
#> $content$records[[5]]$variant_types[[1]]$name
#> [1] "protein_altering_variant"
#> 
#> $content$records[[5]]$variant_types[[1]]$display_name
#> [1] "Protein Altering Variant"
#> 
#> $content$records[[5]]$variant_types[[1]]$so_id
#> [1] "SO:0001818"
#> 
#> $content$records[[5]]$variant_types[[1]]$description
#> [1] "A sequence_variant which is predicted to change the protein encoded in the coding sequence."
#> 
#> $content$records[[5]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001818"
#> 
#> 
#> 
#> $content$records[[5]]$civic_actionability_score
#> [1] 420
#> 
#> $content$records[[5]]$coordinates
#> $content$records[[5]]$coordinates$chromosome
#> [1] "7"
#> 
#> $content$records[[5]]$coordinates$start
#> [1] 140453136
#> 
#> $content$records[[5]]$coordinates$stop
#> [1] 140453137
#> 
#> $content$records[[5]]$coordinates$reference_bases
#> NULL
#> 
#> $content$records[[5]]$coordinates$variant_bases
#> NULL
#> 
#> $content$records[[5]]$coordinates$representative_transcript
#> [1] "ENST00000288602.6"
#> 
#> $content$records[[5]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[5]]$coordinates$start2
#> NULL
#> 
#> $content$records[[5]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[5]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[5]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[5]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[6]]
#> $content$records[[6]]$id
#> [1] 18
#> 
#> $content$records[[6]]$entrez_name
#> [1] "CCND1"
#> 
#> $content$records[[6]]$entrez_id
#> [1] 595
#> 
#> $content$records[[6]]$name
#> [1] "AMPLIFICATION"
#> 
#> $content$records[[6]]$description
#> [1] "CCND1 amplification has been implicated in poorer prognosis in non-small cell lung cancer."
#> 
#> $content$records[[6]]$gene_id
#> [1] 8
#> 
#> $content$records[[6]]$type
#> [1] "variant"
#> 
#> $content$records[[6]]$variant_types
#> $content$records[[6]]$variant_types[[1]]
#> $content$records[[6]]$variant_types[[1]]$id
#> [1] 123
#> 
#> $content$records[[6]]$variant_types[[1]]$name
#> [1] "transcript_amplification"
#> 
#> $content$records[[6]]$variant_types[[1]]$display_name
#> [1] "Transcript Amplification"
#> 
#> $content$records[[6]]$variant_types[[1]]$so_id
#> [1] "SO:0001889"
#> 
#> $content$records[[6]]$variant_types[[1]]$description
#> [1] "A feature amplification of a region containing a transcript."
#> 
#> $content$records[[6]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001889"
#> 
#> 
#> 
#> $content$records[[6]]$civic_actionability_score
#> [1] 47
#> 
#> $content$records[[6]]$coordinates
#> $content$records[[6]]$coordinates$chromosome
#> [1] "11"
#> 
#> $content$records[[6]]$coordinates$start
#> [1] 69455855
#> 
#> $content$records[[6]]$coordinates$stop
#> [1] 69469242
#> 
#> $content$records[[6]]$coordinates$reference_bases
#> NULL
#> 
#> $content$records[[6]]$coordinates$variant_bases
#> NULL
#> 
#> $content$records[[6]]$coordinates$representative_transcript
#> [1] "ENST00000227507.2"
#> 
#> $content$records[[6]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[6]]$coordinates$start2
#> NULL
#> 
#> $content$records[[6]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[6]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[6]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[6]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[7]]
#> $content$records[[7]]$id
#> [1] 19
#> 
#> $content$records[[7]]$entrez_name
#> [1] "CCND1"
#> 
#> $content$records[[7]]$entrez_id
#> [1] 595
#> 
#> $content$records[[7]]$name
#> [1] "EXPRESSION"
#> 
#> $content$records[[7]]$description
#> [1] "CCND1 expression, and its prognositc impact, is still in dispute. Three experiments in non-small cell lung cancer have shown it to have no impact on survival, but three additional studies have shown it results in poorer prognosis. There is also some ambiguity in how the boundaries between expression and overexpression are defined."
#> 
#> $content$records[[7]]$gene_id
#> [1] 8
#> 
#> $content$records[[7]]$type
#> [1] "variant"
#> 
#> $content$records[[7]]$variant_types
#> $content$records[[7]]$variant_types[[1]]
#> $content$records[[7]]$variant_types[[1]]$id
#> [1] 183
#> 
#> $content$records[[7]]$variant_types[[1]]$name
#> [1] "N/A"
#> 
#> $content$records[[7]]$variant_types[[1]]$display_name
#> [1] "N/A"
#> 
#> $content$records[[7]]$variant_types[[1]]$so_id
#> [1] "N/A"
#> 
#> $content$records[[7]]$variant_types[[1]]$description
#> [1] "No suitable Sequence Ontology term exists."
#> 
#> $content$records[[7]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/N/A"
#> 
#> 
#> 
#> $content$records[[7]]$civic_actionability_score
#> [1] 30
#> 
#> $content$records[[7]]$coordinates
#> $content$records[[7]]$coordinates$chromosome
#> [1] "11"
#> 
#> $content$records[[7]]$coordinates$start
#> [1] 69455855
#> 
#> $content$records[[7]]$coordinates$stop
#> [1] 69469242
#> 
#> $content$records[[7]]$coordinates$reference_bases
#> NULL
#> 
#> $content$records[[7]]$coordinates$variant_bases
#> NULL
#> 
#> $content$records[[7]]$coordinates$representative_transcript
#> [1] "ENST00000227507.2"
#> 
#> $content$records[[7]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[7]]$coordinates$start2
#> NULL
#> 
#> $content$records[[7]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[7]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[7]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[7]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[8]]
#> $content$records[[8]]$id
#> [1] 20
#> 
#> $content$records[[8]]$entrez_name
#> [1] "CCND1"
#> 
#> $content$records[[8]]$entrez_id
#> [1] 595
#> 
#> $content$records[[8]]$name
#> [1] "OVEREXPRESSION"
#> 
#> $content$records[[8]]$description
#> [1] "Cyclin D has been shown in many cancer types to be misregulated. Well established for their oncogenic properties, the cyclins and the cyclin-dependent kinases (CDK's) they activate have been the focus of major research and development efforts over the past decade. The methods by which the cyclins are misregulated are widely variable, and range from genomic amplification to promoter methylation changes. While Cyclin D2 has only been found to be significantly misregulated in glioma, Cyclin D1 in particular seems to be a pan-cancer actor. Cyclin D misregulation has been shown to lead to poorer outcomes in a number of studies, and currently there are no FDA-approved targeted therapies."
#> 
#> $content$records[[8]]$gene_id
#> [1] 8
#> 
#> $content$records[[8]]$type
#> [1] "variant"
#> 
#> $content$records[[8]]$variant_types
#> $content$records[[8]]$variant_types[[1]]
#> $content$records[[8]]$variant_types[[1]]$id
#> [1] 183
#> 
#> $content$records[[8]]$variant_types[[1]]$name
#> [1] "N/A"
#> 
#> $content$records[[8]]$variant_types[[1]]$display_name
#> [1] "N/A"
#> 
#> $content$records[[8]]$variant_types[[1]]$so_id
#> [1] "N/A"
#> 
#> $content$records[[8]]$variant_types[[1]]$description
#> [1] "No suitable Sequence Ontology term exists."
#> 
#> $content$records[[8]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/N/A"
#> 
#> 
#> 
#> $content$records[[8]]$civic_actionability_score
#> [1] 106
#> 
#> $content$records[[8]]$coordinates
#> $content$records[[8]]$coordinates$chromosome
#> [1] "11"
#> 
#> $content$records[[8]]$coordinates$start
#> [1] 69455855
#> 
#> $content$records[[8]]$coordinates$stop
#> [1] 69469242
#> 
#> $content$records[[8]]$coordinates$reference_bases
#> NULL
#> 
#> $content$records[[8]]$coordinates$variant_bases
#> NULL
#> 
#> $content$records[[8]]$coordinates$representative_transcript
#> [1] "ENST00000227507.2"
#> 
#> $content$records[[8]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[8]]$coordinates$start2
#> NULL
#> 
#> $content$records[[8]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[8]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[8]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[8]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[9]]
#> $content$records[[9]]$id
#> [1] 21
#> 
#> $content$records[[9]]$entrez_name
#> [1] "CCND2"
#> 
#> $content$records[[9]]$entrez_id
#> [1] 894
#> 
#> $content$records[[9]]$name
#> [1] "OVEREXPRESSION"
#> 
#> $content$records[[9]]$description
#> [1] "Cyclin D has been shown in many cancer types to be misregulated. Well established for their oncogenic properties, the cyclins and the cyclin-dependent kinases (CDK's) they activate have been the focus of major research and development efforts over the past decade. The methods by which the cyclins are misregulated are widely variable, and range from genomic amplification to promoter methylation changes. While Cyclin D2 has only been found to be significantly misregulated in glioma, Cyclin D1 in particular seems to be a pan-cancer actor. Cyclin D misregulation has been shown to lead to poorer outcomes in a number of studies, and currently there are no FDA-approved targeted therapies."
#> 
#> $content$records[[9]]$gene_id
#> [1] 9
#> 
#> $content$records[[9]]$type
#> [1] "variant"
#> 
#> $content$records[[9]]$variant_types
#> $content$records[[9]]$variant_types[[1]]
#> $content$records[[9]]$variant_types[[1]]$id
#> [1] 183
#> 
#> $content$records[[9]]$variant_types[[1]]$name
#> [1] "N/A"
#> 
#> $content$records[[9]]$variant_types[[1]]$display_name
#> [1] "N/A"
#> 
#> $content$records[[9]]$variant_types[[1]]$so_id
#> [1] "N/A"
#> 
#> $content$records[[9]]$variant_types[[1]]$description
#> [1] "No suitable Sequence Ontology term exists."
#> 
#> $content$records[[9]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/N/A"
#> 
#> 
#> 
#> $content$records[[9]]$civic_actionability_score
#> [1] 16
#> 
#> $content$records[[9]]$coordinates
#> $content$records[[9]]$coordinates$chromosome
#> [1] "12"
#> 
#> $content$records[[9]]$coordinates$start
#> [1] 4382938
#> 
#> $content$records[[9]]$coordinates$stop
#> [1] 4414516
#> 
#> $content$records[[9]]$coordinates$reference_bases
#> NULL
#> 
#> $content$records[[9]]$coordinates$variant_bases
#> NULL
#> 
#> $content$records[[9]]$coordinates$representative_transcript
#> [1] "ENST00000261254.3"
#> 
#> $content$records[[9]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[9]]$coordinates$start2
#> NULL
#> 
#> $content$records[[9]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[9]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[9]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[9]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> $content$records[[10]]
#> $content$records[[10]]$id
#> [1] 22
#> 
#> $content$records[[10]]$entrez_name
#> [1] "CCND2"
#> 
#> $content$records[[10]]$entrez_id
#> [1] 894
#> 
#> $content$records[[10]]$name
#> [1] "PROMOTER DEMETHYLATION"
#> 
#> $content$records[[10]]$description
#> [1] "Cyclin D has been shown in many cancer types to be misregulated. Well established for their oncogenic properties, the cyclins and the cyclin-dependent kinases (CDK's) they activate have been the focus of major research and development efforts over the past decade. The methods by which the cyclins are misregulated are widely variable, and range from genomic amplification to promoter methylation changes. While Cyclin D2 has only been found to be significantly misregulated in glioma, Cyclin D1 in particular seems to be a pan-cancer actor. Cyclin D misregulation has been shown to lead to poorer outcomes in a number of studies, and currently there are no FDA-approved targeted therapies."
#> 
#> $content$records[[10]]$gene_id
#> [1] 9
#> 
#> $content$records[[10]]$type
#> [1] "variant"
#> 
#> $content$records[[10]]$variant_types
#> $content$records[[10]]$variant_types[[1]]
#> $content$records[[10]]$variant_types[[1]]$id
#> [1] 183
#> 
#> $content$records[[10]]$variant_types[[1]]$name
#> [1] "N/A"
#> 
#> $content$records[[10]]$variant_types[[1]]$display_name
#> [1] "N/A"
#> 
#> $content$records[[10]]$variant_types[[1]]$so_id
#> [1] "N/A"
#> 
#> $content$records[[10]]$variant_types[[1]]$description
#> [1] "No suitable Sequence Ontology term exists."
#> 
#> $content$records[[10]]$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/N/A"
#> 
#> 
#> 
#> $content$records[[10]]$civic_actionability_score
#> [1] 3
#> 
#> $content$records[[10]]$coordinates
#> $content$records[[10]]$coordinates$chromosome
#> [1] "12"
#> 
#> $content$records[[10]]$coordinates$start
#> [1] 4381437
#> 
#> $content$records[[10]]$coordinates$stop
#> [1] 4382937
#> 
#> $content$records[[10]]$coordinates$reference_bases
#> NULL
#> 
#> $content$records[[10]]$coordinates$variant_bases
#> NULL
#> 
#> $content$records[[10]]$coordinates$representative_transcript
#> [1] "ENST00000261254.3"
#> 
#> $content$records[[10]]$coordinates$chromosome2
#> NULL
#> 
#> $content$records[[10]]$coordinates$start2
#> NULL
#> 
#> $content$records[[10]]$coordinates$stop2
#> NULL
#> 
#> $content$records[[10]]$coordinates$representative_transcript2
#> NULL
#> 
#> $content$records[[10]]$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$records[[10]]$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> 
#> 
#> 
#> $url
#> [1] "https://civicdb.org/api/variants"
#> 
#> $response
#> Response [https://civicdb.org/api/variants?page=2&count=10]
#>   Date: 2019-09-03 16:44
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 11.8 kB
#> 
#> 
#> attr(,"class")
#> [1] "civic_api"