Get a specific assertion metadata information

Retrieve metadata information for a specific variant group from the CIViC DB

getAssertionMetadata(id, type)

Arguments

id
Internal CIViC ID of the assertion of interest

type
Type of metadata (comments, suggested_changes, revisions)

Value

An S3 Object of type civic_api containing the content, url, and response

Examples

getAssertionMetadata(id = 1, type = "comments")

#> $content
#> list()
#> 
#> $url
#> [1] "https://civicdb.org/api/assertions/1/comments"
#> 
#> $response
#> Response [https://civicdb.org/api/assertions/1/comments]
#>   Date: 2019-09-03 16:44
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 2 B
#> 
#> 
#> attr(,"class")
#> [1] "civic_api"

getAssertionMetadata(id = 1, type = "suggested_changes")

#> $content
#> $content[[1]]
#> $content[[1]]$id
#> [1] 31799
#> 
#> $content[[1]]$user
#> $content[[1]]$user$id
#> [1] 3
#> 
#> $content[[1]]$user$name
#> [1] "Obi Griffith"
#> 
#> $content[[1]]$user$last_seen_at
#> [1] "2019-08-26T20:34:15.034Z"
#> 
#> $content[[1]]$user$username
#> [1] "obigriffith"
#> 
#> $content[[1]]$user$role
#> [1] "admin"
#> 
#> $content[[1]]$user$avatar_url
#> [1] "https://secure.gravatar.com/avatar/4c468e9a95d6135e02eb66ee5f2fb574.png?d=identicon&r=pg&s=32"
#> 
#> $content[[1]]$user$avatars
#> $content[[1]]$user$avatars$x128
#> [1] "https://secure.gravatar.com/avatar/4c468e9a95d6135e02eb66ee5f2fb574.png?d=identicon&r=pg&s=128"
#> 
#> $content[[1]]$user$avatars$x64
#> [1] "https://secure.gravatar.com/avatar/4c468e9a95d6135e02eb66ee5f2fb574.png?d=identicon&r=pg&s=64"
#> 
#> $content[[1]]$user$avatars$x32
#> [1] "https://secure.gravatar.com/avatar/4c468e9a95d6135e02eb66ee5f2fb574.png?d=identicon&r=pg&s=32"
#> 
#> $content[[1]]$user$avatars$x14
#> [1] "https://secure.gravatar.com/avatar/4c468e9a95d6135e02eb66ee5f2fb574.png?d=identicon&r=pg&s=14"
#> 
#> 
#> $content[[1]]$user$area_of_expertise
#> [1] "Research Scientist"
#> 
#> $content[[1]]$user$orcid
#> [1] "0000-0002-0843-4271"
#> 
#> $content[[1]]$user$display_name
#> [1] "obigriffith"
#> 
#> $content[[1]]$user$created_at
#> [1] "2015-02-26T02:53:49.147Z"
#> 
#> $content[[1]]$user$url
#> [1] "http://genome.wustl.edu/people/individual/obi-griffith/"
#> 
#> $content[[1]]$user$twitter_handle
#> [1] "obigriffith"
#> 
#> $content[[1]]$user$facebook_profile
#> [1] ""
#> 
#> $content[[1]]$user$linkedin_profile
#> [1] "obigriffith"
#> 
#> $content[[1]]$user$bio
#> [1] "Dr. Griffith is Assistant Professor of Medicine and Assistant Director of the McDonnell Genome Institute at Washington University School of Medicine. He has worked in genomics and bioinformatics since the earliest phase of reference genome sequencing. He contributed to the Mammalian Gene Collection, producing some of the first full-length sequences for many human genes. He also was part of a small team of bioinformaticians that helped sequence and release the first whole genome reference sequence for the severe acute respiratory syndrome (SARS) virus at the height of the 2003 epidemic. He has contributed to the identification of molecular markers at the DNA, RNA and protein level that are useful for diagnosis and prognosis of thyroid, breast, lymphoma and other cancers. His lab’s research is focused on the development of informatics resources and personalized medicine strategies for cancer using genomic technologies. He is a co-creator of the CIViC resource."
#> 
#> $content[[1]]$user$country
#> $content[[1]]$user$country$id
#> [1] 214
#> 
#> $content[[1]]$user$country$iso
#> [1] "US"
#> 
#> $content[[1]]$user$country$name
#> [1] "United States"
#> 
#> 
#> $content[[1]]$user$featured_expert
#> [1] TRUE
#> 
#> $content[[1]]$user$accepted_license
#> NULL
#> 
#> $content[[1]]$user$signup_complete
#> NULL
#> 
#> $content[[1]]$user$affiliation
#> [1] ""
#> 
#> $content[[1]]$user$organization
#> $content[[1]]$user$organization$id
#> [1] 1
#> 
#> $content[[1]]$user$organization$name
#> [1] "The McDonnell Genome Institute"
#> 
#> $content[[1]]$user$organization$url
#> [1] "http://genome.wustl.edu/"
#> 
#> $content[[1]]$user$organization$description
#> [1] "The McDonnell Genome Institute (MGI) is a world leader in the fast-paced, constantly changing field of genomics. A truly unique institution, we are pushing the limits of academic research by creating, testing, and implementing new approaches to the study of biology with the goal of understanding human health and disease, as well as evolution and the biology of other organisms."
#> 
#> $content[[1]]$user$organization$profile_image
#> $content[[1]]$user$organization$profile_image$x256
#> [1] "/system/organizations/profile_images/000/000/001/x256/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content[[1]]$user$organization$profile_image$x128
#> [1] "/system/organizations/profile_images/000/000/001/x128/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content[[1]]$user$organization$profile_image$x64
#> [1] "/system/organizations/profile_images/000/000/001/x64/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content[[1]]$user$organization$profile_image$x32
#> [1] "/system/organizations/profile_images/000/000/001/x32/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content[[1]]$user$organization$profile_image$x14
#> [1] "/system/organizations/profile_images/000/000/001/x14/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> 
#> $content[[1]]$user$organization$parent
#> named list()
#> 
#> 
#> 
#> $content[[1]]$status
#> [1] "applied"
#> 
#> $content[[1]]$created_at
#> [1] "2019-05-22T15:47:05.697Z"
#> 
#> $content[[1]]$suggested_changes
#> $content[[1]]$suggested_changes$variant_id
#> $content[[1]]$suggested_changes$variant_id[[1]]
#> [1] 306
#> 
#> $content[[1]]$suggested_changes$variant_id[[2]]
#> [1] 18
#> 
#> 
#> $content[[1]]$suggested_changes$variant_origin
#> $content[[1]]$suggested_changes$variant_origin[[1]]
#> NULL
#> 
#> $content[[1]]$suggested_changes$variant_origin[[2]]
#> [1] "Somatic Mutation"
#> 
#> 
#> 
#> $content[[1]]$diffs
#> $content[[1]]$diffs$variant
#> $content[[1]]$diffs$variant$diff
#> [1] "<div class=\"diff\"></div>"
#> 
#> $content[[1]]$diffs$variant$final
#> [1] "AMPLIFICATION"
#> 
#> 
#> $content[[1]]$diffs$variant_origin
#> $content[[1]]$diffs$variant_origin$diff
#> [1] "<div class=\"diff\">\n  <ul>\n    <li class=\"ins\"><ins>Somatic Mutation</ins></li>\n  </ul>\n</div>\n"
#> 
#> $content[[1]]$diffs$variant_origin$final
#> [1] "Somatic Mutation"
#> 
#> 
#> 
#> $content[[1]]$moderated_object
#> $content[[1]]$moderated_object$id
#> [1] 1
#> 
#> $content[[1]]$moderated_object$type
#> [1] "Assertion"
#> 
#> 
#> 
#> 
#> $url
#> [1] "https://civicdb.org/api/assertions/1/suggested_changes"
#> 
#> $response
#> Response [https://civicdb.org/api/assertions/1/suggested_changes]
#>   Date: 2019-09-03 16:44
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 3.76 kB
#> 
#> 
#> attr(,"class")
#> [1] "civic_api"