Get a specific evidence item

Retrieve a specific evidence item from the CIViC DB

getEvidenceItem(id)

Arguments

id
Internal CIViC ID of the evidence item of interest

Value

An S3 Object of type civic_api containing the content, url, and response

Examples

getEvidenceItem(id = 1)

#> $content
#> $content$id
#> [1] 1
#> 
#> $content$name
#> [1] "EID1"
#> 
#> $content$description
#> [1] "JAK2 V617F is not associated with lymphoid leukemia (B-lineage ALL, T-ALL or CLL)."
#> 
#> $content$disease
#> $content$disease$id
#> [1] 1
#> 
#> $content$disease$name
#> [1] "Lymphoid Leukemia"
#> 
#> $content$disease$display_name
#> [1] "Lymphoid Leukemia"
#> 
#> $content$disease$doid
#> [1] "10747"
#> 
#> $content$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:10747"
#> 
#> 
#> $content$drugs
#> list()
#> 
#> $content$rating
#> [1] 4
#> 
#> $content$evidence_level
#> [1] "B"
#> 
#> $content$evidence_type
#> [1] "Diagnostic"
#> 
#> $content$clinical_significance
#> [1] "Negative"
#> 
#> $content$evidence_direction
#> [1] "Supports"
#> 
#> $content$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$drug_interaction_type
#> NULL
#> 
#> $content$status
#> [1] "accepted"
#> 
#> $content$open_change_count
#> [1] 0
#> 
#> $content$type
#> [1] "evidence"
#> 
#> $content$source
#> $content$source$id
#> [1] 51
#> 
#> $content$source$name
#> [1] "The JAK2V617F activating mutation occurs in chronic myelomonocytic leukemia and acute myeloid leukemia, but not in acute lymphoblastic leukemia or chronic lymphocytic leukemia."
#> 
#> $content$source$citation
#> [1] "Levine et al., 2005, Blood"
#> 
#> $content$source$citation_id
#> [1] "16081687"
#> 
#> $content$source$source_type
#> [1] "PubMed"
#> 
#> $content$source$asco_abstract_id
#> NULL
#> 
#> $content$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/16081687"
#> 
#> $content$source$open_access
#> [1] TRUE
#> 
#> $content$source$pmc_id
#> [1] "PMC1895066"
#> 
#> $content$source$publication_date
#> $content$source$publication_date$year
#> [1] 2005
#> 
#> $content$source$publication_date$month
#> [1] 11
#> 
#> $content$source$publication_date$day
#> [1] 15
#> 
#> 
#> $content$source$journal
#> [1] "Blood"
#> 
#> $content$source$full_journal_title
#> [1] "Blood"
#> 
#> $content$source$status
#> [1] "fully curated"
#> 
#> $content$source$is_review
#> [1] FALSE
#> 
#> $content$source$clinical_trials
#> list()
#> 
#> 
#> $content$variant_id
#> [1] 64
#> 
#> $content$phenotypes
#> list()
#> 
#> $content$assertions
#> list()
#> 
#> $content$errors
#> named list()
#> 
#> $content$lifecycle_actions
#> $content$lifecycle_actions$submitted
#> $content$lifecycle_actions$submitted$timestamp
#> [1] "2015-06-21T16:49:38.790Z"
#> 
#> $content$lifecycle_actions$submitted$user
#> $content$lifecycle_actions$submitted$user$id
#> [1] 41
#> 
#> $content$lifecycle_actions$submitted$user$name
#> [1] "Nick Spies"
#> 
#> $content$lifecycle_actions$submitted$user$last_seen_at
#> [1] "2019-01-20T03:51:36.969Z"
#> 
#> $content$lifecycle_actions$submitted$user$username
#> [1] "NickSpies"
#> 
#> $content$lifecycle_actions$submitted$user$role
#> [1] "admin"
#> 
#> $content$lifecycle_actions$submitted$user$avatar_url
#> [1] "https://secure.gravatar.com/avatar/3376aeb8439c5ab3e5d72fa2eeed39e5.png?d=identicon&r=pg&s=32"
#> 
#> $content$lifecycle_actions$submitted$user$avatars
#> $content$lifecycle_actions$submitted$user$avatars$x128
#> [1] "https://secure.gravatar.com/avatar/3376aeb8439c5ab3e5d72fa2eeed39e5.png?d=identicon&r=pg&s=128"
#> 
#> $content$lifecycle_actions$submitted$user$avatars$x64
#> [1] "https://secure.gravatar.com/avatar/3376aeb8439c5ab3e5d72fa2eeed39e5.png?d=identicon&r=pg&s=64"
#> 
#> $content$lifecycle_actions$submitted$user$avatars$x32
#> [1] "https://secure.gravatar.com/avatar/3376aeb8439c5ab3e5d72fa2eeed39e5.png?d=identicon&r=pg&s=32"
#> 
#> $content$lifecycle_actions$submitted$user$avatars$x14
#> [1] "https://secure.gravatar.com/avatar/3376aeb8439c5ab3e5d72fa2eeed39e5.png?d=identicon&r=pg&s=14"
#> 
#> 
#> $content$lifecycle_actions$submitted$user$area_of_expertise
#> [1] "Research Scientist"
#> 
#> $content$lifecycle_actions$submitted$user$orcid
#> [1] ""
#> 
#> $content$lifecycle_actions$submitted$user$display_name
#> [1] "NickSpies"
#> 
#> $content$lifecycle_actions$submitted$user$created_at
#> [1] "2015-06-12T18:13:16.508Z"
#> 
#> $content$lifecycle_actions$submitted$user$url
#> [1] ""
#> 
#> $content$lifecycle_actions$submitted$user$twitter_handle
#> [1] "@NickSpies13"
#> 
#> $content$lifecycle_actions$submitted$user$facebook_profile
#> [1] ""
#> 
#> $content$lifecycle_actions$submitted$user$linkedin_profile
#> [1] ""
#> 
#> $content$lifecycle_actions$submitted$user$bio
#> [1] "Nick Spies is a staff analyst at the McDonnell Genome Institute and an MD student at Washington University School of Medicine. He has made substantial contributions to the development of genome analysis tools and resources at the Genome Institute including the Drug-Gene Interaction Database. He is a founding member of the CIViC team, helping to define the CIViC data model, and a leading content curator and a feature development consultant."
#> 
#> $content$lifecycle_actions$submitted$user$country
#> $content$lifecycle_actions$submitted$user$country$id
#> [1] 214
#> 
#> $content$lifecycle_actions$submitted$user$country$iso
#> [1] "US"
#> 
#> $content$lifecycle_actions$submitted$user$country$name
#> [1] "United States"
#> 
#> 
#> $content$lifecycle_actions$submitted$user$featured_expert
#> [1] TRUE
#> 
#> $content$lifecycle_actions$submitted$user$accepted_license
#> NULL
#> 
#> $content$lifecycle_actions$submitted$user$signup_complete
#> NULL
#> 
#> $content$lifecycle_actions$submitted$user$affiliation
#> [1] ""
#> 
#> $content$lifecycle_actions$submitted$user$organization
#> $content$lifecycle_actions$submitted$user$organization$id
#> [1] 1
#> 
#> $content$lifecycle_actions$submitted$user$organization$name
#> [1] "The McDonnell Genome Institute"
#> 
#> $content$lifecycle_actions$submitted$user$organization$url
#> [1] "http://genome.wustl.edu/"
#> 
#> $content$lifecycle_actions$submitted$user$organization$description
#> [1] "The McDonnell Genome Institute (MGI) is a world leader in the fast-paced, constantly changing field of genomics. A truly unique institution, we are pushing the limits of academic research by creating, testing, and implementing new approaches to the study of biology with the goal of understanding human health and disease, as well as evolution and the biology of other organisms."
#> 
#> $content$lifecycle_actions$submitted$user$organization$profile_image
#> $content$lifecycle_actions$submitted$user$organization$profile_image$x256
#> [1] "/system/organizations/profile_images/000/000/001/x256/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$submitted$user$organization$profile_image$x128
#> [1] "/system/organizations/profile_images/000/000/001/x128/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$submitted$user$organization$profile_image$x64
#> [1] "/system/organizations/profile_images/000/000/001/x64/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$submitted$user$organization$profile_image$x32
#> [1] "/system/organizations/profile_images/000/000/001/x32/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$submitted$user$organization$profile_image$x14
#> [1] "/system/organizations/profile_images/000/000/001/x14/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> 
#> $content$lifecycle_actions$submitted$user$organization$parent
#> named list()
#> 
#> 
#> 
#> 
#> $content$lifecycle_actions$accepted
#> $content$lifecycle_actions$accepted$timestamp
#> [1] "2015-06-21T16:49:38.790Z"
#> 
#> $content$lifecycle_actions$accepted$user
#> $content$lifecycle_actions$accepted$user$id
#> [1] 6
#> 
#> $content$lifecycle_actions$accepted$user$name
#> [1] "Kilannin Krysiak"
#> 
#> $content$lifecycle_actions$accepted$user$last_seen_at
#> [1] "2019-09-03T16:28:00.554Z"
#> 
#> $content$lifecycle_actions$accepted$user$username
#> [1] "kkrysiak"
#> 
#> $content$lifecycle_actions$accepted$user$role
#> [1] "admin"
#> 
#> $content$lifecycle_actions$accepted$user$avatar_url
#> [1] "https://secure.gravatar.com/avatar/17180f9afc9f7f04fff97197c1ee5cb6.png?d=identicon&r=pg&s=32"
#> 
#> $content$lifecycle_actions$accepted$user$avatars
#> $content$lifecycle_actions$accepted$user$avatars$x128
#> [1] "https://secure.gravatar.com/avatar/17180f9afc9f7f04fff97197c1ee5cb6.png?d=identicon&r=pg&s=128"
#> 
#> $content$lifecycle_actions$accepted$user$avatars$x64
#> [1] "https://secure.gravatar.com/avatar/17180f9afc9f7f04fff97197c1ee5cb6.png?d=identicon&r=pg&s=64"
#> 
#> $content$lifecycle_actions$accepted$user$avatars$x32
#> [1] "https://secure.gravatar.com/avatar/17180f9afc9f7f04fff97197c1ee5cb6.png?d=identicon&r=pg&s=32"
#> 
#> $content$lifecycle_actions$accepted$user$avatars$x14
#> [1] "https://secure.gravatar.com/avatar/17180f9afc9f7f04fff97197c1ee5cb6.png?d=identicon&r=pg&s=14"
#> 
#> 
#> $content$lifecycle_actions$accepted$user$area_of_expertise
#> [1] "Research Scientist"
#> 
#> $content$lifecycle_actions$accepted$user$orcid
#> [1] "0000-0002-6299-9230"
#> 
#> $content$lifecycle_actions$accepted$user$display_name
#> [1] "kkrysiak"
#> 
#> $content$lifecycle_actions$accepted$user$created_at
#> [1] "2015-02-26T04:14:20.953Z"
#> 
#> $content$lifecycle_actions$accepted$user$url
#> [1] ""
#> 
#> $content$lifecycle_actions$accepted$user$twitter_handle
#> [1] ""
#> 
#> $content$lifecycle_actions$accepted$user$facebook_profile
#> [1] ""
#> 
#> $content$lifecycle_actions$accepted$user$linkedin_profile
#> [1] "kilannin-krysiak-69047819"
#> 
#> $content$lifecycle_actions$accepted$user$bio
#> [1] "Dr. Krysiak is an Instructor at the McDonnell Genome Institute at Washington University School of Medicine where she is involved in the comprehensive genomic analysis of cancer patient cohorts and “n-of-1” studies. She received her PhD in Molecular Genetics and Genomics at Washington University in St. Louis where she focused on the genetics of myelodysplastic syndrome through advanced flow cytometry techniques, primary cell culture and mouse models. She is a founding member of the CIViC team, helping to define the CIViC data model, and a leading content curator and feature development consultant."
#> 
#> $content$lifecycle_actions$accepted$user$country
#> $content$lifecycle_actions$accepted$user$country$id
#> [1] 214
#> 
#> $content$lifecycle_actions$accepted$user$country$iso
#> [1] "US"
#> 
#> $content$lifecycle_actions$accepted$user$country$name
#> [1] "United States"
#> 
#> 
#> $content$lifecycle_actions$accepted$user$featured_expert
#> [1] TRUE
#> 
#> $content$lifecycle_actions$accepted$user$accepted_license
#> NULL
#> 
#> $content$lifecycle_actions$accepted$user$signup_complete
#> NULL
#> 
#> $content$lifecycle_actions$accepted$user$affiliation
#> [1] ""
#> 
#> $content$lifecycle_actions$accepted$user$organization
#> $content$lifecycle_actions$accepted$user$organization$id
#> [1] 1
#> 
#> $content$lifecycle_actions$accepted$user$organization$name
#> [1] "The McDonnell Genome Institute"
#> 
#> $content$lifecycle_actions$accepted$user$organization$url
#> [1] "http://genome.wustl.edu/"
#> 
#> $content$lifecycle_actions$accepted$user$organization$description
#> [1] "The McDonnell Genome Institute (MGI) is a world leader in the fast-paced, constantly changing field of genomics. A truly unique institution, we are pushing the limits of academic research by creating, testing, and implementing new approaches to the study of biology with the goal of understanding human health and disease, as well as evolution and the biology of other organisms."
#> 
#> $content$lifecycle_actions$accepted$user$organization$profile_image
#> $content$lifecycle_actions$accepted$user$organization$profile_image$x256
#> [1] "/system/organizations/profile_images/000/000/001/x256/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$accepted$user$organization$profile_image$x128
#> [1] "/system/organizations/profile_images/000/000/001/x128/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$accepted$user$organization$profile_image$x64
#> [1] "/system/organizations/profile_images/000/000/001/x64/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$accepted$user$organization$profile_image$x32
#> [1] "/system/organizations/profile_images/000/000/001/x32/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$accepted$user$organization$profile_image$x14
#> [1] "/system/organizations/profile_images/000/000/001/x14/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> 
#> $content$lifecycle_actions$accepted$user$organization$parent
#> named list()
#> 
#> 
#> 
#> 
#> 
#> $content$fields_with_pending_changes
#> named list()
#> 
#> $content$gene_id
#> [1] 28
#> 
#> 
#> $url
#> [1] "https://civicdb.org/api/evidence_items/1"
#> 
#> $response
#> Response [https://civicdb.org/api/evidence_items/1]
#>   Date: 2019-09-03 16:44
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 6.72 kB
#> 
#> 
#> attr(,"class")
#> [1] "civic_api"