Get a specific variant

Retrieve a specific variant from the CIViC DB

getVariant(id)

Arguments

id
Internal CIViC ID of the variant of interest

Value

An S3 Object of type civic_api containing the content, url, and response

Examples

getVariant(id = 1)

#> $content
#> $content$id
#> [1] 1
#> 
#> $content$entrez_name
#> [1] "ABL1"
#> 
#> $content$entrez_id
#> [1] 25
#> 
#> $content$name
#> [1] "BCR-ABL"
#> 
#> $content$description
#> [1] "The BCR-ABL fusion protein, commonly referred to as the Philadelphia chromosome, is one of the most studied fusion genes in cancer. It has widely been considered the initiating event in chronic myelogenous leukemia (CML). However, despite its ability to initiate disease in mice, its status as an initiating mutation is in dispute. In what is commonly used as the poster-child for targeted therapeutics, the development and use of imatinib in the clinic has led to profound improvements in the prognosis of the disease. However, imatinib resistance is still seen in patients with mutations in the ABL kinase domain of the fusion, most notably the T315I variant. In patients resistant to imatinib, either acquired or otherwise, second generation BCR-ABL TKI's (dasatinib and nilotinib) have seen some success in delivering a tumor response. Third generation ABL1 inhibitor ponatinib is the only FDA approved drug with activity against T315I . However due to risk of life-threatening blood clots and severe narrowing of blood vessels ponatinib is ONLY approved for T315I-positive CML or T315I-positive Ph+ ALL or in cases of CML, Ph+ ALL with resistance or intolerance to other approved ABL1 inhibitors."
#> 
#> $content$gene_id
#> [1] 4
#> 
#> $content$type
#> [1] "variant"
#> 
#> $content$variant_types
#> $content$variant_types[[1]]
#> $content$variant_types[[1]]$id
#> [1] 120
#> 
#> $content$variant_types[[1]]$name
#> [1] "transcript_fusion"
#> 
#> $content$variant_types[[1]]$display_name
#> [1] "Transcript Fusion"
#> 
#> $content$variant_types[[1]]$so_id
#> [1] "SO:0001886"
#> 
#> $content$variant_types[[1]]$description
#> [1] "A feature fusion where the deletion brings together transcript regions."
#> 
#> $content$variant_types[[1]]$url
#> [1] "http://www.sequenceontology.org/browser/current_svn/term/SO:0001886"
#> 
#> 
#> 
#> $content$civic_actionability_score
#> [1] 238
#> 
#> $content$coordinates
#> $content$coordinates$chromosome
#> [1] "22"
#> 
#> $content$coordinates$start
#> [1] 23522397
#> 
#> $content$coordinates$stop
#> [1] 23632600
#> 
#> $content$coordinates$reference_bases
#> NULL
#> 
#> $content$coordinates$variant_bases
#> NULL
#> 
#> $content$coordinates$representative_transcript
#> [1] "ENST00000305877.8"
#> 
#> $content$coordinates$chromosome2
#> [1] "9"
#> 
#> $content$coordinates$start2
#> [1] 133729451
#> 
#> $content$coordinates$stop2
#> [1] 133763063
#> 
#> $content$coordinates$representative_transcript2
#> [1] "ENST00000318560.5"
#> 
#> $content$coordinates$ensembl_version
#> [1] 75
#> 
#> $content$coordinates$reference_build
#> [1] "GRCh37"
#> 
#> 
#> $content$evidence_items
#> $content$evidence_items[[1]]
#> $content$evidence_items[[1]]$id
#> [1] 344
#> 
#> $content$evidence_items[[1]]$name
#> [1] "EID344"
#> 
#> $content$evidence_items[[1]]$description
#> [1] "BCR-ABL inhibitors such as imatinib have lead to significantly improved prognosis, response rate, overall survival, and patient outcome in CML patients compared to previous therapeutic regimens."
#> 
#> $content$evidence_items[[1]]$disease
#> $content$evidence_items[[1]]$disease$id
#> [1] 4
#> 
#> $content$evidence_items[[1]]$disease$name
#> [1] "Chronic Myeloid Leukemia"
#> 
#> $content$evidence_items[[1]]$disease$display_name
#> [1] "Chronic Myeloid Leukemia"
#> 
#> $content$evidence_items[[1]]$disease$doid
#> [1] "8552"
#> 
#> $content$evidence_items[[1]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:8552"
#> 
#> 
#> $content$evidence_items[[1]]$drugs
#> $content$evidence_items[[1]]$drugs[[1]]
#> $content$evidence_items[[1]]$drugs[[1]]$id
#> [1] 5
#> 
#> $content$evidence_items[[1]]$drugs[[1]]$name
#> [1] "Imatinib"
#> 
#> $content$evidence_items[[1]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$evidence_items[[1]]$rating
#> [1] 5
#> 
#> $content$evidence_items[[1]]$evidence_level
#> [1] "A"
#> 
#> $content$evidence_items[[1]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$evidence_items[[1]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$evidence_items[[1]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$evidence_items[[1]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$evidence_items[[1]]$drug_interaction_type
#> NULL
#> 
#> $content$evidence_items[[1]]$status
#> [1] "accepted"
#> 
#> $content$evidence_items[[1]]$open_change_count
#> [1] 0
#> 
#> $content$evidence_items[[1]]$type
#> [1] "evidence"
#> 
#> $content$evidence_items[[1]]$source
#> $content$evidence_items[[1]]$source$id
#> [1] 156
#> 
#> $content$evidence_items[[1]]$source$name
#> [1] "BCR-ABL tyrosine kinase inhibitors in the treatment of Philadelphia chromosome positive chronic myeloid leukemia: a review."
#> 
#> $content$evidence_items[[1]]$source$citation
#> [1] "An et al., 2010, Leuk. Res."
#> 
#> $content$evidence_items[[1]]$source$citation_id
#> [1] "20537386"
#> 
#> $content$evidence_items[[1]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$evidence_items[[1]]$source$asco_abstract_id
#> NULL
#> 
#> $content$evidence_items[[1]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/20537386"
#> 
#> $content$evidence_items[[1]]$source$open_access
#> NULL
#> 
#> $content$evidence_items[[1]]$source$pmc_id
#> NULL
#> 
#> $content$evidence_items[[1]]$source$publication_date
#> $content$evidence_items[[1]]$source$publication_date$year
#> [1] 2010
#> 
#> $content$evidence_items[[1]]$source$publication_date$month
#> [1] 10
#> 
#> 
#> $content$evidence_items[[1]]$source$journal
#> [1] "Leuk. Res."
#> 
#> $content$evidence_items[[1]]$source$full_journal_title
#> [1] "Leukemia research"
#> 
#> $content$evidence_items[[1]]$source$status
#> [1] "fully curated"
#> 
#> $content$evidence_items[[1]]$source$is_review
#> [1] FALSE
#> 
#> $content$evidence_items[[1]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$evidence_items[[1]]$variant_id
#> [1] 1
#> 
#> $content$evidence_items[[1]]$phenotypes
#> list()
#> 
#> 
#> $content$evidence_items[[2]]
#> $content$evidence_items[[2]]$id
#> [1] 259
#> 
#> $content$evidence_items[[2]]$name
#> [1] "EID259"
#> 
#> $content$evidence_items[[2]]$description
#> [1] "Treatment of Philadelphia chromosome positive leukemias with imatinib results in high rates of complete remission in patients with CML."
#> 
#> $content$evidence_items[[2]]$disease
#> $content$evidence_items[[2]]$disease$id
#> [1] 3
#> 
#> $content$evidence_items[[2]]$disease$name
#> [1] "Acute Myeloid Leukemia"
#> 
#> $content$evidence_items[[2]]$disease$display_name
#> [1] "Acute Myeloid Leukemia"
#> 
#> $content$evidence_items[[2]]$disease$doid
#> [1] "9119"
#> 
#> $content$evidence_items[[2]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:9119"
#> 
#> 
#> $content$evidence_items[[2]]$drugs
#> $content$evidence_items[[2]]$drugs[[1]]
#> $content$evidence_items[[2]]$drugs[[1]]$id
#> [1] 5
#> 
#> $content$evidence_items[[2]]$drugs[[1]]$name
#> [1] "Imatinib"
#> 
#> $content$evidence_items[[2]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$evidence_items[[2]]$rating
#> [1] 5
#> 
#> $content$evidence_items[[2]]$evidence_level
#> [1] "A"
#> 
#> $content$evidence_items[[2]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$evidence_items[[2]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$evidence_items[[2]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$evidence_items[[2]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$evidence_items[[2]]$drug_interaction_type
#> NULL
#> 
#> $content$evidence_items[[2]]$status
#> [1] "accepted"
#> 
#> $content$evidence_items[[2]]$open_change_count
#> [1] 0
#> 
#> $content$evidence_items[[2]]$type
#> [1] "evidence"
#> 
#> $content$evidence_items[[2]]$source
#> $content$evidence_items[[2]]$source$id
#> [1] 164
#> 
#> $content$evidence_items[[2]]$source$name
#> [1] "Novel targeted therapies for Bcr-Abl positive acute leukemias: beyond STI571."
#> 
#> $content$evidence_items[[2]]$source$citation
#> [1] "Nimmanapalli et al., 2002, Oncogene"
#> 
#> $content$evidence_items[[2]]$source$citation_id
#> [1] "12476305"
#> 
#> $content$evidence_items[[2]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$evidence_items[[2]]$source$asco_abstract_id
#> NULL
#> 
#> $content$evidence_items[[2]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/12476305"
#> 
#> $content$evidence_items[[2]]$source$open_access
#> NULL
#> 
#> $content$evidence_items[[2]]$source$pmc_id
#> NULL
#> 
#> $content$evidence_items[[2]]$source$publication_date
#> $content$evidence_items[[2]]$source$publication_date$year
#> [1] 2002
#> 
#> $content$evidence_items[[2]]$source$publication_date$month
#> [1] 12
#> 
#> $content$evidence_items[[2]]$source$publication_date$day
#> [1] 9
#> 
#> 
#> $content$evidence_items[[2]]$source$journal
#> [1] "Oncogene"
#> 
#> $content$evidence_items[[2]]$source$full_journal_title
#> [1] "Oncogene"
#> 
#> $content$evidence_items[[2]]$source$status
#> [1] "fully curated"
#> 
#> $content$evidence_items[[2]]$source$is_review
#> [1] FALSE
#> 
#> $content$evidence_items[[2]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$evidence_items[[2]]$variant_id
#> [1] 1
#> 
#> $content$evidence_items[[2]]$phenotypes
#> list()
#> 
#> 
#> $content$evidence_items[[3]]
#> $content$evidence_items[[3]]$id
#> [1] 260
#> 
#> $content$evidence_items[[3]]$name
#> [1] "EID260"
#> 
#> $content$evidence_items[[3]]$description
#> [1] "The clinical use of imatinib in patients with BCR-ABL fusion has resulted in drastic prognostic improvements in patients with CML."
#> 
#> $content$evidence_items[[3]]$disease
#> $content$evidence_items[[3]]$disease$id
#> [1] 4
#> 
#> $content$evidence_items[[3]]$disease$name
#> [1] "Chronic Myeloid Leukemia"
#> 
#> $content$evidence_items[[3]]$disease$display_name
#> [1] "Chronic Myeloid Leukemia"
#> 
#> $content$evidence_items[[3]]$disease$doid
#> [1] "8552"
#> 
#> $content$evidence_items[[3]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:8552"
#> 
#> 
#> $content$evidence_items[[3]]$drugs
#> $content$evidence_items[[3]]$drugs[[1]]
#> $content$evidence_items[[3]]$drugs[[1]]$id
#> [1] 5
#> 
#> $content$evidence_items[[3]]$drugs[[1]]$name
#> [1] "Imatinib"
#> 
#> $content$evidence_items[[3]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$evidence_items[[3]]$rating
#> [1] 5
#> 
#> $content$evidence_items[[3]]$evidence_level
#> [1] "A"
#> 
#> $content$evidence_items[[3]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$evidence_items[[3]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$evidence_items[[3]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$evidence_items[[3]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$evidence_items[[3]]$drug_interaction_type
#> NULL
#> 
#> $content$evidence_items[[3]]$status
#> [1] "accepted"
#> 
#> $content$evidence_items[[3]]$open_change_count
#> [1] 0
#> 
#> $content$evidence_items[[3]]$type
#> [1] "evidence"
#> 
#> $content$evidence_items[[3]]$source
#> $content$evidence_items[[3]]$source$id
#> [1] 156
#> 
#> $content$evidence_items[[3]]$source$name
#> [1] "BCR-ABL tyrosine kinase inhibitors in the treatment of Philadelphia chromosome positive chronic myeloid leukemia: a review."
#> 
#> $content$evidence_items[[3]]$source$citation
#> [1] "An et al., 2010, Leuk. Res."
#> 
#> $content$evidence_items[[3]]$source$citation_id
#> [1] "20537386"
#> 
#> $content$evidence_items[[3]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$evidence_items[[3]]$source$asco_abstract_id
#> NULL
#> 
#> $content$evidence_items[[3]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/20537386"
#> 
#> $content$evidence_items[[3]]$source$open_access
#> NULL
#> 
#> $content$evidence_items[[3]]$source$pmc_id
#> NULL
#> 
#> $content$evidence_items[[3]]$source$publication_date
#> $content$evidence_items[[3]]$source$publication_date$year
#> [1] 2010
#> 
#> $content$evidence_items[[3]]$source$publication_date$month
#> [1] 10
#> 
#> 
#> $content$evidence_items[[3]]$source$journal
#> [1] "Leuk. Res."
#> 
#> $content$evidence_items[[3]]$source$full_journal_title
#> [1] "Leukemia research"
#> 
#> $content$evidence_items[[3]]$source$status
#> [1] "fully curated"
#> 
#> $content$evidence_items[[3]]$source$is_review
#> [1] FALSE
#> 
#> $content$evidence_items[[3]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$evidence_items[[3]]$variant_id
#> [1] 1
#> 
#> $content$evidence_items[[3]]$phenotypes
#> list()
#> 
#> 
#> $content$evidence_items[[4]]
#> $content$evidence_items[[4]]$id
#> [1] 213
#> 
#> $content$evidence_items[[4]]$name
#> [1] "EID213"
#> 
#> $content$evidence_items[[4]]$description
#> [1] "Data from deep sequencing of a Ph-negative clone of a Ph-positive CML patient has found a driving mutation in DNMT3A that preceeded the BCR-ABL fusion, and may imply the possibility that BCR-ABL is not universally the initiating event in CML."
#> 
#> $content$evidence_items[[4]]$disease
#> $content$evidence_items[[4]]$disease$id
#> [1] 4
#> 
#> $content$evidence_items[[4]]$disease$name
#> [1] "Chronic Myeloid Leukemia"
#> 
#> $content$evidence_items[[4]]$disease$display_name
#> [1] "Chronic Myeloid Leukemia"
#> 
#> $content$evidence_items[[4]]$disease$doid
#> [1] "8552"
#> 
#> $content$evidence_items[[4]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:8552"
#> 
#> 
#> $content$evidence_items[[4]]$drugs
#> list()
#> 
#> $content$evidence_items[[4]]$rating
#> [1] 2
#> 
#> $content$evidence_items[[4]]$evidence_level
#> [1] "C"
#> 
#> $content$evidence_items[[4]]$evidence_type
#> [1] "Diagnostic"
#> 
#> $content$evidence_items[[4]]$clinical_significance
#> [1] "Positive"
#> 
#> $content$evidence_items[[4]]$evidence_direction
#> [1] "Does Not Support"
#> 
#> $content$evidence_items[[4]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$evidence_items[[4]]$drug_interaction_type
#> NULL
#> 
#> $content$evidence_items[[4]]$status
#> [1] "accepted"
#> 
#> $content$evidence_items[[4]]$open_change_count
#> [1] 0
#> 
#> $content$evidence_items[[4]]$type
#> [1] "evidence"
#> 
#> $content$evidence_items[[4]]$source
#> $content$evidence_items[[4]]$source$id
#> [1] 151
#> 
#> $content$evidence_items[[4]]$source$name
#> [1] "Molecular-defined clonal evolution in patients with chronic myeloid leukemia independent of the BCR-ABL status."
#> 
#> $content$evidence_items[[4]]$source$citation
#> [1] "Schmidt et al., 2014, Leukemia"
#> 
#> $content$evidence_items[[4]]$source$citation_id
#> [1] "25212276"
#> 
#> $content$evidence_items[[4]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$evidence_items[[4]]$source$asco_abstract_id
#> NULL
#> 
#> $content$evidence_items[[4]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/25212276"
#> 
#> $content$evidence_items[[4]]$source$open_access
#> NULL
#> 
#> $content$evidence_items[[4]]$source$pmc_id
#> NULL
#> 
#> $content$evidence_items[[4]]$source$publication_date
#> $content$evidence_items[[4]]$source$publication_date$year
#> [1] 2014
#> 
#> $content$evidence_items[[4]]$source$publication_date$month
#> [1] 12
#> 
#> 
#> $content$evidence_items[[4]]$source$journal
#> [1] "Leukemia"
#> 
#> $content$evidence_items[[4]]$source$full_journal_title
#> [1] "Leukemia"
#> 
#> $content$evidence_items[[4]]$source$status
#> [1] "fully curated"
#> 
#> $content$evidence_items[[4]]$source$is_review
#> [1] FALSE
#> 
#> $content$evidence_items[[4]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$evidence_items[[4]]$variant_id
#> [1] 1
#> 
#> $content$evidence_items[[4]]$phenotypes
#> list()
#> 
#> 
#> $content$evidence_items[[5]]
#> $content$evidence_items[[5]]$id
#> [1] 232
#> 
#> $content$evidence_items[[5]]$name
#> [1] "EID232"
#> 
#> $content$evidence_items[[5]]$description
#> [1] "BCR-ABL fusions have been found to contribute to imatinib resistance in AML cell lines."
#> 
#> $content$evidence_items[[5]]$disease
#> $content$evidence_items[[5]]$disease$id
#> [1] 3
#> 
#> $content$evidence_items[[5]]$disease$name
#> [1] "Acute Myeloid Leukemia"
#> 
#> $content$evidence_items[[5]]$disease$display_name
#> [1] "Acute Myeloid Leukemia"
#> 
#> $content$evidence_items[[5]]$disease$doid
#> [1] "9119"
#> 
#> $content$evidence_items[[5]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:9119"
#> 
#> 
#> $content$evidence_items[[5]]$drugs
#> $content$evidence_items[[5]]$drugs[[1]]
#> $content$evidence_items[[5]]$drugs[[1]]$id
#> [1] 5
#> 
#> $content$evidence_items[[5]]$drugs[[1]]$name
#> [1] "Imatinib"
#> 
#> $content$evidence_items[[5]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$evidence_items[[5]]$rating
#> [1] 3
#> 
#> $content$evidence_items[[5]]$evidence_level
#> [1] "D"
#> 
#> $content$evidence_items[[5]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$evidence_items[[5]]$clinical_significance
#> [1] "Resistance"
#> 
#> $content$evidence_items[[5]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$evidence_items[[5]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$evidence_items[[5]]$drug_interaction_type
#> NULL
#> 
#> $content$evidence_items[[5]]$status
#> [1] "accepted"
#> 
#> $content$evidence_items[[5]]$open_change_count
#> [1] 0
#> 
#> $content$evidence_items[[5]]$type
#> [1] "evidence"
#> 
#> $content$evidence_items[[5]]$source
#> $content$evidence_items[[5]]$source$id
#> [1] 164
#> 
#> $content$evidence_items[[5]]$source$name
#> [1] "Novel targeted therapies for Bcr-Abl positive acute leukemias: beyond STI571."
#> 
#> $content$evidence_items[[5]]$source$citation
#> [1] "Nimmanapalli et al., 2002, Oncogene"
#> 
#> $content$evidence_items[[5]]$source$citation_id
#> [1] "12476305"
#> 
#> $content$evidence_items[[5]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$evidence_items[[5]]$source$asco_abstract_id
#> NULL
#> 
#> $content$evidence_items[[5]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/12476305"
#> 
#> $content$evidence_items[[5]]$source$open_access
#> NULL
#> 
#> $content$evidence_items[[5]]$source$pmc_id
#> NULL
#> 
#> $content$evidence_items[[5]]$source$publication_date
#> $content$evidence_items[[5]]$source$publication_date$year
#> [1] 2002
#> 
#> $content$evidence_items[[5]]$source$publication_date$month
#> [1] 12
#> 
#> $content$evidence_items[[5]]$source$publication_date$day
#> [1] 9
#> 
#> 
#> $content$evidence_items[[5]]$source$journal
#> [1] "Oncogene"
#> 
#> $content$evidence_items[[5]]$source$full_journal_title
#> [1] "Oncogene"
#> 
#> $content$evidence_items[[5]]$source$status
#> [1] "fully curated"
#> 
#> $content$evidence_items[[5]]$source$is_review
#> [1] FALSE
#> 
#> $content$evidence_items[[5]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$evidence_items[[5]]$variant_id
#> [1] 1
#> 
#> $content$evidence_items[[5]]$phenotypes
#> list()
#> 
#> 
#> $content$evidence_items[[6]]
#> $content$evidence_items[[6]]$id
#> [1] 220
#> 
#> $content$evidence_items[[6]]$name
#> [1] "EID220"
#> 
#> $content$evidence_items[[6]]$description
#> [1] "The presence of BCR-ABL fusion is considered the characterizing feature of chronic myeloid leukemia, and has been widely thought of as the initiating event in the disease."
#> 
#> $content$evidence_items[[6]]$disease
#> $content$evidence_items[[6]]$disease$id
#> [1] 4
#> 
#> $content$evidence_items[[6]]$disease$name
#> [1] "Chronic Myeloid Leukemia"
#> 
#> $content$evidence_items[[6]]$disease$display_name
#> [1] "Chronic Myeloid Leukemia"
#> 
#> $content$evidence_items[[6]]$disease$doid
#> [1] "8552"
#> 
#> $content$evidence_items[[6]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:8552"
#> 
#> 
#> $content$evidence_items[[6]]$drugs
#> list()
#> 
#> $content$evidence_items[[6]]$rating
#> [1] 4
#> 
#> $content$evidence_items[[6]]$evidence_level
#> [1] "A"
#> 
#> $content$evidence_items[[6]]$evidence_type
#> [1] "Diagnostic"
#> 
#> $content$evidence_items[[6]]$clinical_significance
#> [1] "Positive"
#> 
#> $content$evidence_items[[6]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$evidence_items[[6]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$evidence_items[[6]]$drug_interaction_type
#> NULL
#> 
#> $content$evidence_items[[6]]$status
#> [1] "accepted"
#> 
#> $content$evidence_items[[6]]$open_change_count
#> [1] 0
#> 
#> $content$evidence_items[[6]]$type
#> [1] "evidence"
#> 
#> $content$evidence_items[[6]]$source
#> $content$evidence_items[[6]]$source$id
#> [1] 156
#> 
#> $content$evidence_items[[6]]$source$name
#> [1] "BCR-ABL tyrosine kinase inhibitors in the treatment of Philadelphia chromosome positive chronic myeloid leukemia: a review."
#> 
#> $content$evidence_items[[6]]$source$citation
#> [1] "An et al., 2010, Leuk. Res."
#> 
#> $content$evidence_items[[6]]$source$citation_id
#> [1] "20537386"
#> 
#> $content$evidence_items[[6]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$evidence_items[[6]]$source$asco_abstract_id
#> NULL
#> 
#> $content$evidence_items[[6]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/20537386"
#> 
#> $content$evidence_items[[6]]$source$open_access
#> NULL
#> 
#> $content$evidence_items[[6]]$source$pmc_id
#> NULL
#> 
#> $content$evidence_items[[6]]$source$publication_date
#> $content$evidence_items[[6]]$source$publication_date$year
#> [1] 2010
#> 
#> $content$evidence_items[[6]]$source$publication_date$month
#> [1] 10
#> 
#> 
#> $content$evidence_items[[6]]$source$journal
#> [1] "Leuk. Res."
#> 
#> $content$evidence_items[[6]]$source$full_journal_title
#> [1] "Leukemia research"
#> 
#> $content$evidence_items[[6]]$source$status
#> [1] "fully curated"
#> 
#> $content$evidence_items[[6]]$source$is_review
#> [1] FALSE
#> 
#> $content$evidence_items[[6]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$evidence_items[[6]]$variant_id
#> [1] 1
#> 
#> $content$evidence_items[[6]]$phenotypes
#> list()
#> 
#> 
#> $content$evidence_items[[7]]
#> $content$evidence_items[[7]]$id
#> [1] 261
#> 
#> $content$evidence_items[[7]]$name
#> [1] "EID261"
#> 
#> $content$evidence_items[[7]]$description
#> [1] "The use of nilotinib and dasatinib, second-generation BCR-ABL targeted therapies, has been effective in patients resistant to imatinib, leading to their FDA-approval."
#> 
#> $content$evidence_items[[7]]$disease
#> $content$evidence_items[[7]]$disease$id
#> [1] 4
#> 
#> $content$evidence_items[[7]]$disease$name
#> [1] "Chronic Myeloid Leukemia"
#> 
#> $content$evidence_items[[7]]$disease$display_name
#> [1] "Chronic Myeloid Leukemia"
#> 
#> $content$evidence_items[[7]]$disease$doid
#> [1] "8552"
#> 
#> $content$evidence_items[[7]]$disease$url
#> [1] "http://www.disease-ontology.org/?id=DOID:8552"
#> 
#> 
#> $content$evidence_items[[7]]$drugs
#> $content$evidence_items[[7]]$drugs[[1]]
#> $content$evidence_items[[7]]$drugs[[1]]$id
#> [1] 20
#> 
#> $content$evidence_items[[7]]$drugs[[1]]$name
#> [1] "Dasatinib"
#> 
#> $content$evidence_items[[7]]$drugs[[1]]$pubchem_id
#> NULL
#> 
#> 
#> $content$evidence_items[[7]]$drugs[[2]]
#> $content$evidence_items[[7]]$drugs[[2]]$id
#> [1] 13
#> 
#> $content$evidence_items[[7]]$drugs[[2]]$name
#> [1] "Nilotinib"
#> 
#> $content$evidence_items[[7]]$drugs[[2]]$pubchem_id
#> NULL
#> 
#> 
#> 
#> $content$evidence_items[[7]]$rating
#> [1] 4
#> 
#> $content$evidence_items[[7]]$evidence_level
#> [1] "A"
#> 
#> $content$evidence_items[[7]]$evidence_type
#> [1] "Predictive"
#> 
#> $content$evidence_items[[7]]$clinical_significance
#> [1] "Sensitivity/Response"
#> 
#> $content$evidence_items[[7]]$evidence_direction
#> [1] "Supports"
#> 
#> $content$evidence_items[[7]]$variant_origin
#> [1] "Somatic Mutation"
#> 
#> $content$evidence_items[[7]]$drug_interaction_type
#> [1] "Substitutes"
#> 
#> $content$evidence_items[[7]]$status
#> [1] "accepted"
#> 
#> $content$evidence_items[[7]]$open_change_count
#> [1] 0
#> 
#> $content$evidence_items[[7]]$type
#> [1] "evidence"
#> 
#> $content$evidence_items[[7]]$source
#> $content$evidence_items[[7]]$source$id
#> [1] 156
#> 
#> $content$evidence_items[[7]]$source$name
#> [1] "BCR-ABL tyrosine kinase inhibitors in the treatment of Philadelphia chromosome positive chronic myeloid leukemia: a review."
#> 
#> $content$evidence_items[[7]]$source$citation
#> [1] "An et al., 2010, Leuk. Res."
#> 
#> $content$evidence_items[[7]]$source$citation_id
#> [1] "20537386"
#> 
#> $content$evidence_items[[7]]$source$source_type
#> [1] "PubMed"
#> 
#> $content$evidence_items[[7]]$source$asco_abstract_id
#> NULL
#> 
#> $content$evidence_items[[7]]$source$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/20537386"
#> 
#> $content$evidence_items[[7]]$source$open_access
#> NULL
#> 
#> $content$evidence_items[[7]]$source$pmc_id
#> NULL
#> 
#> $content$evidence_items[[7]]$source$publication_date
#> $content$evidence_items[[7]]$source$publication_date$year
#> [1] 2010
#> 
#> $content$evidence_items[[7]]$source$publication_date$month
#> [1] 10
#> 
#> 
#> $content$evidence_items[[7]]$source$journal
#> [1] "Leuk. Res."
#> 
#> $content$evidence_items[[7]]$source$full_journal_title
#> [1] "Leukemia research"
#> 
#> $content$evidence_items[[7]]$source$status
#> [1] "fully curated"
#> 
#> $content$evidence_items[[7]]$source$is_review
#> [1] FALSE
#> 
#> $content$evidence_items[[7]]$source$clinical_trials
#> list()
#> 
#> 
#> $content$evidence_items[[7]]$variant_id
#> [1] 1
#> 
#> $content$evidence_items[[7]]$phenotypes
#> list()
#> 
#> 
#> 
#> $content$variant_groups
#> list()
#> 
#> $content$assertions
#> list()
#> 
#> $content$variant_aliases
#> $content$variant_aliases[[1]]
#> [1] "T(9;22)(Q34;Q11)"
#> 
#> $content$variant_aliases[[2]]
#> [1] "BCR-ABL1"
#> 
#> 
#> $content$hgvs_expressions
#> list()
#> 
#> $content$clinvar_entries
#> $content$clinvar_entries[[1]]
#> [1] "N/A"
#> 
#> 
#> $content$lifecycle_actions
#> $content$lifecycle_actions$last_modified
#> $content$lifecycle_actions$last_modified$timestamp
#> [1] "2018-10-18T03:10:36.678Z"
#> 
#> $content$lifecycle_actions$last_modified$user
#> $content$lifecycle_actions$last_modified$user$id
#> [1] 706
#> 
#> $content$lifecycle_actions$last_modified$user$name
#> [1] "Shahil Pema"
#> 
#> $content$lifecycle_actions$last_modified$user$last_seen_at
#> [1] "2019-04-10T21:03:01.137Z"
#> 
#> $content$lifecycle_actions$last_modified$user$username
#> [1] "ShahilPema"
#> 
#> $content$lifecycle_actions$last_modified$user$role
#> [1] "curator"
#> 
#> $content$lifecycle_actions$last_modified$user$avatar_url
#> [1] "https://secure.gravatar.com/avatar/ae848c0ef9575f901e4aa5dbf4a457a2.png?d=identicon&r=pg&s=32"
#> 
#> $content$lifecycle_actions$last_modified$user$avatars
#> $content$lifecycle_actions$last_modified$user$avatars$x128
#> [1] "https://secure.gravatar.com/avatar/ae848c0ef9575f901e4aa5dbf4a457a2.png?d=identicon&r=pg&s=128"
#> 
#> $content$lifecycle_actions$last_modified$user$avatars$x64
#> [1] "https://secure.gravatar.com/avatar/ae848c0ef9575f901e4aa5dbf4a457a2.png?d=identicon&r=pg&s=64"
#> 
#> $content$lifecycle_actions$last_modified$user$avatars$x32
#> [1] "https://secure.gravatar.com/avatar/ae848c0ef9575f901e4aa5dbf4a457a2.png?d=identicon&r=pg&s=32"
#> 
#> $content$lifecycle_actions$last_modified$user$avatars$x14
#> [1] "https://secure.gravatar.com/avatar/ae848c0ef9575f901e4aa5dbf4a457a2.png?d=identicon&r=pg&s=14"
#> 
#> 
#> $content$lifecycle_actions$last_modified$user$area_of_expertise
#> NULL
#> 
#> $content$lifecycle_actions$last_modified$user$orcid
#> [1] ""
#> 
#> $content$lifecycle_actions$last_modified$user$display_name
#> [1] "ShahilPema"
#> 
#> $content$lifecycle_actions$last_modified$user$created_at
#> [1] "2018-05-31T18:40:39.787Z"
#> 
#> $content$lifecycle_actions$last_modified$user$url
#> [1] ""
#> 
#> $content$lifecycle_actions$last_modified$user$twitter_handle
#> [1] ""
#> 
#> $content$lifecycle_actions$last_modified$user$facebook_profile
#> [1] ""
#> 
#> $content$lifecycle_actions$last_modified$user$linkedin_profile
#> [1] ""
#> 
#> $content$lifecycle_actions$last_modified$user$bio
#> [1] ""
#> 
#> $content$lifecycle_actions$last_modified$user$country
#> $content$lifecycle_actions$last_modified$user$country$id
#> [1] 214
#> 
#> $content$lifecycle_actions$last_modified$user$country$iso
#> [1] "US"
#> 
#> $content$lifecycle_actions$last_modified$user$country$name
#> [1] "United States"
#> 
#> 
#> $content$lifecycle_actions$last_modified$user$featured_expert
#> [1] FALSE
#> 
#> $content$lifecycle_actions$last_modified$user$accepted_license
#> NULL
#> 
#> $content$lifecycle_actions$last_modified$user$signup_complete
#> NULL
#> 
#> $content$lifecycle_actions$last_modified$user$affiliation
#> [1] ""
#> 
#> $content$lifecycle_actions$last_modified$user$organization
#> $content$lifecycle_actions$last_modified$user$organization$id
#> [1] 1
#> 
#> $content$lifecycle_actions$last_modified$user$organization$name
#> [1] "The McDonnell Genome Institute"
#> 
#> $content$lifecycle_actions$last_modified$user$organization$url
#> [1] "http://genome.wustl.edu/"
#> 
#> $content$lifecycle_actions$last_modified$user$organization$description
#> [1] "The McDonnell Genome Institute (MGI) is a world leader in the fast-paced, constantly changing field of genomics. A truly unique institution, we are pushing the limits of academic research by creating, testing, and implementing new approaches to the study of biology with the goal of understanding human health and disease, as well as evolution and the biology of other organisms."
#> 
#> $content$lifecycle_actions$last_modified$user$organization$profile_image
#> $content$lifecycle_actions$last_modified$user$organization$profile_image$x256
#> [1] "/system/organizations/profile_images/000/000/001/x256/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$last_modified$user$organization$profile_image$x128
#> [1] "/system/organizations/profile_images/000/000/001/x128/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$last_modified$user$organization$profile_image$x64
#> [1] "/system/organizations/profile_images/000/000/001/x64/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$last_modified$user$organization$profile_image$x32
#> [1] "/system/organizations/profile_images/000/000/001/x32/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$last_modified$user$organization$profile_image$x14
#> [1] "/system/organizations/profile_images/000/000/001/x14/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> 
#> $content$lifecycle_actions$last_modified$user$organization$parent
#> named list()
#> 
#> 
#> 
#> 
#> $content$lifecycle_actions$last_reviewed
#> $content$lifecycle_actions$last_reviewed$timestamp
#> [1] "2019-03-12T00:22:25.589Z"
#> 
#> $content$lifecycle_actions$last_reviewed$user
#> $content$lifecycle_actions$last_reviewed$user$id
#> [1] 6
#> 
#> $content$lifecycle_actions$last_reviewed$user$name
#> [1] "Kilannin Krysiak"
#> 
#> $content$lifecycle_actions$last_reviewed$user$last_seen_at
#> [1] "2019-09-03T16:28:00.554Z"
#> 
#> $content$lifecycle_actions$last_reviewed$user$username
#> [1] "kkrysiak"
#> 
#> $content$lifecycle_actions$last_reviewed$user$role
#> [1] "admin"
#> 
#> $content$lifecycle_actions$last_reviewed$user$avatar_url
#> [1] "https://secure.gravatar.com/avatar/17180f9afc9f7f04fff97197c1ee5cb6.png?d=identicon&r=pg&s=32"
#> 
#> $content$lifecycle_actions$last_reviewed$user$avatars
#> $content$lifecycle_actions$last_reviewed$user$avatars$x128
#> [1] "https://secure.gravatar.com/avatar/17180f9afc9f7f04fff97197c1ee5cb6.png?d=identicon&r=pg&s=128"
#> 
#> $content$lifecycle_actions$last_reviewed$user$avatars$x64
#> [1] "https://secure.gravatar.com/avatar/17180f9afc9f7f04fff97197c1ee5cb6.png?d=identicon&r=pg&s=64"
#> 
#> $content$lifecycle_actions$last_reviewed$user$avatars$x32
#> [1] "https://secure.gravatar.com/avatar/17180f9afc9f7f04fff97197c1ee5cb6.png?d=identicon&r=pg&s=32"
#> 
#> $content$lifecycle_actions$last_reviewed$user$avatars$x14
#> [1] "https://secure.gravatar.com/avatar/17180f9afc9f7f04fff97197c1ee5cb6.png?d=identicon&r=pg&s=14"
#> 
#> 
#> $content$lifecycle_actions$last_reviewed$user$area_of_expertise
#> [1] "Research Scientist"
#> 
#> $content$lifecycle_actions$last_reviewed$user$orcid
#> [1] "0000-0002-6299-9230"
#> 
#> $content$lifecycle_actions$last_reviewed$user$display_name
#> [1] "kkrysiak"
#> 
#> $content$lifecycle_actions$last_reviewed$user$created_at
#> [1] "2015-02-26T04:14:20.953Z"
#> 
#> $content$lifecycle_actions$last_reviewed$user$url
#> [1] ""
#> 
#> $content$lifecycle_actions$last_reviewed$user$twitter_handle
#> [1] ""
#> 
#> $content$lifecycle_actions$last_reviewed$user$facebook_profile
#> [1] ""
#> 
#> $content$lifecycle_actions$last_reviewed$user$linkedin_profile
#> [1] "kilannin-krysiak-69047819"
#> 
#> $content$lifecycle_actions$last_reviewed$user$bio
#> [1] "Dr. Krysiak is an Instructor at the McDonnell Genome Institute at Washington University School of Medicine where she is involved in the comprehensive genomic analysis of cancer patient cohorts and “n-of-1” studies. She received her PhD in Molecular Genetics and Genomics at Washington University in St. Louis where she focused on the genetics of myelodysplastic syndrome through advanced flow cytometry techniques, primary cell culture and mouse models. She is a founding member of the CIViC team, helping to define the CIViC data model, and a leading content curator and feature development consultant."
#> 
#> $content$lifecycle_actions$last_reviewed$user$country
#> $content$lifecycle_actions$last_reviewed$user$country$id
#> [1] 214
#> 
#> $content$lifecycle_actions$last_reviewed$user$country$iso
#> [1] "US"
#> 
#> $content$lifecycle_actions$last_reviewed$user$country$name
#> [1] "United States"
#> 
#> 
#> $content$lifecycle_actions$last_reviewed$user$featured_expert
#> [1] TRUE
#> 
#> $content$lifecycle_actions$last_reviewed$user$accepted_license
#> NULL
#> 
#> $content$lifecycle_actions$last_reviewed$user$signup_complete
#> NULL
#> 
#> $content$lifecycle_actions$last_reviewed$user$affiliation
#> [1] ""
#> 
#> $content$lifecycle_actions$last_reviewed$user$organization
#> $content$lifecycle_actions$last_reviewed$user$organization$id
#> [1] 1
#> 
#> $content$lifecycle_actions$last_reviewed$user$organization$name
#> [1] "The McDonnell Genome Institute"
#> 
#> $content$lifecycle_actions$last_reviewed$user$organization$url
#> [1] "http://genome.wustl.edu/"
#> 
#> $content$lifecycle_actions$last_reviewed$user$organization$description
#> [1] "The McDonnell Genome Institute (MGI) is a world leader in the fast-paced, constantly changing field of genomics. A truly unique institution, we are pushing the limits of academic research by creating, testing, and implementing new approaches to the study of biology with the goal of understanding human health and disease, as well as evolution and the biology of other organisms."
#> 
#> $content$lifecycle_actions$last_reviewed$user$organization$profile_image
#> $content$lifecycle_actions$last_reviewed$user$organization$profile_image$x256
#> [1] "/system/organizations/profile_images/000/000/001/x256/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$last_reviewed$user$organization$profile_image$x128
#> [1] "/system/organizations/profile_images/000/000/001/x128/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$last_reviewed$user$organization$profile_image$x64
#> [1] "/system/organizations/profile_images/000/000/001/x64/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$last_reviewed$user$organization$profile_image$x32
#> [1] "/system/organizations/profile_images/000/000/001/x32/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$last_reviewed$user$organization$profile_image$x14
#> [1] "/system/organizations/profile_images/000/000/001/x14/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> 
#> $content$lifecycle_actions$last_reviewed$user$organization$parent
#> named list()
#> 
#> 
#> 
#> 
#> $content$lifecycle_actions$last_commented_on
#> $content$lifecycle_actions$last_commented_on$timestamp
#> [1] "2019-03-12T00:22:25.589Z"
#> 
#> $content$lifecycle_actions$last_commented_on$user
#> $content$lifecycle_actions$last_commented_on$user$id
#> [1] 6
#> 
#> $content$lifecycle_actions$last_commented_on$user$name
#> [1] "Kilannin Krysiak"
#> 
#> $content$lifecycle_actions$last_commented_on$user$last_seen_at
#> [1] "2019-09-03T16:28:00.554Z"
#> 
#> $content$lifecycle_actions$last_commented_on$user$username
#> [1] "kkrysiak"
#> 
#> $content$lifecycle_actions$last_commented_on$user$role
#> [1] "admin"
#> 
#> $content$lifecycle_actions$last_commented_on$user$avatar_url
#> [1] "https://secure.gravatar.com/avatar/17180f9afc9f7f04fff97197c1ee5cb6.png?d=identicon&r=pg&s=32"
#> 
#> $content$lifecycle_actions$last_commented_on$user$avatars
#> $content$lifecycle_actions$last_commented_on$user$avatars$x128
#> [1] "https://secure.gravatar.com/avatar/17180f9afc9f7f04fff97197c1ee5cb6.png?d=identicon&r=pg&s=128"
#> 
#> $content$lifecycle_actions$last_commented_on$user$avatars$x64
#> [1] "https://secure.gravatar.com/avatar/17180f9afc9f7f04fff97197c1ee5cb6.png?d=identicon&r=pg&s=64"
#> 
#> $content$lifecycle_actions$last_commented_on$user$avatars$x32
#> [1] "https://secure.gravatar.com/avatar/17180f9afc9f7f04fff97197c1ee5cb6.png?d=identicon&r=pg&s=32"
#> 
#> $content$lifecycle_actions$last_commented_on$user$avatars$x14
#> [1] "https://secure.gravatar.com/avatar/17180f9afc9f7f04fff97197c1ee5cb6.png?d=identicon&r=pg&s=14"
#> 
#> 
#> $content$lifecycle_actions$last_commented_on$user$area_of_expertise
#> [1] "Research Scientist"
#> 
#> $content$lifecycle_actions$last_commented_on$user$orcid
#> [1] "0000-0002-6299-9230"
#> 
#> $content$lifecycle_actions$last_commented_on$user$display_name
#> [1] "kkrysiak"
#> 
#> $content$lifecycle_actions$last_commented_on$user$created_at
#> [1] "2015-02-26T04:14:20.953Z"
#> 
#> $content$lifecycle_actions$last_commented_on$user$url
#> [1] ""
#> 
#> $content$lifecycle_actions$last_commented_on$user$twitter_handle
#> [1] ""
#> 
#> $content$lifecycle_actions$last_commented_on$user$facebook_profile
#> [1] ""
#> 
#> $content$lifecycle_actions$last_commented_on$user$linkedin_profile
#> [1] "kilannin-krysiak-69047819"
#> 
#> $content$lifecycle_actions$last_commented_on$user$bio
#> [1] "Dr. Krysiak is an Instructor at the McDonnell Genome Institute at Washington University School of Medicine where she is involved in the comprehensive genomic analysis of cancer patient cohorts and “n-of-1” studies. She received her PhD in Molecular Genetics and Genomics at Washington University in St. Louis where she focused on the genetics of myelodysplastic syndrome through advanced flow cytometry techniques, primary cell culture and mouse models. She is a founding member of the CIViC team, helping to define the CIViC data model, and a leading content curator and feature development consultant."
#> 
#> $content$lifecycle_actions$last_commented_on$user$country
#> $content$lifecycle_actions$last_commented_on$user$country$id
#> [1] 214
#> 
#> $content$lifecycle_actions$last_commented_on$user$country$iso
#> [1] "US"
#> 
#> $content$lifecycle_actions$last_commented_on$user$country$name
#> [1] "United States"
#> 
#> 
#> $content$lifecycle_actions$last_commented_on$user$featured_expert
#> [1] TRUE
#> 
#> $content$lifecycle_actions$last_commented_on$user$accepted_license
#> NULL
#> 
#> $content$lifecycle_actions$last_commented_on$user$signup_complete
#> NULL
#> 
#> $content$lifecycle_actions$last_commented_on$user$affiliation
#> [1] ""
#> 
#> $content$lifecycle_actions$last_commented_on$user$organization
#> $content$lifecycle_actions$last_commented_on$user$organization$id
#> [1] 1
#> 
#> $content$lifecycle_actions$last_commented_on$user$organization$name
#> [1] "The McDonnell Genome Institute"
#> 
#> $content$lifecycle_actions$last_commented_on$user$organization$url
#> [1] "http://genome.wustl.edu/"
#> 
#> $content$lifecycle_actions$last_commented_on$user$organization$description
#> [1] "The McDonnell Genome Institute (MGI) is a world leader in the fast-paced, constantly changing field of genomics. A truly unique institution, we are pushing the limits of academic research by creating, testing, and implementing new approaches to the study of biology with the goal of understanding human health and disease, as well as evolution and the biology of other organisms."
#> 
#> $content$lifecycle_actions$last_commented_on$user$organization$profile_image
#> $content$lifecycle_actions$last_commented_on$user$organization$profile_image$x256
#> [1] "/system/organizations/profile_images/000/000/001/x256/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$last_commented_on$user$organization$profile_image$x128
#> [1] "/system/organizations/profile_images/000/000/001/x128/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$last_commented_on$user$organization$profile_image$x64
#> [1] "/system/organizations/profile_images/000/000/001/x64/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$last_commented_on$user$organization$profile_image$x32
#> [1] "/system/organizations/profile_images/000/000/001/x32/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content$lifecycle_actions$last_commented_on$user$organization$profile_image$x14
#> [1] "/system/organizations/profile_images/000/000/001/x14/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> 
#> $content$lifecycle_actions$last_commented_on$user$organization$parent
#> named list()
#> 
#> 
#> 
#> 
#> 
#> $content$sources
#> $content$sources[[1]]
#> $content$sources[[1]]$id
#> [1] 156
#> 
#> $content$sources[[1]]$name
#> [1] "BCR-ABL tyrosine kinase inhibitors in the treatment of Philadelphia chromosome positive chronic myeloid leukemia: a review."
#> 
#> $content$sources[[1]]$citation
#> [1] "An et al., 2010, Leuk. Res."
#> 
#> $content$sources[[1]]$citation_id
#> [1] "20537386"
#> 
#> $content$sources[[1]]$source_type
#> [1] "PubMed"
#> 
#> $content$sources[[1]]$asco_abstract_id
#> NULL
#> 
#> $content$sources[[1]]$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/20537386"
#> 
#> $content$sources[[1]]$open_access
#> NULL
#> 
#> $content$sources[[1]]$pmc_id
#> NULL
#> 
#> $content$sources[[1]]$publication_date
#> $content$sources[[1]]$publication_date$year
#> [1] 2010
#> 
#> $content$sources[[1]]$publication_date$month
#> [1] 10
#> 
#> 
#> $content$sources[[1]]$journal
#> [1] "Leuk. Res."
#> 
#> $content$sources[[1]]$full_journal_title
#> [1] "Leukemia research"
#> 
#> $content$sources[[1]]$status
#> [1] "fully curated"
#> 
#> $content$sources[[1]]$is_review
#> [1] FALSE
#> 
#> $content$sources[[1]]$clinical_trials
#> list()
#> 
#> 
#> $content$sources[[2]]
#> $content$sources[[2]]$id
#> [1] 1747
#> 
#> $content$sources[[2]]$name
#> [1] "Ponatinib in the therapy of chronic myeloid leukemia."
#> 
#> $content$sources[[2]]$citation
#> [1] "Poch Martell et al., 2016, Expert Rev Hematol"
#> 
#> $content$sources[[2]]$citation_id
#> [1] "27590270"
#> 
#> $content$sources[[2]]$source_type
#> [1] "PubMed"
#> 
#> $content$sources[[2]]$asco_abstract_id
#> NULL
#> 
#> $content$sources[[2]]$source_url
#> [1] "http://www.ncbi.nlm.nih.gov/pubmed/27590270"
#> 
#> $content$sources[[2]]$open_access
#> NULL
#> 
#> $content$sources[[2]]$pmc_id
#> NULL
#> 
#> $content$sources[[2]]$publication_date
#> $content$sources[[2]]$publication_date$year
#> [1] 2016
#> 
#> $content$sources[[2]]$publication_date$month
#> [1] 10
#> 
#> 
#> $content$sources[[2]]$journal
#> [1] "Expert Rev Hematol"
#> 
#> $content$sources[[2]]$full_journal_title
#> [1] "Expert review of hematology"
#> 
#> $content$sources[[2]]$status
#> [1] "fully curated"
#> 
#> $content$sources[[2]]$is_review
#> [1] FALSE
#> 
#> $content$sources[[2]]$clinical_trials
#> list()
#> 
#> 
#> 
#> $content$allele_registry_id
#> NULL
#> 
#> $content$allele_registry_hgvs
#> NULL
#> 
#> $content$provisional_values
#> named list()
#> 
#> $content$errors
#> named list()
#> 
#> 
#> $url
#> [1] "https://civicdb.org/api/variants/1"
#> 
#> $response
#> Response [https://civicdb.org/api/variants/1]
#>   Date: 2019-09-03 16:44
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 19.7 kB
#> 
#> 
#> attr(,"class")
#> [1] "civic_api"