Get a specific variant group metadata information

Retrieve metadata information for a specific variant group from the CIViC DB

getVariantGroupMetadata(id, type)

Arguments

id
Internal CIViC ID of the variant group of interest

type
Type of metadata (comments, suggested_changes, revisions)

Value

An S3 Object of type civic_api containing the content, url, and response

Examples

getVariantGroupMetadata(id = 1, type = "comments")

#> $content
#> list()
#> 
#> $url
#> [1] "https://civicdb.org/api/variant_groups/1/comments"
#> 
#> $response
#> Response [https://civicdb.org/api/variant_groups/1/comments]
#>   Date: 2019-09-03 16:44
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 2 B
#> 
#> 
#> attr(,"class")
#> [1] "civic_api"

getVariantGroupMetadata(id = 1, type = "revisions")

#> $content
#> $content[[1]]
#> $content[[1]]$id
#> [1] 230
#> 
#> $content[[1]]$action
#> [1] "create"
#> 
#> $content[[1]]$user
#> $content[[1]]$user$id
#> [1] 41
#> 
#> $content[[1]]$user$name
#> [1] "Nick Spies"
#> 
#> $content[[1]]$user$last_seen_at
#> [1] "2019-01-20T03:51:36.969Z"
#> 
#> $content[[1]]$user$username
#> [1] "NickSpies"
#> 
#> $content[[1]]$user$role
#> [1] "admin"
#> 
#> $content[[1]]$user$avatar_url
#> [1] "https://secure.gravatar.com/avatar/3376aeb8439c5ab3e5d72fa2eeed39e5.png?d=identicon&r=pg&s=32"
#> 
#> $content[[1]]$user$avatars
#> $content[[1]]$user$avatars$x128
#> [1] "https://secure.gravatar.com/avatar/3376aeb8439c5ab3e5d72fa2eeed39e5.png?d=identicon&r=pg&s=128"
#> 
#> $content[[1]]$user$avatars$x64
#> [1] "https://secure.gravatar.com/avatar/3376aeb8439c5ab3e5d72fa2eeed39e5.png?d=identicon&r=pg&s=64"
#> 
#> $content[[1]]$user$avatars$x32
#> [1] "https://secure.gravatar.com/avatar/3376aeb8439c5ab3e5d72fa2eeed39e5.png?d=identicon&r=pg&s=32"
#> 
#> $content[[1]]$user$avatars$x14
#> [1] "https://secure.gravatar.com/avatar/3376aeb8439c5ab3e5d72fa2eeed39e5.png?d=identicon&r=pg&s=14"
#> 
#> 
#> $content[[1]]$user$area_of_expertise
#> [1] "Research Scientist"
#> 
#> $content[[1]]$user$orcid
#> [1] ""
#> 
#> $content[[1]]$user$display_name
#> [1] "NickSpies"
#> 
#> $content[[1]]$user$created_at
#> [1] "2015-06-12T18:13:16.508Z"
#> 
#> $content[[1]]$user$url
#> [1] ""
#> 
#> $content[[1]]$user$twitter_handle
#> [1] "@NickSpies13"
#> 
#> $content[[1]]$user$facebook_profile
#> [1] ""
#> 
#> $content[[1]]$user$linkedin_profile
#> [1] ""
#> 
#> $content[[1]]$user$bio
#> [1] "Nick Spies is a staff analyst at the McDonnell Genome Institute and an MD student at Washington University School of Medicine. He has made substantial contributions to the development of genome analysis tools and resources at the Genome Institute including the Drug-Gene Interaction Database. He is a founding member of the CIViC team, helping to define the CIViC data model, and a leading content curator and a feature development consultant."
#> 
#> $content[[1]]$user$country
#> $content[[1]]$user$country$id
#> [1] 214
#> 
#> $content[[1]]$user$country$iso
#> [1] "US"
#> 
#> $content[[1]]$user$country$name
#> [1] "United States"
#> 
#> 
#> $content[[1]]$user$featured_expert
#> [1] TRUE
#> 
#> $content[[1]]$user$accepted_license
#> NULL
#> 
#> $content[[1]]$user$signup_complete
#> NULL
#> 
#> $content[[1]]$user$affiliation
#> [1] ""
#> 
#> $content[[1]]$user$organization
#> $content[[1]]$user$organization$id
#> [1] 1
#> 
#> $content[[1]]$user$organization$name
#> [1] "The McDonnell Genome Institute"
#> 
#> $content[[1]]$user$organization$url
#> [1] "http://genome.wustl.edu/"
#> 
#> $content[[1]]$user$organization$description
#> [1] "The McDonnell Genome Institute (MGI) is a world leader in the fast-paced, constantly changing field of genomics. A truly unique institution, we are pushing the limits of academic research by creating, testing, and implementing new approaches to the study of biology with the goal of understanding human health and disease, as well as evolution and the biology of other organisms."
#> 
#> $content[[1]]$user$organization$profile_image
#> $content[[1]]$user$organization$profile_image$x256
#> [1] "/system/organizations/profile_images/000/000/001/x256/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content[[1]]$user$organization$profile_image$x128
#> [1] "/system/organizations/profile_images/000/000/001/x128/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content[[1]]$user$organization$profile_image$x64
#> [1] "/system/organizations/profile_images/000/000/001/x64/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content[[1]]$user$organization$profile_image$x32
#> [1] "/system/organizations/profile_images/000/000/001/x32/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content[[1]]$user$organization$profile_image$x14
#> [1] "/system/organizations/profile_images/000/000/001/x14/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> 
#> $content[[1]]$user$organization$parent
#> named list()
#> 
#> 
#> 
#> $content[[1]]$created_at
#> [1] "2015-06-21T16:49:38.943Z"
#> 
#> $content[[1]]$changes
#> $content[[1]]$changes$name
#> [1] "Imatinib Resistance"
#> 
#> $content[[1]]$changes$description
#> NULL
#> 
#> $content[[1]]$changes$deleted
#> [1] FALSE
#> 
#> $content[[1]]$changes$deleted_at
#> NULL
#> 
#> 
#> $content[[1]]$diffs
#> $content[[1]]$diffs$name
#> $content[[1]]$diffs$name$diff
#> [1] "<div class=\"diff\">\n  <ul>\n    <li class=\"ins\"><ins>Imatinib Resistance</ins></li>\n  </ul>\n</div>\n"
#> 
#> $content[[1]]$diffs$name$final
#> [1] "Imatinib Resistance"
#> 
#> 
#> $content[[1]]$diffs$description
#> $content[[1]]$diffs$description$diff
#> [1] "<div class=\"diff\"></div>"
#> 
#> $content[[1]]$diffs$description$final
#> NULL
#> 
#> 
#> $content[[1]]$diffs$deleted
#> $content[[1]]$diffs$deleted$diff
#> [1] "<div class=\"diff\">\n  <ul>\n    <li class=\"ins\"><ins>false</ins></li>\n  </ul>\n</div>\n"
#> 
#> $content[[1]]$diffs$deleted$final
#> [1] FALSE
#> 
#> 
#> $content[[1]]$diffs$deleted_at
#> $content[[1]]$diffs$deleted_at$diff
#> [1] "<div class=\"diff\"></div>"
#> 
#> $content[[1]]$diffs$deleted_at$final
#> NULL
#> 
#> 
#> 
#> 
#> $content[[2]]
#> $content[[2]]$id
#> [1] 674
#> 
#> $content[[2]]$action
#> [1] "update"
#> 
#> $content[[2]]$user
#> $content[[2]]$user$id
#> [1] 41
#> 
#> $content[[2]]$user$name
#> [1] "Nick Spies"
#> 
#> $content[[2]]$user$last_seen_at
#> [1] "2019-01-20T03:51:36.969Z"
#> 
#> $content[[2]]$user$username
#> [1] "NickSpies"
#> 
#> $content[[2]]$user$role
#> [1] "admin"
#> 
#> $content[[2]]$user$avatar_url
#> [1] "https://secure.gravatar.com/avatar/3376aeb8439c5ab3e5d72fa2eeed39e5.png?d=identicon&r=pg&s=32"
#> 
#> $content[[2]]$user$avatars
#> $content[[2]]$user$avatars$x128
#> [1] "https://secure.gravatar.com/avatar/3376aeb8439c5ab3e5d72fa2eeed39e5.png?d=identicon&r=pg&s=128"
#> 
#> $content[[2]]$user$avatars$x64
#> [1] "https://secure.gravatar.com/avatar/3376aeb8439c5ab3e5d72fa2eeed39e5.png?d=identicon&r=pg&s=64"
#> 
#> $content[[2]]$user$avatars$x32
#> [1] "https://secure.gravatar.com/avatar/3376aeb8439c5ab3e5d72fa2eeed39e5.png?d=identicon&r=pg&s=32"
#> 
#> $content[[2]]$user$avatars$x14
#> [1] "https://secure.gravatar.com/avatar/3376aeb8439c5ab3e5d72fa2eeed39e5.png?d=identicon&r=pg&s=14"
#> 
#> 
#> $content[[2]]$user$area_of_expertise
#> [1] "Research Scientist"
#> 
#> $content[[2]]$user$orcid
#> [1] ""
#> 
#> $content[[2]]$user$display_name
#> [1] "NickSpies"
#> 
#> $content[[2]]$user$created_at
#> [1] "2015-06-12T18:13:16.508Z"
#> 
#> $content[[2]]$user$url
#> [1] ""
#> 
#> $content[[2]]$user$twitter_handle
#> [1] "@NickSpies13"
#> 
#> $content[[2]]$user$facebook_profile
#> [1] ""
#> 
#> $content[[2]]$user$linkedin_profile
#> [1] ""
#> 
#> $content[[2]]$user$bio
#> [1] "Nick Spies is a staff analyst at the McDonnell Genome Institute and an MD student at Washington University School of Medicine. He has made substantial contributions to the development of genome analysis tools and resources at the Genome Institute including the Drug-Gene Interaction Database. He is a founding member of the CIViC team, helping to define the CIViC data model, and a leading content curator and a feature development consultant."
#> 
#> $content[[2]]$user$country
#> $content[[2]]$user$country$id
#> [1] 214
#> 
#> $content[[2]]$user$country$iso
#> [1] "US"
#> 
#> $content[[2]]$user$country$name
#> [1] "United States"
#> 
#> 
#> $content[[2]]$user$featured_expert
#> [1] TRUE
#> 
#> $content[[2]]$user$accepted_license
#> NULL
#> 
#> $content[[2]]$user$signup_complete
#> NULL
#> 
#> $content[[2]]$user$affiliation
#> [1] ""
#> 
#> $content[[2]]$user$organization
#> $content[[2]]$user$organization$id
#> [1] 1
#> 
#> $content[[2]]$user$organization$name
#> [1] "The McDonnell Genome Institute"
#> 
#> $content[[2]]$user$organization$url
#> [1] "http://genome.wustl.edu/"
#> 
#> $content[[2]]$user$organization$description
#> [1] "The McDonnell Genome Institute (MGI) is a world leader in the fast-paced, constantly changing field of genomics. A truly unique institution, we are pushing the limits of academic research by creating, testing, and implementing new approaches to the study of biology with the goal of understanding human health and disease, as well as evolution and the biology of other organisms."
#> 
#> $content[[2]]$user$organization$profile_image
#> $content[[2]]$user$organization$profile_image$x256
#> [1] "/system/organizations/profile_images/000/000/001/x256/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content[[2]]$user$organization$profile_image$x128
#> [1] "/system/organizations/profile_images/000/000/001/x128/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content[[2]]$user$organization$profile_image$x64
#> [1] "/system/organizations/profile_images/000/000/001/x64/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content[[2]]$user$organization$profile_image$x32
#> [1] "/system/organizations/profile_images/000/000/001/x32/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> $content[[2]]$user$organization$profile_image$x14
#> [1] "/system/organizations/profile_images/000/000/001/x14/MGI_STANDARD4_logo_brown-example_v1b.png?1494525976"
#> 
#> 
#> $content[[2]]$user$organization$parent
#> named list()
#> 
#> 
#> 
#> $content[[2]]$created_at
#> [1] "2015-06-21T16:49:55.148Z"
#> 
#> $content[[2]]$changes
#> $content[[2]]$changes$description
#> $content[[2]]$changes$description[[1]]
#> NULL
#> 
#> $content[[2]]$changes$description[[2]]
#> [1] "While imatinib has shown to be incredibly successful in treating philadelphia chromosome positive CML, patients that have shown primary or secondary resistance to the drug have been observed to harbor T315I and E255K ABL kinase domain mutations. These mutations, among others, have been observed both in primary refractory disease and acquired resistance. In gastrointestinal stromal tumors (GIST), PDGFRA 842 mutations have also been shown to confer resistance to imatinib. "
#> 
#> 
#> 
#> $content[[2]]$diffs
#> $content[[2]]$diffs$description
#> $content[[2]]$diffs$description$diff
#> [1] "<div class=\"diff\">\n  <ul>\n    <li class=\"ins\"><ins>While imatinib has shown to be incredibly successful in treating philadelphia chromosome positive CML, patients that have shown primary or secondary resistance to the drug have been observed to harbor T315I and E255K ABL kinase domain mutations. These mutations, among others, have been observed both in primary refractory disease and acquired resistance. In gastrointestinal stromal tumors (GIST), PDGFRA 842 mutations have also been shown to confer resistance to imatinib. </ins></li>\n  </ul>\n</div>\n"
#> 
#> $content[[2]]$diffs$description$final
#> [1] "While imatinib has shown to be incredibly successful in treating philadelphia chromosome positive CML, patients that have shown primary or secondary resistance to the drug have been observed to harbor T315I and E255K ABL kinase domain mutations. These mutations, among others, have been observed both in primary refractory disease and acquired resistance. In gastrointestinal stromal tumors (GIST), PDGFRA 842 mutations have also been shown to confer resistance to imatinib. "
#> 
#> 
#> 
#> 
#> 
#> $url
#> [1] "https://civicdb.org/api/variant_groups/1/revisions"
#> 
#> $response
#> Response [https://civicdb.org/api/variant_groups/1/revisions]
#>   Date: 2019-09-03 16:44
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 7.71 kB
#> 
#> 
#> attr(,"class")
#> [1] "civic_api"